Bates/Granger (1969) Forecast Combination Approach

Computes forecast combination weights according to the approach by Bates and Granger (1969) and produces forecasts for the test set, if provided.

comb_BG(x)

Arguments

x

An object of class foreccomb. Contains training set (actual values + matrix of model forecasts) and optionally a test set.

Value

Returns an object of class foreccomb_res with the following components:

Method

Returns the used forecast combination method.

Models

Returns the individual input models that were used for the forecast combinations.

Weights

Returns the combination weights obtained by applying the combination method to the training set.

Fitted

Returns the fitted values of the combination method for the training set.

Accuracy_Train

Returns range of summary measures of the forecast accuracy for the training set.

Forecasts_Test

Returns forecasts produced by the combination method for the test set. Only returned if input included a forecast matrix for the test set.

Accuracy_Test

Returns range of summary measures of the forecast accuracy for the test set. Only returned if input included a forecast matrix and a vector of actual values for the test set.

Input_Data

Returns the data forwarded to the method.

Details

In their seminal paper, Bates and Granger (1969) introduce the idea of combining forecasts. Their approach builds on portfolio diversification theory and uses the diagonal elements of the estimated mean squared prediction error matrix in order to compute combination weights:

$$w_i^{BG} = \frac{\hat{\sigma}^{-2} (i)}{\Sigma_{j=1}^N \hat{\sigma}^{-2} (j)}$$

where \(\hat{\sigma}^{-2} (i)\) is the estimated mean squared prediction error of the i-th model.

The combined forecast is then obtained by:

$$\hat{y}_t = {\mathbf{f}_{t}}'\mathbf{w}^{BG}$$

Their approach ignores correlation between forecast models due to difficulties in precisely estimating the covariance matrix.

References

Bates, J. M., and Granger, C. W. (1969). The Combination of Forecasts. Journal of the Operational Research Society, 20(4), 451--468.

Timmermann, A. (2006). Forecast Combinations. In: Elliott, G., Granger, C. W. J., and Timmermann, A. (Eds.), Handbook of Economic Forecasting, 1, 135--196.

See also

Author

Christoph E. Weiss and Gernot R. Roetzer

Examples

obs <- rnorm(100)
preds <- matrix(rnorm(1000, 1), 100, 10)
train_o<-obs[1:80]
train_p<-preds[1:80,]
test_o<-obs[81:100]
test_p<-preds[81:100,]

data<-foreccomb(train_o, train_p, test_o, test_p)
comb_BG(data)
#> $Method
#> [1] "Bates/Granger (1969)"
#> 
#> $Models
#>  [1] "Series 1"  "Series 2"  "Series 3"  "Series 4"  "Series 5"  "Series 6" 
#>  [7] "Series 7"  "Series 8"  "Series 9"  "Series 10"
#> 
#> $Fitted
#> Time Series:
#> Start = 1 
#> End = 80 
#> Frequency = 1 
#>  [1] 0.6675734 1.5833194 0.6443819 0.9179441 0.6822914 1.3402995 1.1454922
#>  [8] 1.2916645 1.0098423 1.0418799 0.8958282 0.8517471 1.2506701 1.1785563
#> [15] 2.0159862 0.6754769 0.5947357 0.3630619 1.1624175 1.3150091 0.9455511
#> [22] 0.8059200 1.1651766 1.3422940 1.1522935 0.7241763 1.2131771 0.4675045
#> [29] 1.2753373 1.4011339 0.8569864 1.0537832 0.8606707 1.4398236 1.2543896
#> [36] 0.7458595 0.8411663 0.6743347 0.4393090 0.5828574 1.3803559 0.8744679
#> [43] 0.7513465 1.0457438 0.8586695 0.6219768 0.7663450 1.5735356 0.5962535
#> [50] 1.0814976 1.1386730 1.4010448 1.3399097 1.0837587 0.5572391 0.9879321
#> [57] 1.0085031 0.9794061 0.5403466 1.1535570 1.1834569 1.0933564 0.5373302
#> [64] 1.1737329 1.9545219 1.2780164 1.0062971 1.3020793 1.1561934 0.6661839
#> [71] 1.2464826 0.9842618 1.0832465 0.8567118 0.9483771 1.3009404 1.0351426
#> [78] 1.2038885 1.4916500 1.4533319
#> 
#> $Accuracy_Train
#>                 ME     RMSE      MAE      MPE     MAPE        ACF1 Theil's U
#> Test set -1.188559 1.592273 1.378676 126.6177 386.5069 -0.06061653  1.444645
#> 
#> $Input_Data
#> $Input_Data$Actual_Train
#> Time Series:
#> Start = 1 
#> End = 80 
#> Frequency = 1 
#>  [1] -1.70558168 -0.85541313 -0.14490163 -0.32444696 -0.17256490 -1.23606292
#>  [7] -1.90230421 -0.09450402  0.03255579  0.46129012  1.38140030 -0.41647627
#> [13]  0.68094267 -0.41437304 -0.51834551 -0.68401973 -0.88564860  0.04923707
#> [19]  0.18556122 -0.60865779 -0.73110285  2.71514421 -1.33938704 -0.64601525
#> [25] -0.93245461 -0.76908693  0.37157978  0.35543278 -0.98399849  0.21472959
#> [31] -0.08008508 -1.42236955  1.10814896  1.07847764 -0.44025034 -0.77816901
#> [37] -1.81859001 -1.12408090  1.06052384 -1.47870016 -1.55156017  0.77750668
#> [43]  1.06844014 -0.18358770  1.55824293 -0.21324238  0.93053526  0.41081180
#> [49] -1.27984430 -0.78236663 -2.27608345 -0.12639591  1.44814671 -1.44275737
#> [55]  1.46718718 -0.74329990 -0.30422384  0.33765806 -0.60750209 -0.29556027
#> [61] -0.13453714  0.81478437 -0.27292173  2.15948580  1.09173757  0.74338485
#> [67] -1.20785935  0.32781805 -0.53416511  1.28394672  0.02893134 -0.39567707
#> [73] -0.69486833 -1.49208070  1.44425724 -0.34701739 -0.04025917  1.24663211
#> [79] -1.34630152 -0.57390128
#> 
#> $Input_Data$Forecasts_Train
#> Time Series:
#> Start = 1 
#> End = 80 
#> Frequency = 1 
#>       Series 1    Series 2      Series 3    Series 4    Series 5    Series 6
#>  1  0.09015089 -0.15018052 -0.1370307332  0.42147510  0.29151525  1.47852876
#>  2  4.10918268  0.72840328  0.7202802421  0.83089095  2.47509246  2.66985957
#>  3 -0.06955277  1.45742422  0.1058709495 -0.91923252  1.84500418  0.96354445
#>  4  0.29989048  0.98304206  1.1367018463 -0.53426639  2.29399443  0.55906373
#>  5  0.80155553  0.45840564  0.2508345818 -0.11476122  1.29816114  1.73440889
#>  6  1.74362012  1.87581342  1.5181990767  2.59781164  0.59431751  0.42469054
#>  7  0.94984023  1.74083815  0.8076627857  0.36019492  0.86119242  1.07489755
#>  8  2.31314355  0.83518761  1.0288098118  2.56669055  0.77740727  1.05930184
#>  9  1.08601352  0.24903572  1.3585908857 -0.44915571  2.74715026  1.84826470
#> 10  0.37014863 -0.25336629  0.9710049654  0.20849604  0.91829533  3.57539659
#> 11  1.65507033 -0.03623626  2.1470420721  0.49551927  1.12414159  1.09425996
#> 12  1.01939355  0.97056143  1.3735889351  1.40182670  1.39223614  0.48005047
#> 13  1.69626996  0.72505622  1.3233392113  1.97139647  1.37841997  1.77442862
#> 14  3.00886989  1.48545147  0.1701806797  0.42033698  1.16617567  2.22828926
#> 15  0.94302172  1.91698203  2.3944625816  2.60417979  2.15338737  2.88060688
#> 16  1.24041550  0.01465608  0.8084564229  1.22597334  1.01474830  0.87475828
#> 17  0.97810983 -0.49118975  1.2722770209  0.48514288  1.02144084  0.24090523
#> 18  0.16919184  1.81959245 -0.0816590083  0.17621176  0.17364783  0.49069640
#> 19  2.40247432  2.01340453 -1.3293993614  1.33441525  1.04739913  0.84396033
#> 20  0.27613089  2.05360525  0.4503740414  0.90650924  0.70689263  1.13258244
#> 21  2.33058037  0.92857136  0.9274200090  1.30406208  2.20763218  0.02591598
#> 22  0.18866698  1.95244461  2.0322884037  0.52349247  1.03200138 -0.25094966
#> 23  2.79503611  0.61596248  1.2151385258  0.75868828  1.18039755  0.95366396
#> 24  1.68662633  0.58224001  0.5056598754  1.82415583  1.07716058  2.13738777
#> 25  1.08937800  0.52973183  2.5121382812 -0.55564397  1.55839534  2.26444989
#> 26  1.32454774 -0.69018029  0.4205367358  1.09350128  0.83314563  0.45125792
#> 27  1.07131812  0.05450234  2.6747896339  0.63305058  0.77201938  1.68239143
#> 28  1.23304838  0.65698804 -0.0009802585  0.87059120  1.24266606  0.07022664
#> 29  2.98637658  1.35846241  2.2227028352  1.40779552  0.07151703  0.02485131
#> 30  0.04832921  1.04818886  2.0771881661  0.41603105  2.51540668  0.97288292
#> 31  0.44135768  2.13017228  0.3880545405  0.80655011  0.18840118  1.44667189
#> 32  1.25581454  0.47236769  1.5068669683  0.73045333  2.31124741  0.67088775
#> 33  0.80192383  1.47957533  1.4600683731  1.07365823  1.81556680  1.11304908
#> 34  1.42878914  2.41752930  2.4843918615  1.35723614  1.12003267  2.05122939
#> 35  1.07077011  0.91990923  1.8819632260  1.55042842  1.88178985  1.32697740
#> 36 -0.53665380  0.60581300  0.4632977606  1.03840179 -0.79874312  1.78703744
#> 37  1.45398437  1.52525764  2.2855537065 -0.60957529  0.04294892  1.44155962
#> 38  1.87439857  0.20406490  1.5878485326 -0.04971011  0.52594139 -0.01930451
#> 39 -0.29773757  1.24509645 -0.3084820255  3.05203397 -0.86046128  0.84130167
#> 40 -0.07896524  0.22068016  1.3167263249  1.17643641  1.25369007 -0.63770129
#> 41 -0.13484994  1.59492387  2.1941583035  2.12830736  0.78559829  3.83926623
#> 42  1.02261431  2.10889090  1.9130571478  1.43500174 -1.37018145  1.96181916
#> 43  0.97853878  0.05703097  0.2132470073  1.54883349  1.00325808  1.50843200
#> 44  0.72122756  1.70034399  0.5890703003  1.64741830  1.66974557  1.25952319
#> 45  0.97579359  0.57532517  1.4763736818  1.87846345  1.22682084  0.22659587
#> 46  1.82502030 -0.14313760  1.1859158675  1.35079787  0.40813151 -0.80552753
#> 47  0.24981722  1.23344022 -0.3247162530  1.04987972  0.59458551  1.17838323
#> 48  1.53621047  0.81076342  2.1723710551  1.83574945  2.41422466  0.47422269
#> 49 -0.57271748 -0.66341126  1.2314927722  0.71827480  1.24162583  1.40018632
#> 50  0.01357258  2.91569077  1.4830719041  0.20832054 -0.43608748  0.87741241
#> 51  2.98351232  0.18310556  0.4646804198  1.00165373  0.68544726  0.92865497
#> 52 -0.85137844  1.38365049  2.3781358190 -0.18753013  0.51945719  3.09573436
#> 53  0.09021182  0.54136775 -0.3026716711  1.36241747  1.83625425  2.32208261
#> 54 -0.95144002  0.28429602  1.6348575353  0.45056468  0.94765071  2.20055900
#> 55  0.19972354  1.42954854  1.9996551613  1.69294984  1.95195762  0.21812543
#> 56 -0.86936160 -0.16985274  0.6622510472  0.93915608  2.49451280  0.55852355
#> 57  0.24942852  2.32208831  0.9139663930 -0.19354093  0.65578527  0.75397187
#> 58  0.40906161  2.36596464 -0.7188175835  0.88009194  1.64577541  1.36579219
#> 59  0.25790073  0.75027629  0.0708784276  0.29189205 -0.53450055  0.96586712
#> 60  1.69351208  0.57527865  1.8136760465 -0.61666397  1.59967251  1.29622700
#> 61  0.94053603  2.09843934  1.5264135457  1.49584782  0.49564111  0.77894023
#> 62 -0.86389510  1.67092304  2.0119573903  2.29975097  0.59309038  1.18770616
#> 63 -0.27450892  1.03095466  1.8313798093 -0.61598551  0.53906479  0.56193783
#> 64 -0.78177081  3.36851196  1.4151341447 -0.25036794  3.21490843  0.91126642
#> 65  0.49140661  3.58578811  0.3387429391  2.58213231  2.48696782  1.63425003
#> 66 -0.73687829  2.21228095  1.5488112863  2.25252912  0.64490114  1.86553824
#> 67  1.04048130 -0.25119684  0.4606322004  0.79060430  1.87620374  0.56940691
#> 68  0.87582054  1.40026604  0.8286815736  0.36096631  0.78530335  1.01303089
#> 69  0.38724657  1.93470763  0.2332653948  1.39890294  3.02079836  1.05124963
#> 70  1.16066248  0.52733015  0.3934251544  2.34197350 -0.83866156  1.53863815
#> 71  0.33764037  2.58301322  2.6473128210  0.96239258  1.17293716  1.51967261
#> 72  0.66514834  1.59775422  1.8410693481  1.54367011  0.23675496 -1.42291128
#> 73  1.62301159  1.46608459  1.3381750972  2.74021839 -0.11849291  0.19370975
#> 74  2.02819429  0.25563246  0.5951169242  0.67554406  2.19797746  0.51532001
#> 75 -0.13457824  1.95905526  1.9000090391  0.55248867  1.98119391  0.34692272
#> 76  1.91691105  1.15263708  2.1906666552  1.41608253  2.19174921  0.70623575
#> 77  2.21387479  0.23063130  0.9524609165 -0.29586537  1.63283357  0.72301100
#> 78  0.31406754  2.39308074  1.7869254248  1.68263309  1.53282595  1.13646959
#> 79 -0.09456819  1.91866877  2.1666841120  1.49691333  0.17987810  2.06720772
#> 80  1.36487410  2.01608706  0.7281761287  2.43966066  0.84067432  1.00182810
#>       Series 7    Series 8    Series 9   Series 10
#>  1  2.00046921  0.28781792  0.83375382  1.84244375
#>  2  1.52799951 -0.36265667  1.12021853  2.26409418
#>  3  2.16360593  0.49930961  0.33826569  0.28965273
#>  4  1.17313701  2.43847286  0.46937743  0.64617038
#>  5  1.57791725  0.48836231  0.69858130 -0.08349099
#>  6  0.88399129  0.99803230  0.39769839  1.96685776
#>  7  1.90452114  2.78313450  0.68179192  0.51946922
#>  8  0.98341329  2.03138098  1.30819604 -0.04103618
#>  9  0.07770259  0.41929531  1.79865009  1.19962995
#> 10 -0.06011455 -0.59308498  2.75057447  2.71290679
#> 11 -0.78719964  0.51095746  1.26533026  1.55273636
#> 12 -0.06655178  0.24729987  1.31715464  0.13126419
#> 13 -0.12982917  1.66537880  0.85126362  1.09863716
#> 14  1.38726780  1.73566389  1.32221895 -0.83487472
#> 15  2.27026338  1.73079934  0.67579087  2.41119492
#> 16 -0.64611783  0.47127153  0.21664073  1.37983102
#> 17  0.87639593 -0.21231512  0.98191288  0.98432814
#> 18  0.02835935  0.38010097  0.83080126 -0.57458146
#> 19  2.06858736  1.78089750  0.83525700  0.64745722
#> 20  3.31625287  1.09622971  1.71362242  1.62181043
#> 21 -0.68287339 -0.56135026  1.23720904  1.41343327
#> 22  1.13587543  1.69837422 -1.32268642  0.90080473
#> 23  1.49843777  1.03819575  1.07904427  0.80397665
#> 24  2.33346297  1.23466066  0.57119365  1.62116236
#> 25  0.97704543  2.02132847  0.62728561  0.95250291
#> 26  0.80847676  2.47512357  0.95508139 -0.09567871
#> 27  0.55938511  0.95160161  2.40134303  1.59783572
#> 28 -0.94140598  0.40532774  0.69912789  0.18987199
#> 29  2.41570672  0.52599391  0.03424559  1.81911755
#> 30  1.59639966  0.67363872  2.68321848  2.10470063
#> 31  0.34707771 -0.18566168  0.22344942  2.41998840
#> 32  0.84585533  0.02271188  1.24730137  1.50181391
#> 33  0.55635434  0.31732257  0.03285304 -0.29236569
#> 34  0.48176622  0.96747231 -0.03785442  1.81758778
#> 35  1.20610706  2.48148112 -0.05359998  0.29912774
#> 36  2.63945367  1.02825460  0.79449256  0.63910243
#> 37 -0.56473471  0.31195409  2.30826655  0.31767825
#> 38  1.51158992  0.10526889  1.51372190 -0.13663536
#> 39 -0.53966137  0.46682941 -0.38535523  0.43577663
#> 40  1.36279556  1.58385837 -0.00123881 -0.34905562
#> 41  1.77073594 -0.26285999  1.30132568  0.34911989
#> 42 -0.28898622  1.89319710  1.43768205 -1.56959690
#> 43  0.28680563  0.10281649  0.34972485  1.33158034
#> 44  0.89761661  1.55461273 -1.41967894  1.50872867
#> 45  0.35238825 -0.16400329  1.27554607  0.50289441
#> 46  0.62674616  1.01865667  0.78265361  0.05561020
#> 47  1.08334721  1.32882120  1.42081028 -0.19445969
#> 48  1.91721627  0.70505369  1.51449255  2.35163231
#> 49  0.74104670  1.64159990  0.02387975  0.38611621
#> 50  1.67702573  1.92667781  1.82132277  0.33329832
#> 51  0.57989926  2.77212400  2.48745742 -0.29254753
#> 52  2.04720542  2.66209268  1.52952179  1.82272239
#> 53 -0.54657635  1.02792891  2.88518177  4.00809274
#> 54 -0.04600568  2.61549511  1.72341993  2.11924176
#> 55 -1.27266976 -0.34752241  0.88469426 -1.82179908
#> 56  0.68334806  1.19362683  1.78134967  2.61728753
#> 57 -0.47619064  1.42150530  2.80293717  1.51382026
#> 58  0.89961217  0.49986680  0.61331078  1.50278931
#> 59  0.25377739 -0.59442601  3.71554023  0.26470842
#> 60  1.20364588  2.41707609  2.24977062 -0.10851818
#> 61  1.49745036  2.05056411 -0.95674563  1.71705719
#> 62 -0.01428664  0.41556809  1.66675609  1.40585690
#> 63  2.14576035  0.47328513  0.46306769 -0.49918152
#> 64  0.78558123  0.87775928  2.12178686 -0.22357851
#> 65  2.67489934  0.41367172  2.39415872  2.42534798
#> 66  1.11636015  0.75213312  1.11220949  1.53548399
#> 67  2.42656559  1.80125715  1.26312335  0.52087590
#> 68  0.76033267  3.52510877  2.00675166  1.70875558
#> 69 -0.10547392  1.27003304  0.95379909  0.98905929
#> 70  1.32289773  0.11364297  0.34766935 -0.39260904
#> 71  1.02174680  0.23047615  0.02275085  1.62494754
#> 72  1.31762482  1.45405702  0.74006285  1.68071956
#> 73  0.03029916  0.29334275 -0.55323598  3.25114000
#> 74  2.43446772  0.05060454  0.95474719 -0.84212098
#> 75  0.01119515 -0.06640060  1.67735566  0.91106797
#> 76  1.39634960  0.11354881  1.53612501  0.34581417
#> 77  1.52005742  1.66859230  1.61687349  0.62891030
#> 78  1.71647258 -1.06756433  0.33856346  1.73214268
#> 79  0.93888419  2.69414641  2.33858671  1.15818293
#> 80  3.27437658  0.81022964 -0.11781788  2.00547316
#> 
#> $Input_Data$Actual_Test
#>  [1] -0.7059593 -1.8574045 -0.2469705 -0.3355609 -0.2493769  0.4595226
#>  [7] -0.4600755  0.0589328 -0.7425911 -2.1635740 -0.2328965  0.8749068
#> [13] -0.9734940 -0.5127469 -0.9234880 -1.4944257 -0.1659304  1.5822383
#> [19]  1.0146023 -1.1464783
#> 
#> $Input_Data$Forecasts_Test
#>           Series 1     Series 2   Series 3    Series 4   Series 5    Series 6
#>  [1,] -0.008017083  1.534635439  0.6862957  2.20735856  0.9824679  1.14009225
#>  [2,]  1.554996436  0.012371440  0.7136998  1.23851469  0.6093358  0.47940146
#>  [3,]  1.436065921  2.187792273  0.1965597  0.70116212  0.5972249  3.36581164
#>  [4,]  1.105373296  0.482457788  0.9054445  2.38462105  2.2767070 -0.69647441
#>  [5,]  0.746262147  0.740439749  0.9259535  0.29993098  1.8210758  1.68258558
#>  [6,]  0.023508768  0.671935329  1.7144907  2.86093151  1.4197851  2.36156198
#>  [7,]  1.394217766  1.073432395  1.1247184  2.80392496  2.7116612  2.90595576
#>  [8,]  2.014903026  0.752136980  2.2965943 -0.35604349 -1.2156327  0.32716146
#>  [9,]  1.631443810 -0.373862256 -0.1144929  0.68524893  0.8151218 -0.06291143
#> [10,]  0.516601745  0.959554182  0.1579412 -0.01926518  1.6711682  2.63368316
#> [11,] -0.526333936  1.421538239 -0.5042942  0.38775034  0.2048598  1.54083036
#> [12,]  1.616842281  1.201597506  0.7159735  0.71093125 -0.5662578  2.45685797
#> [13,]  2.229286839 -0.697191918  1.0428690  2.50698329  2.2137325 -0.93378424
#> [14,]  1.157453027  1.642287683  0.9911336  1.36772269  2.5038101  0.11959920
#> [15,]  2.406342056  0.004760387 -1.9490838  0.97996506 -0.1694250  0.20226984
#> [16,]  1.396800917  1.963813900  1.0203583  0.02007804  1.6133772  0.11386903
#> [17,]  1.326422964 -0.656037231  0.9015530 -0.40091233  2.4839313  0.70517221
#> [18,] -0.459343308  2.070861091  1.5891091  2.44501491  0.5562263  0.11334831
#> [19,]  0.200704473  0.890973640  0.5742484  0.57651838  1.2685897  1.09908467
#> [20,]  1.470012796  2.899186393  1.6346590  0.26625040 -0.2282168  1.72054690
#>          Series 7    Series 8    Series 9  Series 10
#>  [1,]  0.10236280  1.42216699  1.68473354  0.7026386
#>  [2,]  1.61803781  0.33688427  0.38944366  0.7645648
#>  [3,]  1.07756352  0.33723454  0.65069317  0.7368292
#>  [4,]  1.71350068  1.11329221  2.13396930  0.6795360
#>  [5,]  1.52620982  1.94295108  2.53375658  2.9306592
#>  [6,]  0.52944427 -0.21587333  2.10666929 -1.1877745
#>  [7,]  1.60882855  0.95273726  1.34406175  1.7009740
#>  [8,]  1.77324476  0.80795504 -1.03857975  0.6236985
#>  [9,]  0.18025180  1.63214857  0.50654237  1.3351810
#> [10,]  0.87157726  0.84809505  0.93920704  0.7064743
#> [11,]  0.63497720  1.01520037  1.43822179  1.0111422
#> [12,]  0.92521058  1.88871818 -0.21489123  2.1980173
#> [13,] -0.09698788  1.01033495  1.49864688  0.5322665
#> [14,]  1.96671333  1.34493232 -0.08286996  1.9163829
#> [15,]  1.93992695  0.08682158  1.42216645  0.2939397
#> [16,] -0.60017550  0.07560190  0.77711128  1.7869355
#> [17,]  1.20853659  0.62792281  0.49967531  1.3998839
#> [18,]  2.12126035  0.70220215 -0.10958579  0.3612947
#> [19,]  0.34800460  2.57449189  1.90920901  0.6568786
#> [20,] -0.09013527 -0.98066169  0.76603472 -0.1182784
#> 
#> 
#> $Weights
#>  prediction_matrix.Series 1  prediction_matrix.Series 2 
#>                  0.09326928                  0.11532144 
#>  prediction_matrix.Series 3  prediction_matrix.Series 4 
#>                  0.09832551                  0.11718641 
#>  prediction_matrix.Series 5  prediction_matrix.Series 6 
#>                  0.10238573                  0.09702742 
#>  prediction_matrix.Series 7  prediction_matrix.Series 8 
#>                  0.08785250                  0.09244420 
#>  prediction_matrix.Series 9 prediction_matrix.Series 10 
#>                  0.09572415                  0.10046337 
#> 
#> $Forecasts_Test
#>  [1] 1.0859154 0.7580560 1.1376104 1.2187274 1.4815552 1.0669117 1.7815500
#>  [8] 0.5597530 0.6047896 0.9146068 0.6700637 1.0796205 0.9417087 1.3047423
#> [15] 0.4945791 0.8438357 0.7651525 0.9837715 0.9993377 0.7698193
#> 
#> $Accuracy_Test
#>                 ME     RMSE      MAE      MPE     MAPE
#> Test set -1.384144 1.651062 1.445517 196.7319 297.2711
#> 
#> attr(,"class")
#> [1] "foreccomb_res"