Constrained Least Squares Forecast Combination

Computes forecast combination weights using constrained least squares (CLS) regression.

comb_CLS(x)

Arguments

x

An object of class 'foreccomb'. Contains training set (actual values + matrix of model forecasts) and optionally a test set.

Value

Returns an object of class foreccomb_res with the following components:

Method

Returns the best-fit forecast combination method.

Models

Returns the individual input models that were used for the forecast combinations.

Weights

Returns the combination weights obtained by applying the combination method to the training set.

Fitted

Returns the fitted values of the combination method for the training set.

Accuracy_Train

Returns range of summary measures of the forecast accuracy for the training set.

Forecasts_Test

Returns forecasts produced by the combination method for the test set. Only returned if input included a forecast matrix for the test set.

Accuracy_Test

Returns range of summary measures of the forecast accuracy for the test set. Only returned if input included a forecast matrix and a vector of actual values for the test set.

Input_Data

Returns the data forwarded to the method.

Details

The function integrates the constrained least squares (CLS) forecast combination implementation of the ForecastCombinations package into ForecastComb. The implementation has improved robustness regarding multicollinearity.

Compared to the ordinary least squares forecast combination method, CLS forecast combination has the additional requirement that the weights, \(\mathbf{w}^{CLS} = (w_1, \ldots, w_N)'\), sum up to 1 and that there is no intercept. That is, the combinations of comb_CLS are affine combinations.

This method was first introduced by Granger and Ramanathan (1984). The general appeal of the method is its ease of interpretation (the weights can be interpreted as percentages) and often produces better forecasts than the OLS method when the individual forecasts are highly correlated. A disadvantage is that if one or more individual forecasts are biased, this bias is not corrected through the forecast combination due to the lack of an intercept.

In addition to the version presented by Granger and Ramanathan (1984), this variant of the method adds the restriction that combination weights must be non-negative, which has been found to be almost always outperform unconstrained OLS by Aksu and Gunter (1992) and was combined with the condition of forcing the weights to sum up to one by Nowotarski et al. (2014), who conclude that even though the method provides a suboptimal solution in-sample, it almost always produces better forecasts than unrestricted OLS out-of-sample.

The results are stored in an object of class 'foreccomb_res', for which separate plot and summary functions are provided.

References

Aksu, C., and Gunter, S. I. (1992). An Empirical Analysis of the Accuracy of SA, OLS, ERLS and NRLS Combination Forecasts. International Journal of Forecasting, 8(1), 27--43.

Granger, C., and Ramanathan, R. (1984). Improved Methods Of Combining Forecasts. Journal of Forecasting, 3(2), 197--204.

Nowotarski, J., Raviv, E., Tr\"uck, S., and Weron, R. (2014). An Empirical Comparison of Alternative Schemes for Combining Electricity Spot Price Forecasts. Energy Economics, 46, 395--412.

See also

Examples

obs <- rnorm(100)
preds <- matrix(rnorm(1000, 1), 100, 10)
train_o<-obs[1:80]
train_p<-preds[1:80,]
test_o<-obs[81:100]
test_p<-preds[81:100,]

data<-foreccomb(train_o, train_p, test_o, test_p)
comb_CLS(data)
#> $Method
#> [1] "Constrained Least Squares Regression"
#> 
#> $Models
#>  [1] "Series 1"  "Series 2"  "Series 3"  "Series 4"  "Series 5"  "Series 6" 
#>  [7] "Series 7"  "Series 8"  "Series 9"  "Series 10"
#> 
#> $Fitted
#> Time Series:
#> Start = 1 
#> End = 80 
#> Frequency = 1 
#>  [1]  1.64054849  0.24394745  0.26913638 -0.02951454  1.76138804  1.24411152
#>  [7]  1.61672215  1.28201668  0.70244806  0.21913510  1.58621644  0.75532049
#> [13]  1.28473849  1.51561372  0.25099004  0.95310497  0.74928637  0.99417013
#> [19]  1.17487700  1.49592316  0.89304976  0.36009837  0.29938808  1.34356953
#> [25]  1.15101407  1.17849874  0.10140875  1.09720470  1.00667194  0.76678042
#> [31]  1.25466651  0.85936105  1.23679783  1.38845299  0.57647112  1.46125231
#> [37]  0.60039067  0.86588272  0.93947879  0.58238030  0.83494213  1.39678758
#> [43]  0.85266312  0.66755729  0.69502150  1.24468131  1.87608447  0.61385013
#> [49]  1.03617717  1.07398686  1.98965549  0.85636443  0.86236194  1.19637663
#> [55]  1.19179796  0.56518163  1.68469894  1.49049108  1.06029600  0.86978273
#> [61]  0.47766729  1.06881756  1.29289449  0.96042559  1.06676189  1.21898200
#> [67]  0.61300428  1.21683971  0.88662744  1.13733841  0.47829390  0.80986156
#> [73]  0.54288779  0.26271040  1.06631281  0.56836324  0.48724304  1.26746025
#> [79]  1.07885119  1.61320481
#> 
#> $Accuracy_Train
#>                  ME     RMSE      MAE       MPE     MAPE       ACF1 Theil's U
#> Test set -0.9256253 1.405946 1.176104 -278.1881 716.8338 -0.1262019 0.5776145
#> 
#> $Input_Data
#> $Input_Data$Actual_Train
#> Time Series:
#> Start = 1 
#> End = 80 
#> Frequency = 1 
#>  [1]  0.578376820  0.113186948  0.440206334 -2.002320938  1.041633836
#>  [6] -1.013748921 -0.504802858  0.089953043 -1.292000906 -1.730530080
#> [11]  1.132424775  0.955398518 -0.714008278 -0.130393939  0.950804885
#> [16]  0.470794223  0.334838125 -1.549423258 -0.621150422 -1.615618228
#> [21]  0.853465374  1.888819161 -0.867441159  0.714420496  0.614427022
#> [26]  0.305098743 -0.925999799  1.359534396  0.203698336  1.288543508
#> [31] -0.212971834  0.487549814 -1.067709165 -1.180488666  2.159175760
#> [36] -0.417812648  0.087654958 -0.543709068 -0.119867022  1.800624858
#> [41] -0.794247350 -0.360360559  0.631460646 -1.663279352 -0.488005245
#> [46] -0.020752219  0.933871038 -0.232670878 -1.019630064 -1.680328321
#> [51]  2.281394917 -1.042285303  0.712307192  1.220747981  0.336018683
#> [56] -0.279147937  0.076422131 -0.354509030  1.432658547  0.002671474
#> [61]  0.344012511 -0.830323380  0.482115591 -0.688660369 -1.093359720
#> [66]  0.841060978  1.454136238  0.137240272  0.325544925  1.223728019
#> [71] -1.053314136  1.840332640 -2.231484275 -0.595844719 -0.230472754
#> [76]  1.756315261  0.475705566  0.782644366 -1.146245649  0.950180682
#> 
#> $Input_Data$Forecasts_Train
#> Time Series:
#> Start = 1 
#> End = 80 
#> Frequency = 1 
#>       Series 1     Series 2    Series 3    Series 4    Series 5    Series 6
#>  1  0.88988051  2.143623543  1.38599496 -0.10735295  2.07465712  0.82352105
#>  2 -0.05157582 -0.381713624  0.16115737  0.37211721  1.65505667  1.75631487
#>  3  0.10859866  0.155930448  0.46997241 -0.47950195  0.87686634  0.01999229
#>  4  0.38086196 -0.836583479 -1.29318208  1.15434317  0.07027820  0.61357562
#>  5  1.25519905  2.044170831  2.85526364 -1.29948441  2.51655653  1.38151625
#>  6  0.56068408  1.215210107  0.37528504  0.56127232  0.92320704  2.34576854
#>  7  2.46814921  1.800349614  1.17587258  0.35031939  0.63020806  1.98374868
#>  8  1.83250613  1.714933031 -0.36578885  1.42404095  1.60623267  1.08663730
#>  9  0.08621259  0.682828791 -0.34551847  1.04704973  2.52673800  1.46966993
#> 10  0.72740457 -1.252045124  1.94726527 -0.28504407  0.91848915  0.10709983
#> 11  1.75108655  2.505248197  0.88426033  1.04553236  2.42835885  1.00222962
#> 12  0.92314374  0.580852128  0.57059551 -0.44942970  2.01357184  1.15074768
#> 13  1.14710153  0.791820013  1.07963198 -0.63370440  0.32342217  0.42544545
#> 14  0.67876180  2.025008622  2.00646399  0.48271504  0.76343433  1.23516043
#> 15 -0.70314007 -1.109352248  2.75451875  0.00884857  1.50559912  2.26585841
#> 16  0.23896156  1.581155983  0.88063869  0.27151496 -1.52775293 -0.39029804
#> 17  0.24097461  1.427476044  2.64399863  1.45195163  0.29933133  1.11579572
#> 18  1.29260865  1.309712183 -0.31099776  1.88718133  1.09160758  1.89257730
#> 19  1.33223639  0.651319199  1.36270142  2.26457340  1.85217358  1.03875288
#> 20  0.76806523  1.472684011  2.36966549 -0.29141939  2.12390932  0.96162730
#> 21  0.18419283  1.702742649  0.65148760  1.18251031  1.82786930  1.81145335
#> 22  1.21607480 -0.543797498  0.45452935  0.91019033 -0.36811281  1.03348690
#> 23 -0.13886579 -0.139419424  0.97948032  0.88737072  1.18975968  1.66331887
#> 24  1.14010501  1.370802634  0.66115011  1.16280134  2.96075593  0.77990820
#> 25  0.33786609  1.249291812  1.18559210  1.55586101  0.19398228  0.98426345
#> 26  0.46560823  2.240415284  1.34586266  1.85339468  1.24735027  3.08771350
#> 27 -0.41283088  0.628405955  0.20666763  0.80510064  1.50782823  2.49658665
#> 28  2.21720350  1.908843016  0.25136839  0.95658809  0.50419825  0.69023598
#> 29  1.42553189  1.772073385 -0.84666664  0.47056168  0.09789763  2.58429093
#> 30  0.87907139  1.366647135 -0.55682442  0.40148940  2.01253254 -0.14727838
#> 31  1.65143072  1.234063054  1.00235494  1.14929768  2.63654057  1.93799414
#> 32  1.49020556  0.175362506  2.47020628  2.25582092  0.17447949 -0.17647338
#> 33  1.70519246  1.793613760  0.22251551  0.76563047 -0.30128401  2.40213301
#> 34  1.80565067  1.815375576  1.70461343  2.42234028  0.51595739  0.82282462
#> 35  0.82000942  0.359587441  1.69660640  1.29212754  0.80843806  0.34939574
#> 36  2.07286188  1.154668564  1.93197293  0.26311325 -0.20399135 -0.22936441
#> 37  0.91747654 -0.319062104  3.61617869  0.87000345 -0.55456583  1.25132590
#> 38  2.07259876  0.440555761 -0.47749228  3.20088565  0.85071030 -0.56321535
#> 39  0.99792026 -0.278683713  1.07467723  2.61308129  0.01175105  1.81307999
#> 40  0.14009468  0.838145365  0.35922983  0.73250042 -0.30053093  0.88915132
#> 41  1.74807413 -0.212698055  1.11492547  1.12760317  1.76400192  0.23882794
#> 42  2.35651545  2.371345622  1.74702130  0.62910726  0.23577998  1.82951258
#> 43  1.40000160  1.082627704 -0.11847926  0.54759772  2.05984222 -0.05517196
#> 44  1.63625661 -0.265853773  1.34100482  2.19747825  1.57440044  1.22296796
#> 45 -0.86603591  2.174275795  1.12503119  1.96133970  2.45435482  3.44119504
#> 46  0.41907276  1.695991675  1.31136972  2.09924455  2.92660650  3.00064170
#> 47  2.09389613  1.800070376  1.42644380  1.05353454  1.37769313  0.24519417
#> 48  0.46730890 -0.002196891 -0.26957932  0.92891744  1.45781732 -0.30323896
#> 49  1.68398409  0.322884333  1.86981105  0.73478738  0.68552183  1.24602616
#> 50  0.31573633  1.720963599  0.97642537 -0.08747279  0.80366278  1.54889371
#> 51  2.62594327  1.960168144  1.64606826  1.11299534  1.53220725  1.48641429
#> 52  0.36826555  0.854334333  2.23267300  0.63123839 -0.29693082  0.14275554
#> 53 -0.35377999  2.505255540 -0.28041054 -0.81381842  0.88354012  0.79111037
#> 54  1.38712223  1.678180406  0.82391383  0.27200563 -0.79143581  0.97817165
#> 55  0.03233248  0.734447804  2.14891232  0.67665654  2.03763845  1.37172886
#> 56  0.64121468  0.541867463  1.10640942  1.11817806  0.24826269 -1.30061488
#> 57  0.41360257  2.738431533  2.05259936  1.52272118  2.44332021  0.31360851
#> 58  2.61291348  0.986894051  1.06350867  0.60528796  1.73418909  1.82395471
#> 59  1.50567134  0.205646646  1.76028867  0.26173764 -0.05455799  1.67156095
#> 60  1.02410609  1.255333393  0.41660877  2.31970010  1.03042473  0.86318891
#> 61  1.29535965  0.121222352  0.16682303  0.45228253  0.37940834  0.67514458
#> 62  2.27535036  0.497615986  0.04472074  3.11653074  0.80649798  2.16536442
#> 63  1.20447718  1.918235108  1.46819746  1.23888968  1.39360687  0.09625856
#> 64  1.18674791  0.584443890  2.48936160  0.48403063  0.96677384  1.98561211
#> 65  1.70974984  0.502545873  1.84563780  2.15875218  0.47127743  0.87367647
#> 66  1.37584673  1.122983974  0.90966964  0.36564219  4.07100715 -0.44396717
#> 67  1.40541324  0.005906159  0.23902591  1.86271083 -0.42472377  1.03215408
#> 68  1.33098575  1.485937859  0.57729651 -0.68157303  1.33070126 -1.11771645
#> 69  2.59376051  0.448958242 -0.11699055 -0.63679212 -0.75947902  2.66581693
#> 70  1.55507576  1.166600822 -0.37320544  2.43029016 -0.01638420  1.57567105
#> 71 -0.20947832  1.095564919  0.53537280  2.53937650  2.03686204  0.99776404
#> 72  0.77444205  1.937361737  0.24801763  0.92278914  1.03471228  0.27798501
#> 73  0.65761816  0.221407756  0.59202546  1.72223242 -0.69369136  1.44867236
#> 74 -0.73655950  0.653985037  1.17718257  1.76201239  0.62591787  1.05173624
#> 75  1.16356090  1.145927947  1.99932634 -0.10521674 -0.25836245 -0.25364945
#> 76  1.29829351  0.320621723  1.11626926 -0.46116859  0.39189437  1.55837496
#> 77  1.17587003  0.386301747 -0.83506531  1.29556190 -0.36870786  1.29870240
#> 78 -0.40560101  3.202935147  0.70130432  1.65852651  0.05783655  0.58447117
#> 79  0.78474411  0.671227469  1.31465272  1.64514272  1.73985773 -0.62537447
#> 80  0.49731973  2.751776850 -0.08246616  1.62365346  1.02209434  1.29964747
#>       Series 7    Series 8    Series 9   Series 10
#>  1  1.60521064  2.12061565 -0.38866883  2.03200271
#>  2  1.80068548  1.39989766  1.43721421 -0.58287658
#>  3  1.31836219  0.01547234  1.31586209  1.08496778
#>  4  1.10347211  0.49743782  1.19461053  2.21079462
#>  5  0.34236340  1.98714844  0.54423787  0.97653743
#>  6  1.03387470  3.19148101  1.81253499  1.08469392
#>  7  0.35024344  0.83495779  1.27504259  0.67493449
#>  8  1.91123965  0.31395920  1.00600941  1.50592356
#>  9  0.95272708  1.94139941  3.01018641  1.41527205
#> 10 -0.17851501  0.83595716  1.31380882 -0.17741390
#> 11  3.24266097 -0.30233941  0.15383729  1.03344533
#> 12  2.46748561  0.27651600  0.86535895  0.32012662
#> 13  1.55206173  2.39008874  2.47083220  1.97781403
#> 14  1.35962252  1.68184063 -0.47512574  2.77247739
#> 15  2.83752167  1.46826028  1.20449968  1.27793329
#> 16  1.88327330  1.42092109  0.65823158  0.51898069
#> 17 -0.71320146  0.19966874  2.84215765 -0.79106060
#> 18  1.29372860  0.51154249  0.79409046  1.53066370
#> 19  2.33080754  1.53900452  2.50223128 -0.12476618
#> 20  1.00102550  2.43517102  1.24218806  1.04022615
#> 21  1.94401212  0.73816128  1.05345572 -0.18796740
#> 22  0.83688984 -0.41862372  0.87485369  3.24883111
#> 23  1.02093633  0.48620728  1.24778724  2.55316175
#> 24  1.55105923  1.77230122  0.37338742  1.32302195
#> 25  1.75481473  2.40368234  1.68287753  1.02275970
#> 26 -0.01790465  0.98419509  1.58960606  0.75057213
#> 27  0.67947445 -0.02328956  0.18503268 -1.17231912
#> 28  0.54276043 -1.11639216  0.65420101 -0.33965912
#> 29 -1.16867569  1.15037144  1.05652905  0.01951425
#> 30  1.06640965  0.47912880  0.43323382 -0.79278068
#> 31  2.51311497  0.09454516  0.95716254  1.52281281
#> 32  1.24256364  0.25001310 -0.28155930 -0.59132848
#> 33  1.19609740  0.21342398  1.96758224  1.80626313
#> 34  0.34397622  0.44031813  2.02904885  0.31618693
#> 35  0.52164495 -0.39085131 -1.16642525  1.33438806
#> 36  1.56887874  0.71736940  0.69665178  1.94300642
#> 37 -1.10262681  0.51971500  1.17926804  0.10161233
#> 38  0.99669850  0.03237899  2.42739582  1.76733587
#> 39  1.41691801  2.77794098  0.38005340  1.09233868
#> 40  1.47386760  0.93091096  0.86815755  0.64830397
#> 41  0.93707488  0.57381175  1.26755175  1.11537300
#> 42 -0.63621935 -1.25140373 -0.45677852  0.85673865
#> 43  2.16822965  0.08582905  1.23434052 -1.41982035
#> 44  0.59131381  0.19947542  0.04847195 -0.16379404
#> 45  0.90588294  0.24669410  2.89843940  1.10342665
#> 46  1.46758931  1.66064917 -0.04277055  0.87385667
#> 47  0.17180462  2.72769564 -0.33374629  0.85825659
#> 48  1.39681323  1.78267858  2.76472481  1.91043704
#> 49  1.24447551  0.21818159  1.50606343  2.47761791
#> 50  1.62743759  0.87420245  0.34470778  2.64624260
#> 51  0.19479353  2.01523607  1.66750884  0.82000698
#> 52 -0.93205307  1.29467627  2.37287615  1.97685639
#> 53  1.71708201  0.70354121  1.59458198  0.81339727
#> 54  2.79248656  0.17771018  1.80930744  0.96590161
#> 55  2.12319487  2.69154608  0.07064828  2.48873220
#> 56  2.60994768 -0.47654571  0.03111743  0.81016070
#> 57  2.73279950  1.92969356  0.43651695 -0.08930962
#> 58  2.27570372  0.38616791  2.88848871  1.37108948
#> 59  1.42400799  1.61759809  0.88323039 -0.02777011
#> 60  0.86222187  0.16938482  0.95022156  1.01491585
#> 61  0.40206268 -0.13337593  1.41253090  0.47970144
#> 62  0.27792673  0.84362168  1.33900622  0.94775557
#> 63  1.07276248  0.75690825  0.78053606 -0.43507957
#> 64  2.47336251 -0.12926441  0.72007802  0.67202093
#> 65  0.68379392  0.93780752  0.81553338 -0.28534762
#> 66 -0.41472141  1.48708263 -0.13104119  0.57446276
#> 67 -0.98617904  0.94536305  1.59906814  1.25747566
#> 68  1.70573403  0.80137579  0.65497554  1.20131940
#> 69  0.20957237 -0.45337500  1.32868363  1.48207740
#> 70  1.75848016  1.19843462  0.99799821  1.50663659
#> 71  3.07837562 -0.39100169 -0.64364047  0.47330864
#> 72  2.91913490 -1.24423510  1.83014458  0.43522602
#> 73  1.07552982  0.76820756  2.31130078  1.04039804
#> 74  1.26775153  0.31344864  0.27429102  0.79921862
#> 75  1.43629074  0.51708472  0.61165493  0.09578281
#> 76  2.20974878 -1.24350827 -0.04240835  1.16665552
#> 77 -0.25305556  0.61892143  1.73996554  0.47571522
#> 78  2.76502863  1.11604773  0.62077201  1.01756390
#> 79  0.84067797  1.89249289  1.85504041  1.94851719
#> 80  1.89887422  2.80852040  1.92972681  1.27026578
#> 
#> $Input_Data$Actual_Test
#>  [1] -0.34522799 -2.47355149  1.37626416  2.47714507 -1.33723420  2.02415520
#>  [7]  0.85959318  0.72847134  1.70297114 -1.34967197  0.08596097  0.20420709
#> [13] -0.97905044  0.59465080  1.03931710 -1.27883428 -0.31997882 -1.87282861
#> [19]  0.09132266  0.71236377
#> 
#> $Input_Data$Forecasts_Test
#>         Series 1    Series 2     Series 3   Series 4   Series 5   Series 6
#>  [1,]  2.6838817  0.02834826  1.221793014  3.3158493  1.5488827  0.8539921
#>  [2,] -0.7250366 -0.68954689  0.950148697  0.6184011 -0.3506981  1.3304704
#>  [3,]  0.7474452 -0.48138618  2.107534869  1.0316882  0.3555324  0.6556415
#>  [4,] -0.6048680  3.42689478  0.006639381  1.7766254  2.0478447  1.8917623
#>  [5,]  2.1533763  0.77153998  2.924843474 -0.9830478  0.6786066  0.8637057
#>  [6,] -0.2663631  2.11328352  0.443237596  0.9945533  1.0343259  1.2811912
#>  [7,]  0.7498721  0.56293615  1.329474662  1.7024117  2.0639484  0.8891875
#>  [8,]  0.4610114  2.26488461  2.758286351  2.2231640 -0.5876082  0.9637526
#>  [9,]  0.3888406  2.43659773  2.273909790  0.8586025 -0.5834624  1.0249772
#> [10,]  1.8078373  0.08133568  0.416114805  0.5046210  1.1105313  1.1310536
#> [11,]  2.3579676 -0.91460251  0.672307622  1.8405226  1.2462988  3.5955239
#> [12,]  2.0009284  1.11496953  1.145241262  1.5600501  1.0571499  1.4192840
#> [13,] -0.2587053  2.51034812  1.214101138  2.1370459 -0.0833035  0.7183016
#> [14,] -0.3954291  0.99094739  1.011554669  1.4738013  1.2163011  1.5448631
#> [15,]  1.3267751  0.86414326  2.219699084  0.8495596  0.1763145  1.8682307
#> [16,]  0.4337302  0.66374933 -0.030643641  3.1389952 -0.5007804  0.7933247
#> [17,]  0.9599068  1.35779653  1.764734975  1.2179965  2.2265670  1.0749473
#> [18,]  2.0074259  1.34049858  0.485017396 -0.8959692  0.7841325 -1.2509013
#> [19,] -0.3735209  4.09723070  2.740986972  0.8572616  2.7337848  2.5009612
#> [20,]  1.3308582  1.67735274  2.005845241  2.4624257 -0.4325357  0.5172311
#>          Series 7   Series 8   Series 9   Series 10
#>  [1,]  0.29500467  2.0796635  0.4885797  0.83905860
#>  [2,]  2.12689657  2.4520553  1.5756067  2.96519951
#>  [3,] -0.97432221  3.5569236 -0.1797607  1.39554804
#>  [4,] -0.24477283  2.0727281  1.8739227  1.87371937
#>  [5,]  2.03125039 -0.1789150  1.7526550  1.47232416
#>  [6,]  1.57909705  2.0754171  1.6319529 -0.48522645
#>  [7,] -0.03789001  1.0538551  1.4890517  0.16383770
#>  [8,]  2.56162232  0.7656562  1.4934662  0.35373079
#>  [9,]  1.10195962  1.9507597 -1.0246687  0.95256296
#> [10,]  3.32423830  1.0730219  0.8846534  1.16775999
#> [11,]  3.32585155  1.6635280  1.4029612  0.90026133
#> [12,] -0.29314996  0.1838013 -0.3295532  1.08599646
#> [13,]  0.95498115  1.2955443  1.8500323  0.62742597
#> [14,]  1.19569423  2.3463577  2.1056819  1.18133254
#> [15,]  1.48062339  0.7199846  1.4041419  0.09319359
#> [16,]  3.01734246  0.9369564  0.8768585  1.59115715
#> [17,]  2.19560865  0.8671662  1.4488477  1.79251323
#> [18,]  1.65657552  1.6293618  3.0681537  1.07687238
#> [19,]  2.02629785  0.3587685 -0.2495368  0.93216280
#> [20,]  1.52019266  0.8959814  0.7804688  1.43328805
#> 
#> 
#> $Weights
#>  [1]  3.058392e-01  2.926303e-01  1.078093e-01  0.000000e+00  2.714791e-02
#>  [6] -1.214720e-17  5.078292e-02  1.732178e-01  7.802808e-18  4.257254e-02
#> 
#> $Forecasts_Test
#>  [1] 1.4138383 0.3283729 0.9506493 1.3005035 1.3529531 1.0318445 0.7810314
#>  [8] 1.3629570 1.4956739 1.0561141 1.0552095 1.1535876 1.0837307 0.8285626
#> [15] 1.1066172 0.6932526 1.2796309 1.4920024 1.6591879 1.3957977
#> 
#> $Accuracy_Test
#>                 ME     RMSE      MAE      MPE     MAPE
#> Test set -1.044074 1.640463 1.332116 -67.1334 286.6812
#> 
#> attr(,"class")
#> [1] "foreccomb_res"