Standard Eigenvector Forecast Combination

Computes forecast combination weights according to the standard eigenvector approach by Hsiao and Wan (2014) and produces forecasts for the test set, if provided.

comb_EIG1(x)

Arguments

x

An object of class foreccomb. Contains training set (actual values + matrix of model forecasts) and optionally a test set.

Value

Returns an object of class foreccomb_res with the following components:

Method

Returns the used forecast combination method.

Models

Returns the individual input models that were used for the forecast combinations.

Weights

Returns the combination weights obtained by applying the combination method to the training set.

Fitted

Returns the fitted values of the combination method for the training set.

Accuracy_Train

Returns range of summary measures of the forecast accuracy for the training set.

Forecasts_Test

Returns forecasts produced by the combination method for the test set. Only returned if input included a forecast matrix for the test set.

Accuracy_Test

Returns range of summary measures of the forecast accuracy for the test set. Only returned if input included a forecast matrix and a vector of actual values for the test set.

Input_Data

Returns the data forwarded to the method.

Details

The standard eigenvector approach retrieves combination weights from the sample estimated mean squared prediction error matrix as follows: Suppose \(y_t\) is the variable of interest, there are \(N\) not perfectly collinear predictors, \(\mathbf{f}_t = (f_{1t}, \ldots, f_{Nt})'\), \(\Sigma\) is the (positive definite) mean squared prediction error matrix of \(\mathbf{f}_t\) and \(\mathbf{e}\) is an \(N \times 1\) vector of \((1, \ldots, 1)'\). The \(N\) positive eigenvalues are then arranged in increasing order \((\Phi_1 = \Phi_{min}, \Phi_2, \ldots, \Phi_N)\), and \(\mathbf{w^j}\) is defined as the eigenvector corresponding to \(\Phi_j\). The combination weights \(\mathbf{w}^{EIG1} = (w_1, \ldots, w_N)'\) are then chosen corresponding to the minimum of \(\left(\frac{\Phi_1}{d_1^2}, \frac{\Phi_2}{d_2^2},\ldots,\frac{\Phi_N}{d_N^2}\right)\), denoted as \(\mathbf{w}^l\), where \(d_j = \mathbf{e}'\mathbf{w}^j\), as: $$\mathbf{w}^{EIG1} = \frac{1}{d_l} \mathbf{w}^l$$

The combined forecast is then obtained by:

$$\hat{y}_t = {\mathbf{f}_{t}}'\mathbf{w}^{EIG1}$$

The difference to extant methods that minimize the population mean squared prediction error (e.g., Newbold and Granger, 1974) is the normalization function. While previous approaches optimize MSPE under the constraint of \(\mathbf{e}'\mathbf{w} = 1\), Hsiao and Wan (2014) show that this is dominated by using \(\mathbf{w}'\mathbf{w} = 1\) as constraint in the optimization problem.

References

Hsiao, C., and Wan, S. K. (2014). Is There An Optimal Forecast Combination? Journal of Econometrics, 178(2), 294--309.

Newbold, P., and Granger, C. W. J. (1974). Experience with Forecasting Univariate Time Series and the Combination of Forecasts. Journal of the Royal Statistical Society, Series A, 137(2), 131--165.

See also

Author

Christoph E. Weiss and Gernot R. Roetzer

Examples

obs <- rnorm(100)
preds <- matrix(rnorm(1000, 1), 100, 10)
train_o<-obs[1:80]
train_p<-preds[1:80,]
test_o<-obs[81:100]
test_p<-preds[81:100,]

data<-foreccomb(train_o, train_p, test_o, test_p)
comb_EIG1(data)
#> $Method
#> [1] "Standard Eigenvector Approach"
#> 
#> $Models
#>  [1] "Series 1"  "Series 2"  "Series 3"  "Series 4"  "Series 5"  "Series 6" 
#>  [7] "Series 7"  "Series 8"  "Series 9"  "Series 10"
#> 
#> $Fitted
#> Time Series:
#> Start = 1 
#> End = 80 
#> Frequency = 1 
#>  [1] 1.17990749 0.56782250 0.63451748 1.01904703 0.66231288 0.64547268
#>  [7] 0.25432241 1.40553495 1.10098134 0.95930636 1.25673125 1.14324898
#> [13] 1.66160609 0.85339039 1.79923019 0.43408067 0.71203962 1.15653944
#> [19] 1.05201960 0.87527359 0.82780413 1.04063103 1.08021442 0.60619766
#> [25] 0.65410412 1.10062804 0.52953810 1.03627558 1.14503726 0.75368764
#> [31] 0.53601325 0.76387640 0.61551643 1.33850066 1.10578403 1.22446986
#> [37] 0.91626682 0.97206099 1.19129833 1.39783789 1.65410177 0.83221690
#> [43] 1.77028405 1.12358417 0.50018413 1.06149876 1.06470962 0.56929274
#> [49] 1.63159202 0.65898866 0.84784947 0.88145509 1.22775551 1.19648085
#> [55] 1.32521537 0.96695921 1.13189042 0.64211788 1.10401573 0.76726222
#> [61] 1.14846817 1.22381090 1.17620702 1.85032974 1.37247034 1.06658105
#> [67] 0.89117621 0.49443089 1.22602705 1.59209931 0.88448192 0.05469283
#> [73] 1.28169321 0.92578848 0.87413541 0.91707761 1.04004060 0.80050220
#> [79] 1.05486906 0.71641626
#> 
#> $Accuracy_Train
#>                  ME    RMSE      MAE       MPE     MAPE         ACF1 Theil's U
#> Test set -0.9530954 1.47675 1.221519 -12.65019 351.2687 -0.004732714 0.7240841
#> 
#> $Input_Data
#> $Input_Data$Actual_Train
#> Time Series:
#> Start = 1 
#> End = 80 
#> Frequency = 1 
#>  [1]  1.20978354  0.14037244 -0.87582787  0.96236474  1.19026769  0.19197782
#>  [7] -1.82760053 -0.84007682  1.62339399  0.16004145  1.60015127  0.11427471
#> [13] -1.36542455  0.43853873 -0.96150727  1.21634889 -0.78071197 -0.68229321
#> [19] -2.24342637  0.18027639  0.78150664 -0.95723134  0.09140705 -0.41574198
#> [25] -0.66142355  0.68397164  0.89238434 -0.72690757  1.01574403 -0.39559343
#> [31]  2.15951107  2.21558158 -0.25571707  0.30229170 -0.66580505 -0.69606121
#> [37]  0.68334193 -0.58927487 -0.07409355 -1.11227103 -0.59106225  0.85694461
#> [43]  0.50889339  0.39710390  0.38675725  0.37013618  1.03672310  2.35876506
#> [49] -0.91360208  1.27559942 -1.43909009  0.03814221 -0.50357219 -1.71317745
#> [55]  0.75245163 -0.88340211  0.13651878  0.42628188 -0.92513975 -0.96909574
#> [61] -0.22658278 -0.97822396 -0.53335413  0.86639293 -0.49338094 -1.22861584
#> [67]  1.28456315  1.21178017  0.01786708  1.70024778  1.34411535  0.24724950
#> [73] -0.09585099 -0.10589477 -1.26115471 -1.76021040  1.79143081 -0.80103458
#> [79] -0.55793318  0.75611749
#> 
#> $Input_Data$Forecasts_Train
#> Time Series:
#> Start = 1 
#> End = 80 
#> Frequency = 1 
#>       Series 1    Series 2    Series 3    Series 4     Series 5    Series 6
#>  1  1.68571475  0.83543625 -0.55205289 -0.15193116  2.997741562  2.90029082
#>  2 -0.57419018  0.49480268  1.96151600  0.12538566  0.707229407  0.19312233
#>  3 -0.03920264  0.59758010 -0.08631527  0.94399454  2.030537614 -0.39633665
#>  4  2.06289269  3.29525136 -0.71625725  0.35187922  0.222543917  0.02515786
#>  5 -0.27938089  0.94195983  0.64477382  1.26751883  1.536711136  1.43501099
#>  6 -0.76525996  1.12237766  1.97927600  1.34975619 -0.035791458  0.24952342
#>  7 -1.39468494  0.79922232 -0.17407070 -0.21730203 -0.184131555  1.18028656
#>  8  1.66185142  1.46080357  0.60353914  1.08192791  2.496071761  1.60573214
#>  9  1.27920214  1.64591565  1.02895934  0.20435658  1.694575009  0.22989794
#> 10 -0.97579813  1.28653862 -0.47050643  0.41526206  0.123351803  3.64719149
#> 11  1.49650693 -0.95530250  1.96517223  1.53480656 -0.339466676  1.71171769
#> 12  2.35303163  0.92987748  0.21855831  0.46930419  2.647912882  0.83927436
#> 13  1.65534828 -0.99347696  2.25731934  1.13069949  2.371481222  2.61826196
#> 14  1.20874837  1.12113268 -1.04624731  1.87706131 -0.069745688  1.45653664
#> 15  1.47930452  1.43559586  2.81907819  1.55789393  1.716576522  2.70934812
#> 16  2.11972189  1.13890432 -0.36738211  2.15001554  0.004710155 -1.39021458
#> 17  1.35897006  1.98249750  1.37197567  0.71769532  0.290872827  0.09948432
#> 18  2.19858349  0.71360860  1.75249610 -0.74263112  0.860499807  0.45425011
#> 19  1.56301946  2.69415206  0.15589730  1.40134331  0.164366441  1.45684325
#> 20  1.36036431  1.52131080 -0.47470376  1.64905980  1.471752612  1.71317272
#> 21  1.82536399  0.52839978  0.87271349  2.08786520 -1.694768511  0.42972462
#> 22  0.95960742  0.61324156  0.88691666  2.15911751  1.034812537  1.65360900
#> 23  0.39454326  0.96707851  2.80629661 -0.51050449  1.362560413  0.48820064
#> 24  0.52135116 -0.58992520  0.08368637  1.11420660  0.667673503  0.86745882
#> 25 -0.62703687 -0.63163038  1.32603605  0.83812181  1.343610367  1.22946312
#> 26  2.09597335  0.38078307  2.72940636  1.55216266  1.273656710  0.57299749
#> 27 -0.13061577  0.57143614  1.61863103  1.34873345  1.293555193 -0.14875313
#> 28  1.74940760  3.14362344  0.75648962  0.93732078 -0.001603188  0.37393249
#> 29  1.50704090  2.38105533  1.30415358  0.51717243 -0.118568138  1.80048647
#> 30  1.14922777  0.01817687  0.84388947  2.99597980 -0.535709243  0.02633468
#> 31  1.75691489  2.31291728  1.66641160 -0.25150957 -0.303462906  0.28983995
#> 32  0.57277547  1.37237404 -1.14302489  1.52848796  0.562001551  1.01809067
#> 33 -0.34783112 -0.56624353  0.21431595 -0.24761627  2.940102483 -0.44506438
#> 34  1.08259559  0.17273766  1.85631614  0.95834866  1.339760884  1.93872973
#> 35  2.15372812  2.04343224  0.64527868 -0.05473729  0.796342273  1.03078885
#> 36  3.58684190  0.43818129  1.18610895  0.90182632 -0.224985183  1.86838197
#> 37 -0.64136882  1.45899022  1.38589772  1.26143179  2.746340578  0.41314355
#> 38  1.18557675  1.46117549  1.60550490  1.91011249  1.482781860  0.77765782
#> 39  1.80813951  1.58820968  1.52212061  2.70788034  1.223189693  0.86079440
#> 40  1.46274190  2.36518764  2.60783185  0.88314993  0.052756175  1.67666016
#> 41  0.85913705  1.23758497  2.52866477  2.67862224  1.039435646  2.09656773
#> 42 -0.83016730 -0.30193238  1.28576087  2.01476604  1.591092492  0.99459824
#> 43  1.40408788  2.73797718  1.96647886  1.14258127  1.902114720  2.23094336
#> 44  0.98624933  2.46415154  1.04167686  1.02544607  1.358692081  0.87904883
#> 45  0.89775876 -0.34449753  0.81343195  0.32154826  1.617214775  0.45947163
#> 46  1.20245444  0.99214646 -0.35109974  0.62793576  0.999698490  1.96318107
#> 47  1.80332055  0.29063926  0.04677411  2.21136061  2.116019831  0.25469089
#> 48  0.41220675  2.41991032 -0.28260715  0.74892342  0.867543676 -1.43703682
#> 49  3.11095854  3.10075387  2.34873878  0.45854274  1.888134388  2.26925259
#> 50  1.10631918  1.05642966  1.08748952  0.90666240  0.944953700  0.69040649
#> 51  1.28863828  1.80554512  0.43324178  0.45861601 -0.144717157  0.63591756
#> 52  0.17187098  0.80727002  1.57397293  1.07153290  1.031086651  0.02230817
#> 53  1.25924619  0.87942187  1.91685224 -0.82262195 -0.522310495  1.65730567
#> 54  0.50476926  0.75881469  2.18252112  2.09916976  0.469685570  0.20810579
#> 55  0.71049833  0.74427044  1.08992096  1.75587684  1.010262759  2.79512205
#> 56  1.64068395  0.79122132 -0.42392593  0.61512039 -0.585926944  3.13085949
#> 57  0.85853714  2.63874899  1.18584016  2.01679022  2.150015160  0.37114260
#> 58  0.12015261  1.71425048  1.64734362  0.47637746  0.854588993  1.54783581
#> 59  0.47636039  1.02173027  1.12338155  1.02745382  2.617246199 -0.08167315
#> 60  0.38656340  0.28913584  0.44295938  1.57984851  1.135390016  1.05988434
#> 61  1.74949240  1.06765025  0.57559621  2.33528304  1.440550012  1.50413197
#> 62  2.36816180  2.86731298  2.42145636  1.35862513  1.638182647  0.21268581
#> 63  1.80904192  1.73932627  0.18910251  2.55262013 -0.087071921 -0.58860298
#> 64  2.05204392  1.96568744  0.32146539  2.56769961  1.324650337  2.12243660
#> 65  2.57671017  1.46933493  2.87299949  1.79857974 -1.020733876  1.84003525
#> 66  2.15202991  1.73292806  0.08425034  3.60505344 -0.580309329  1.07464561
#> 67  0.48225714  2.33379773  0.98787200  1.71011685  0.799550568 -1.80083420
#> 68  0.36498718  2.33018264 -0.11723798  1.15292829 -0.131138060 -0.70188801
#> 69  0.59987609  2.94267087  1.26996560  0.89475193  0.213194693  1.90621436
#> 70  1.59324146  1.45475641  2.51165399  1.44967922  2.762824348  1.38585822
#> 71  0.99890509  0.56412173  0.40684432  1.49960522  0.683850642 -0.23604639
#> 72 -0.89022776  2.64225074 -0.61495437  0.81399185  1.669813830 -0.79380612
#> 73  2.46147713  0.88060870 -0.37389655  2.24476382  2.855952275  1.15840270
#> 74  0.60626575  1.00340336  2.19273754 -1.63185225  0.563349346  0.45110716
#> 75  1.58857701  0.71233752  1.11645395 -1.77102642  2.683487931  0.91977819
#> 76 -0.05946131  2.00491484  1.19714199 -0.01858378  0.771708787  0.37768906
#> 77  1.70208360 -0.51968149 -0.53435721  1.67406362  1.834798318  0.27897061
#> 78  0.74560991  2.08058887  1.02661874  2.12914819  0.372624143  1.43215334
#> 79  0.61954892  1.43537143  1.36292709  0.33790901  0.877100313  1.99559764
#> 80  1.72205180 -0.38914741  0.32267567  0.69919048  1.635431456  0.47401848
#>       Series 7    Series 8    Series 9   Series 10
#>  1  0.76158568  1.04247089  2.21887431  0.08551056
#>  2  1.03399153  1.07196321  0.91317587 -0.28102497
#>  3  0.73827360  1.28950002  1.53134235  0.10649986
#>  4  0.75126728  1.33165697  2.16134845  0.70698965
#>  5 -1.40005737  1.73398444  0.02209036  0.94853761
#>  6  0.68525116  0.74476662  0.06506040  1.02690030
#>  7  0.65103579 -0.07299659  1.52034247  0.41058842
#>  8  0.20453165  1.04890013  2.19957560  1.90559239
#>  9  1.64993356  0.74700311  1.66660465  0.89136614
#> 10  1.23672828 -0.49995253  2.14222658  2.58809294
#> 11  2.80949448  1.86163666  1.71514190  0.61051985
#> 12  1.05108052 -0.56587808  1.45018485  2.15945057
#> 13  1.67434295  1.78739241  1.56279014  2.71977526
#> 14 -0.39363130  2.76884602  1.13119979  0.68194352
#> 15  2.16575457  1.27318478  0.66526594  1.92741742
#> 16 -0.59192653  2.29463774 -1.18227275  0.39674968
#> 17  1.03933481  1.24247379 -1.27280875  0.02220706
#> 18  1.28734529  0.88271145  2.60388472  1.54643546
#> 19  0.33875170  1.74198475  0.13820911  0.70612345
#> 20  0.07015918  2.07047680 -0.15027596 -0.49463565
#> 21 -0.04121988  1.30826577  1.77623802  1.16332195
#> 22  1.95092647  1.71311907 -1.03826438  0.33426532
#> 23  1.92934463  2.01946591  0.40424801  0.88347108
#> 24  2.10942854 -1.31672304  0.70471240  1.90263848
#> 25  1.48741325  1.89832886 -0.01009139 -0.25290071
#> 26  1.32407930  0.71005819 -0.06534187  0.21081831
#> 27  0.66428216  0.24925934 -1.94842238  1.85473440
#> 28  1.39292133  1.74095320 -1.80445483  1.86776478
#> 29  1.73474981  1.09338261  0.88672295 -0.11078728
#> 30  1.93244653 -0.82495709  0.75640755  1.06853924
#> 31  0.06850568 -0.07155595  0.79543391 -1.41642139
#> 32 -0.78207210  1.82933495  1.46287384  1.56952916
#> 33  0.31494973  1.11974261  1.80139488  2.04368729
#> 34  1.94355278  1.27518326  2.13540035  0.60339118
#> 35  0.19222023  0.69740429  2.18987474  1.30628159
#> 36  0.55106636  0.37703023  1.57304512  1.73316346
#> 37  2.42603777  0.68943368 -1.03688917  0.47244338
#> 38  0.54873270  0.25743872  0.40358446 -0.06268802
#> 39  0.82501549 -1.41807206  1.41871879  1.19559182
#> 40  2.39544945  0.84502644  0.74713035  0.42582802
#> 41  1.75358788  1.70596465  1.23842012  1.29293560
#> 42  1.55935077  2.66847176  0.07054422 -0.55875291
#> 43  3.08003990  0.25950181  1.43576247  1.18512205
#> 44  1.29242072  1.03564444  0.83031766  0.18304512
#> 45 -1.05766984  0.97449068  0.95365038  0.61256745
#> 46  1.71702546  2.40479144  0.58191143  0.46683317
#> 47  0.36838845  2.58357257  1.40240920 -0.07037896
#> 48  2.26511727 -0.54087935  0.37585914  0.90052259
#> 49  1.14839717 -0.41248999  0.83057415  1.03706845
#> 50 -0.68327799 -2.15382827  0.16467887  3.47387447
#> 51  2.18838811  0.15156558  1.29261126  0.07470964
#> 52 -0.45381379  2.87674257  1.33795787  0.70633177
#> 53  3.24682245  2.47413049  0.92570466  0.94406971
#> 54  1.69302158  0.68282601  2.76650371  0.62383667
#> 55  2.15871966  1.51740523  1.31658628 -0.06485919
#> 56  0.85001713 -1.18380821  2.16633077  2.39569734
#> 57  1.77541957  0.32209576  0.43180165 -0.60329913
#> 58 -1.00402100 -0.19397452 -0.58330604  1.69472859
#> 59 -0.10279800  1.57041723  1.81844582  2.06773904
#> 60  1.46684941 -0.37593125  0.51150533  1.17515767
#> 61  1.29543940  1.08542654  0.70053774 -0.39947336
#> 62 -1.01940758  1.00483283  0.07414196  1.23312997
#> 63  2.07618713  1.87478684  2.03081813  0.25369128
#> 64  1.51921557  3.52581503  1.77534177  1.48378679
#> 65  1.46653036  1.67222646  0.30693677  0.18747365
#> 66 -0.55210539  0.62224673  1.11378643  1.38707250
#> 67  0.93271083  1.82042576  0.55469443  1.40521631
#> 68 -0.24511450 -0.23313393  2.34018328  0.25504122
#> 69  1.67041555  1.85331650 -0.23803473  0.82995406
#> 70  2.47966304 -0.40232461  0.43833777  2.09048065
#> 71  0.64919170  0.49745982  1.79744745  2.29207019
#> 72  0.88986023 -1.38325758 -0.67776729 -1.22489886
#> 73 -0.01755973  2.02610504  1.16655344  0.73630809
#> 74  0.92931149  2.21331624  1.67386007  1.28976658
#> 75  0.99283028  0.47671378  1.11052578  0.91172969
#> 76  1.30334959  0.78387001  1.44451338  1.38314426
#> 77  1.89937176  1.04616714  0.75759111  2.68574621
#> 78  0.18475393 -0.79567879  0.85750184 -0.37907090
#> 79  2.89466441  0.91227697  0.39429898 -0.71779093
#> 80  1.41467151  0.83650464  0.29895302  0.20515431
#> 
#> $Input_Data$Actual_Test
#>  [1] -0.06565428  1.40382510  0.80612311 -0.12533444 -0.56442254  0.94077799
#>  [7]  1.26980895  1.26503849  0.69953040 -0.47893229 -1.31420274 -0.90364582
#> [13] -0.68267777  0.37459966  0.80978442 -1.02731189  0.27120472  0.37216339
#> [19] -1.47380351 -1.42279915
#> 
#> $Input_Data$Forecasts_Test
#>          Series 1    Series 2   Series 3   Series 4    Series 5     Series 6
#>  [1,]  2.32784279  0.35569507  0.6132068  1.3112786  2.09852053  0.520616953
#>  [2,]  0.87450750  0.68221811  0.7892259  0.5541351  0.75854399  0.085398626
#>  [3,]  2.13979470 -1.61114258  0.3663475  0.4606983 -1.05220527  1.023512013
#>  [4,]  0.21559336  0.59592982  1.0795693  0.4318598  0.16142510  1.911207183
#>  [5,]  0.52941036  2.96393939  1.2386371  1.9870993  0.13871477  1.116916346
#>  [6,]  3.61884348  0.09196176  0.4884536 -1.0407828  1.15870612  0.595551706
#>  [7,]  2.78197043  1.23100282  0.1793576  0.8936354  1.14151503  1.728545506
#>  [8,]  1.31122790  2.08303139 -0.4975464  2.7262202  1.54919748  0.530879581
#>  [9,]  1.60465292 -0.02306279  0.8476980  1.4177740 -0.25469101 -0.694862105
#> [10,]  1.80292536  2.97588492  2.3186660  0.7038226  1.65287800  1.285189240
#> [11,]  0.25781792  1.50174849  0.2240381  2.0235029  0.82919862  0.404561801
#> [12,] -0.08648627  1.93378365  2.1481348  0.8302830 -0.71282287  0.004881956
#> [13,]  2.58588835  0.56960515  2.1409608  0.1261706  0.90848103  0.827946161
#> [14,]  0.85929583  2.31892143  0.6650988  1.2393699  0.06039643  1.535371046
#> [15,]  0.02472337  1.78666256  1.6296591  0.9754117  1.83190259  0.539847456
#> [16,]  0.90815328  1.39090820  1.7340168  1.8585745  1.35367663  0.176096597
#> [17,]  1.69744715  1.55501663  2.8020496  2.3034761  1.81208047  1.104849405
#> [18,]  1.85019426  1.49586149 -0.1484801  0.6113281  0.94946399  1.001218301
#> [19,]  0.71753004 -0.58883685  2.2163814  0.9745766  2.87464215  0.548400251
#> [20,]  2.05682051  0.67873552  0.3265345  1.2447490  2.51944240  1.800731333
#>          Series 7      Series 8     Series 9  Series 10
#>  [1,] -0.76765703  1.9711049692  0.051063954  0.5594439
#>  [2,]  1.58868313  2.0106731480  2.005357428  1.3571067
#>  [3,]  1.89988660  1.6854386920  0.097176932  3.1535520
#>  [4,]  0.29998330  0.4393845861  1.015924940  0.7969465
#>  [5,]  0.20905110  1.5730131381  0.846375161  1.6421096
#>  [6,]  1.71742108 -0.8805651620 -0.168192848  1.5945465
#>  [7,]  0.35875605  0.8582544122  1.956962827  1.6591017
#>  [8,]  0.65196734  0.6899992743  0.622905516  0.9348575
#>  [9,]  0.91081076  1.8183243297  1.251409581  0.8665688
#> [10,]  1.43445420  2.7325625652  0.008066258  0.4737906
#> [11,]  1.18780136  2.6101804846  1.127217622  1.4496606
#> [12,]  0.09023753  0.5481903328  1.488781522  2.1142512
#> [13,]  0.12003103  2.6016983516  0.500329442  0.5220895
#> [14,] -0.45144626  0.0963342305  2.372691566  1.3018534
#> [15,]  1.45395884  0.4780933246  0.432644172  1.8554982
#> [16,] -0.03868829  2.8260265857 -0.325028224 -0.2325536
#> [17,]  1.59345586  0.0008895807  0.780941012  0.7283828
#> [18,]  1.67374625  0.8507603357  1.107024938  2.7830112
#> [19,]  1.59384119  0.5320654166 -0.025905527  1.5381370
#> [20,]  2.01579013  0.0820144856 -0.138437873  1.7781755
#> 
#> 
#> $Weights
#>  [1] 0.10553673 0.10619171 0.10707328 0.09690586 0.09139680 0.11000344
#>  [7] 0.10465603 0.09192013 0.09498281 0.09133323
#> 
#> $Forecasts_Test
#>  [1] 0.8820314 1.0471752 0.8060435 0.7095381 1.2281875 0.7430001 1.2784478
#>  [8] 1.0466924 0.7558134 1.5605446 1.1307567 0.8364224 1.0958863 1.0085460
#> [15] 1.0948435 0.9560208 1.4622215 1.2070693 1.0217476 1.2408490
#> 
#> $Accuracy_Test
#>                 ME     RMSE      MAE      MPE     MAPE
#> Test set -1.047888 1.435433 1.125174 161.4387 249.1044
#> 
#> attr(,"class")
#> [1] "foreccomb_res"