Bias-Corrected Eigenvector Forecast Combination

Computes forecast combination weights according to the bias-corrected eigenvector approach by Hsiao and Wan (2014) and produces forecasts for the test set, if provided.

comb_EIG2(x)

Arguments

x

An object of class foreccomb. Contains training set (actual values + matrix of model forecasts) and optionally a test set.

Value

Returns an object of class foreccomb_res with the following components:

Method

Returns the used forecast combination method.

Models

Returns the individual input models that were used for the forecast combinations.

Intercept

Returns the intercept (bias correction).

Weights

Returns the combination weights obtained by applying the combination method to the training set.

Fitted

Returns the fitted values of the combination method for the training set.

Accuracy_Train

Returns range of summary measures of the forecast accuracy for the training set.

Forecasts_Test

Returns forecasts produced by the combination method for the test set. Only returned if input included a forecast matrix for the test set.

Accuracy_Test

Returns range of summary measures of the forecast accuracy for the test set. Only returned if input included a forecast matrix and a vector of actual values for the test set.

Input_Data

Returns the data forwarded to the method.

Details

The bias-corrected eigenvector approach builds on the idea that if one or more of the predictive models yield biased predictions, the accuracy of the standard eigenvector approach can be improved by eliminating the bias. The optimization procedure to obtain combination weights coincides with the standard eigenvector approach, except that it is applied to the centered MSPE matrix after extracting the bias (by subtracting the column means of the MSPE).

The combination weights are calculated as:

$$\mathbf{w}^{EIG2} = \frac{1}{\tilde{d}_l} \tilde{\mathbf{w}}^l$$

where \(\tilde{d}_j\) and \(\tilde{\mathbf{w}}^j\) are defined analogously to \(d_j\) and \(\mathbf{w}^j\) in the standard eigenvector approach, with the only difference that they correspond to the spectral decomposition of the centered MSPE matrix rather than the uncentered one.

The combined forecast is then obtained by:

$$\hat{y}_t = a + {\mathbf{f}_t}'\mathbf{w}^{EIG2}$$

where \(a = E(y_t) - E(\mathbf{f}_t)' \mathbf{w}^{EIG2}\) is the intercept for bias correction. If the actual series and the forecasts are stationary, the expectations can be approximated by the time series means, i.e. the intercept is obtained by subtracting the weighted sum of column means of the MSPE matrix from the mean of the actual series. Forecast combination methods including intercepts therefore usually require stationarity.

References

Hsiao, C., and Wan, S. K. (2014). Is There An Optimal Forecast Combination? Journal of Econometrics, 178(2), 294--309.

See also

Author

Christoph E. Weiss and Gernot R. Roetzer

Examples

obs <- rnorm(100)
preds <- matrix(rnorm(1000, 1), 100, 10)
train_o<-obs[1:80]
train_p<-preds[1:80,]
test_o<-obs[81:100]
test_p<-preds[81:100,]

data<-foreccomb(train_o, train_p, test_o, test_p)
comb_EIG2(data)
#> $Method
#> [1] "Bias-Corrected Eigenvector Approach"
#> 
#> $Models
#>  [1] "Series 1"  "Series 2"  "Series 3"  "Series 4"  "Series 5"  "Series 6" 
#>  [7] "Series 7"  "Series 8"  "Series 9"  "Series 10"
#> 
#> $Fitted
#> Time Series:
#> Start = 1 
#> End = 80 
#> Frequency = 1 
#>  [1] -0.044173243 -0.267361320  0.317130878  0.112423988  0.026106361
#>  [6]  0.174271661 -0.025925372 -0.278906380 -0.520899115  0.129062149
#> [11] -0.298349257 -0.008705583 -0.082425064  0.030612038 -0.680765634
#> [16] -0.146337944  0.095779915  0.163247415  0.226456160 -0.457108212
#> [21]  0.086002222  0.367779256 -0.043353049  0.384194899  0.133443240
#> [26]  0.050644328  0.158498503 -0.618387995 -0.583679772 -0.248725769
#> [31]  0.161718355  0.024021768 -0.332255866  0.412071758 -0.108168128
#> [36]  0.468645118  0.027701518 -0.621343147 -0.222524086 -0.775665352
#> [41] -0.067560145  0.246576078  0.013227233 -0.356353872 -0.288611044
#> [46] -0.147875886 -0.246448305 -0.166522683 -0.228333825 -0.272782337
#> [51] -0.433409947 -0.065280698 -0.105106086 -0.205755707 -0.244490727
#> [56] -0.384678509 -0.045057387 -0.336099945 -0.123139606 -0.045739198
#> [61] -0.476160992 -0.330674020 -0.444002304 -0.540750526 -0.374230754
#> [66] -0.402918669 -0.530493564 -0.271121607  0.078908162 -0.057526416
#> [71]  0.017532267  0.232104062  0.408778198  0.002270658 -0.329776273
#> [76]  0.534331518  0.589032478 -0.499909303 -0.021920490 -0.479764219
#> 
#> $Accuracy_Train
#>                     ME     RMSE       MAE      MPE    MAPE      ACF1 Theil's U
#> Test set -2.969088e-17 1.141543 0.9383163 355.6106 374.737 0.2133673  1.252081
#> 
#> $Input_Data
#> $Input_Data$Actual_Train
#> Time Series:
#> Start = 1 
#> End = 80 
#> Frequency = 1 
#>  [1] -1.541826390  0.773002880  0.487531964  0.178393460 -0.923606427
#>  [6] -0.253597462  0.638103993  0.382320075 -1.255662345 -0.650728871
#> [11] -1.493193245  1.160338189  0.455165981  0.955826032  0.359364524
#> [16] -0.028823078 -1.971234964 -1.195571616 -0.009020818  0.494183074
#> [21]  0.248420186  0.149623216  2.057033527 -0.398318253 -1.298431511
#> [26] -0.049131313  0.781283578  2.458614064  1.001553071 -1.020365211
#> [31] -0.695242053 -1.063623038 -0.455460665 -0.869726405 -2.467041559
#> [36] -1.175795448 -0.987050069  0.294136724  2.376871649 -0.645460605
#> [41] -0.516268778 -1.583576846 -0.174756197 -2.016767882  0.692724243
#> [46] -0.398988011 -1.286793073 -1.223062889 -1.111333701  0.351989798
#> [51] -2.211100501 -1.283910408  1.021198955  0.001285460  0.982247399
#> [56] -1.063700999 -0.115923363  1.741668734 -0.955842309  0.524433299
#> [61]  0.626373400 -0.966326089  0.070059144  0.392404536  0.117934881
#> [66]  1.330731932  0.227375618 -0.506615171  0.447135591  1.627108103
#> [71] -0.280043874  1.136901960  0.834057146 -2.575019606 -0.677689895
#> [76] -0.049648882  1.724863987 -0.171572570  0.535735781  0.764873090
#> 
#> $Input_Data$Forecasts_Train
#> Time Series:
#> Start = 1 
#> End = 80 
#> Frequency = 1 
#>      Series 1     Series 2    Series 3    Series 4    Series 5    Series 6
#>  1  1.4581024  0.295322041  1.80887652  0.43337574  0.60542543  1.49202009
#>  2  1.1748092  1.119089180  2.79607311  0.59498802 -1.15355858  0.62412356
#>  3  1.5616182  0.668610426  4.08689327  0.47893423  0.12874426  2.35434541
#>  4  2.1694207  1.703582624  1.32796650  2.41054629  0.93869374  2.71987875
#>  5  0.8519223  1.528391727  0.89433924  1.57089544 -0.80455688  1.72946329
#>  6  1.3310973  0.774067287  1.50663176  2.05449968  2.12201031  1.87057538
#>  7  1.4264978 -0.232467030  2.16997286  1.66698602 -0.06545074  2.07099473
#>  8  1.2486235  0.166102719  1.75438825  1.40095596  0.15322438  0.14043252
#>  9  0.1145385  0.538350791 -0.46262571  0.07704928  0.19970407  3.78097086
#> 10  1.4564305  1.620959383  0.89158549  1.70914830  1.03175689  0.20826861
#> 11  1.4561762  2.570296716  1.77612964  0.97464603 -0.28234303 -1.41083226
#> 12  2.7488128  0.120588888  1.60559391  1.79887958 -1.32753861  0.71705911
#> 13  1.0475970  1.318067458  1.48277723 -0.86034371  1.09259009  2.13100508
#> 14  1.5020839  1.025875535  2.68051058  1.69569188  1.88956465  0.39999177
#> 15  1.0057179 -0.095246613  0.84918095 -1.34254848  1.06104341  0.98678082
#> 16 -0.3846600  0.070810349  1.67959992  1.78466037  0.63662053  1.49708319
#> 17  0.6403744  2.028420789  1.44990018  1.86845871  2.09508635  3.00331877
#> 18  1.5830267  1.511972624  1.07943202  1.65508233  3.30713242 -0.31684552
#> 19  1.2764839  2.547562621  1.34625071  0.50447732  0.11890115  2.30198163
#> 20  1.5899776  2.803318212  1.35028358 -0.48162430  0.91015839 -0.93966990
#> 21  1.9978452 -0.667684828  1.04273343  1.39433271  0.85275107  0.36787067
#> 22  2.3016948  1.230666303  1.11790671  1.01099831  1.37659630  0.80674827
#> 23  1.7733545  2.241133429  0.60877381  2.68863401  1.51699471 -0.91908999
#> 24  2.0346831  1.168681767  1.84974677  1.46577370  3.50302166  1.34200138
#> 25  2.0254720  2.174129559  1.64126774  2.48538161  0.03595806  0.97836339
#> 26  0.3474148  1.158000377  0.14667947  0.54680140  2.92823762  0.59872690
#> 27  0.6116130  1.775553531  2.45024674  1.93924409  0.87791464  0.93007992
#> 28  0.6565888  1.014091334  0.41652238  1.36907388 -0.22760185  1.15145220
#> 29  1.1416242  0.717369530 -0.51547176  0.30513268  1.54829456 -0.92894477
#> 30  1.5298778 -0.208235904  2.45072146  0.46651217  0.24624931  0.59641570
#> 31  1.1688995  2.407412134  0.12146327 -0.18985935  2.33527827  1.57128851
#> 32  0.8709273  1.227540728  1.47516763  1.79849982  2.19244703  1.88946649
#> 33  0.7960543  0.841147099 -0.72687657  0.27704949 -0.03460444  2.40219039
#> 34  1.9499072  1.455855929  1.35481554  1.25756241  2.52570674  2.16868906
#> 35  0.2061256  1.699231005  1.47661284  0.41047774  2.85002566 -0.89339810
#> 36  2.1349152  2.458685309  2.10021776  1.63144331  1.25499421  1.09715707
#> 37  1.3263366  2.678513788  0.89369368 -0.17472543  0.62804380  1.25676544
#> 38  0.5302766  0.559343076  0.38862690  0.41630061  0.19508870  0.39324481
#> 39  1.7385470  1.346396311  0.01600350  0.90209834  1.18455543  2.75062459
#> 40  2.1096022 -1.118092762  1.30980935 -0.20774780 -0.09776203 -0.03996979
#> 41  2.7754676  1.268025444  0.06627182  2.54849216  0.17657368  1.93689116
#> 42  0.8925396  1.386697624  2.07371086  3.51840061  0.24063710 -0.43090043
#> 43  2.2848751  1.621078262  1.51700555  2.61074834 -0.88811745  1.06586736
#> 44  2.0803738  0.987740516  0.31046253  1.09869387 -0.18465343  1.12886080
#> 45  1.2897888  0.774337295 -0.26797467  0.60095422  1.25822354  1.94598414
#> 46  0.6236876  3.043020997  1.43169581  0.60800714  0.98582735  0.60299562
#> 47  3.1670642  0.026279263  2.04395737  2.10854946  0.59412483 -0.31670836
#> 48  1.6262958  1.015397138  2.28146334  0.77289850  0.57848569  0.97865736
#> 49  0.7384270  1.495188990  0.68053500 -1.39299246  0.60299623  2.42101277
#> 50  2.0051896 -0.432592497  1.40799626  1.13497936  1.66440898  0.81961804
#> 51  0.4954618 -0.421748185  0.26565606 -1.01975858  0.56509675  0.93729185
#> 52  0.1710774  0.007644122  0.71857698  1.25590436  1.49890089  0.36921977
#> 53  1.3519370  1.649618182  0.85299354  1.39044639  2.19792245  1.16457261
#> 54  0.8635664  1.247651565  1.18970476  0.16886595  0.82923629  0.93902862
#> 55  0.7783463 -0.381419577  1.65248489  1.51955530  1.57452042  1.25309425
#> 56  0.3049434  1.645473264  0.89927940  0.15447831  0.46579504  1.92618354
#> 57  2.2950082 -0.710004047  1.85951673  0.08180097  1.75873620  1.56286290
#> 58  0.9311538 -0.105111075 -0.62651083  0.36923210  2.18599484 -0.47255842
#> 59  1.4229096 -0.490317651  1.53876609  1.77843529  1.36346559  2.22160218
#> 60  0.1444460  0.126863290  2.80184513 -0.11433499  0.84619404  1.80905394
#> 61  1.9023736 -0.319450204  0.76782778  1.17748068 -0.07004688  2.62131378
#> 62  1.0124890  0.372887322  1.75792924  1.54915725  0.97434666  0.65413923
#> 63  0.3724599  1.907897340  1.55227909  0.47249840 -0.50847489  1.34862440
#> 64  2.3895399  1.456003324  0.13604993 -0.64940047  0.42456760  0.63147601
#> 65  0.8715452  1.899749654  0.65236896  1.53437891  0.12824723  1.55451290
#> 66  0.5200283  2.530010506 -0.17972177 -0.92317091 -0.01130706  1.33360477
#> 67 -0.6108871  0.905760834 -1.38311598  0.97129788  0.62546602  1.42028568
#> 68  0.7988600 -0.134574733  0.95484131  0.23137474  0.19370802  0.37063699
#> 69  1.7105320  1.875055187  0.79435775  2.12438890  1.18907184  0.87790332
#> 70  0.7622105  2.295102981  0.38102140  0.63604118  2.30908303  0.44957670
#> 71  1.2019295  0.882356733 -1.15980205  1.73809545  0.86787826  1.48554864
#> 72  2.2257616  1.534351136  2.16117442  1.31692404  0.70855623  2.33201103
#> 73  0.5886528  1.401662921  0.03890743  1.01798004  2.82940773  1.06564227
#> 74  0.8262043  0.646449855  0.40212502  3.90208410  1.51692204  3.29193774
#> 75  1.4323397  0.584716950 -1.02034424  2.09950358  3.00778279  0.22874497
#> 76  0.3634122  2.936239452  0.50001533  0.73646903  3.31486107  1.91919617
#> 77  2.4912990  0.909089606  3.33199706  1.82278033  1.60152954  0.67844532
#> 78  0.4772386  0.703365297  0.73329844  2.04418690  0.18419846  1.67468694
#> 79  2.2945369  0.913769856  0.30753420  1.64481258  1.69110470 -0.46689596
#> 80 -0.8796096  1.270238393  1.65100530 -0.70285587  2.04478360  1.56599983
#>       Series 7    Series 8    Series 9   Series 10
#>  1  1.44742147  0.29113985  0.83692862  2.74865145
#>  2  2.49675135  0.61616844 -0.21518875  1.13069396
#>  3  1.26498093  0.89877423  1.40330955  2.18237586
#>  4  0.98555987  0.33621988  0.18003304  0.13761257
#>  5 -0.57493885  3.40779880  1.43795481  2.01485846
#>  6  1.03815098  1.60221822  1.11374859  0.13100200
#>  7 -1.97217641  2.68468461  1.80100306  1.92358098
#>  8  0.12306261  0.70254090  0.31673184  2.88750551
#>  9  1.60128937  1.30216084 -0.82753660  0.28665283
#> 10  2.04683552  0.67792657  1.22018476  2.29569604
#> 11  1.23670427  1.51647160  1.63085178 -0.35896479
#> 12  2.05047231  0.15199453  1.59502649  2.42836336
#> 13  0.27985708  0.76569822  1.10476537  2.76093742
#> 14  0.90164047  0.50138522  0.17245610  1.23902275
#> 15  1.27048526  2.25269011 -1.27042372  0.39476288
#> 16  0.92578763  1.41146436  1.48985358  1.13071862
#> 17 -0.09155494 -0.03394171 -0.44881241  1.97837061
#> 18  1.57651113  0.91136466  0.36362743  1.72937017
#> 19  2.05103791  0.02869772  1.61475532  2.46736323
#> 20  0.58015548  1.11357386  0.88424853 -0.24242212
#> 21  1.60591965  2.89103347  0.64726154  2.54020940
#> 22  2.03606238  2.16780272  2.68075882  1.09138947
#> 23  0.16006346  0.57026861  1.20725295  1.51667814
#> 24  2.52044145  0.58493084  0.84734635  0.41979375
#> 25  3.05108009  0.04913940  0.97792301 -0.13436036
#> 26  1.94657051  2.33539297  1.28784128  1.12983030
#> 27  0.21654362  0.09137472  2.86600740  1.67456822
#> 28  2.09777050 -0.82701133  0.71819507 -0.59954875
#> 29  3.12993156  0.89566824  1.23392137 -1.16558368
#> 30  0.50426042  3.06370243 -0.95516305  1.52214452
#> 31  1.13378458  2.45666950  1.05701211  1.56112242
#> 32 -1.25244569  1.24434463  0.60217971  1.80367593
#> 33  1.32977917  1.42922421  0.94399466  1.40292124
#> 34  1.61701429  2.39130239  0.23654209  0.98897639
#> 35  0.45799056  0.77963242  1.24823809  2.48532499
#> 36  1.29994623  3.05644001  1.98998712 -0.30521432
#> 37  2.00484448  0.96005747  2.72644567  0.24724679
#> 38  1.33954365  1.68004516 -0.31536164  0.45594164
#> 39 -0.12837807  0.68962273  0.91775584  0.30915906
#> 40  1.46110120  0.73050318  0.22593908 -0.06916646
#> 41  1.79862889  0.15037500 -0.37998189  0.80713162
#> 42  2.62296322  0.71401057  1.89078673  1.31246466
#> 43  1.28286478 -0.16423325  1.50005920  1.20594435
#> 44 -0.46228763  0.27316651  1.54884818  1.63280574
#> 45 -0.21769676  1.76480291  0.42915190  1.39372912
#> 46  1.25184115  0.97318182  0.80036235  0.14515140
#> 47  1.41326017 -0.04372242  0.25535176  0.18309771
#> 48  1.31286575  1.01728912  0.89613878 -0.17679794
#> 49 -0.13867139  1.20950199  1.15917207  2.93566324
#> 50  0.73133721  0.03366900  0.56036241  1.19098285
#> 51  3.01468258  2.78725299  1.33216256 -0.10959384
#> 52  4.40945488  0.43779186  1.79605596  0.66371132
#> 53  1.72589893 -0.58874392  1.63792817 -0.42410878
#> 54  2.08406169  0.69286323  0.77114363  1.11907351
#> 55  1.81194978  0.57208395 -1.40671159  1.72254673
#> 56  1.13874083  0.40632796  0.63082899  0.50387012
#> 57  0.89850117  1.46900842  0.91273148  1.37714403
#> 58  1.19658025  1.81249252  3.49527695  0.08364813
#> 59  1.36045016  0.39128066  0.74079132  0.24179017
#> 60  1.71110980  0.80760135  1.15548010  2.08728589
#> 61  0.11304184  1.07830933 -0.09250935 -0.09418779
#> 62 -0.78725339  1.71680340  0.59970766  0.56986021
#> 63  0.67811873  0.47129848  0.62832691  0.52739612
#> 64  1.26428062  0.17317303  1.08012199 -0.05531191
#> 65  0.43209075  0.89184640  0.63419427 -0.46458122
#> 66  2.10123551  1.05607656  0.93059020  0.76307495
#> 67  2.06970941  0.17819161  1.40791824  1.03235157
#> 68  2.11756080  3.05028321  0.65561835  0.99375804
#> 69  0.13176807  2.41581576  1.01386383  0.49535438
#> 70 -1.10790563  2.86751282  1.20930430  1.47095940
#> 71  3.13124718  1.68841780 -0.45928146  2.56120814
#> 72  2.99120145  0.99701576  0.89702370 -0.80898748
#> 73  1.59666108  2.25668115  3.03594538  2.24757325
#> 74  0.24901170  1.12165929 -0.86795827  0.34923141
#> 75 -0.07295951  0.18273049  1.13058558  0.95467814
#> 76  1.84272209  0.49064251  3.10626719  2.15025682
#> 77  3.07187756  1.43163086  1.59916644  0.89672292
#> 78  0.69855000 -0.22248941 -0.65635607  0.97484951
#> 79 -0.93850080  3.23335837  0.91566920  2.00708963
#> 80  0.63710546  0.75122452 -0.66375542  1.22330392
#> 
#> $Input_Data$Actual_Test
#>  [1] -0.95262390  0.30807936  0.38740109 -0.60278300  1.08398798 -1.07821652
#>  [7]  0.88857733 -1.12467826  0.56274446 -0.37992467 -0.45792974 -0.15015838
#> [13] -1.02456171 -0.22594046  0.20514499 -0.16299230 -0.09093013  2.04480031
#> [19] -0.04121660 -0.64704005
#> 
#> $Input_Data$Forecasts_Test
#>          Series 1    Series 2   Series 3   Series 4   Series 5     Series 6
#>  [1,]  1.29187192  2.02780387  2.1983748  1.3844619  0.5903510  1.326430294
#>  [2,]  2.40421475  1.25796037  1.3787277  1.1533839  2.1017226  1.346628776
#>  [3,]  1.40551176  1.36642089  0.9326244  1.7441460  2.6796755  1.019316699
#>  [4,]  0.72155790  2.46881989  1.5604005  1.3354670  0.1490883  1.688894695
#>  [5,]  0.01474742 -0.13614527  2.9959845  2.7300101  1.3021082 -0.630031658
#>  [6,]  2.44440038  0.50732702  0.9239109  2.1421840  1.9469220  1.163411600
#>  [7,]  1.31570082  1.38268097  1.8080551  0.9475185  2.8292827  2.614134175
#>  [8,]  1.55749607  1.48463796  0.4836805  0.5930117  1.3006919  1.447730784
#>  [9,]  1.96145554  0.47574142  1.0514562  0.1731203  0.1209056  1.483797932
#> [10,]  1.33799932  1.04498342  1.2086729  0.3802640  1.5040007 -0.802311586
#> [11,]  0.53150646  1.19567585  0.2778696 -0.2413099  0.8048701  1.010899838
#> [12,]  1.17253640 -0.09492323  2.8128926  1.5827318  1.8034500  0.652432196
#> [13,]  1.04943398  1.84438922 -1.6147970  2.1610015  0.1803457  0.428978808
#> [14,]  1.62373918 -0.58392531  2.5054219  0.5739342 -1.4304187  0.868191868
#> [15,]  2.39312530  2.82294230  0.7265964  1.0532808  1.9913070  2.910318311
#> [16,]  0.33781842  0.95985855  0.4987037  1.9294043  0.4361207  1.203381598
#> [17,]  1.77008194  0.07877925  0.6225584  0.3358177  2.4397394  0.255220312
#> [18,]  1.88986699  1.64888805  1.3532341  0.1628779  0.8362233  0.530204085
#> [19,]  1.59987538 -0.06471897  1.5631547  0.6943394 -0.7798312 -0.009502284
#> [20,] -0.40200788  0.64148941  2.6865573  1.0617708  2.3181819  1.327770721
#>          Series 7   Series 8   Series 9   Series 10
#>  [1,]  2.04306730  3.8005026  1.4840957  1.74673837
#>  [2,]  1.83523851 -1.4891750  1.0604779  1.97191685
#>  [3,]  1.48032749  1.1339088  0.1930517  1.01028866
#>  [4,]  0.23461273  1.1729035 -0.2321511  0.40207793
#>  [5,]  0.72057283  1.6213578  0.2035393  1.62094918
#>  [6,] -0.61769524  1.0414846  0.3970283  0.89280002
#>  [7,]  1.20980671  1.1809202  2.4106901  1.71692481
#>  [8,]  2.52698170  2.1365925  1.5580513 -0.01546665
#>  [9,]  0.38531305  0.3510019  1.2022808  1.77327345
#> [10,] -0.71715408  0.2501373  1.5372892 -0.15015568
#> [11,]  2.36126500  0.7720822  2.0558563  1.09077936
#> [12,]  0.52762050  2.3641599  2.5316783  0.31503576
#> [13,]  0.39708320  1.6156517  0.6106443  0.83903665
#> [14,]  2.00035266  0.3506853  1.0256722  0.89003830
#> [15,] -0.09986257  1.0592063  0.3825066  0.33979180
#> [16,] -0.80145436  1.6017145  2.9110351  2.59919440
#> [17,]  0.53755847  0.1150179  2.0521937  0.67287488
#> [18,] -0.25059469  1.2873018  0.4255367  0.05084726
#> [19,]  2.28402743  2.2294391  1.1046685  2.12285566
#> [20,]  2.53703579  0.9984527  1.9303857 -0.06345866
#> 
#> 
#> $Predict
#> function (object, newpreds) 
#> {
#>     pred <- as.vector(object$Intercept + newpreds %*% object$Weights)
#>     return(pred)
#> }
#> <bytecode: 0x560496024318>
#> <environment: namespace:ForecastComb>
#> 
#> $Intercept
#> [1] -1.188348
#> 
#> $Weights
#>  [1] 0.09271550 0.09835513 0.10340386 0.10943230 0.10202297 0.10115132
#>  [7] 0.09693824 0.10041778 0.09117347 0.10438944
#> 
#> $Forecasts_Test
#>  [1]  0.60143930  0.10860761  0.12058579 -0.22375787 -0.10227521 -0.09144148
#>  [7]  0.54823772  0.09607904 -0.30228024 -0.64254855 -0.22703999  0.17734079
#> [13] -0.43435151 -0.41592299  0.16614827 -0.01380787 -0.31962914 -0.40517109
#> [19] -0.12167715  0.11785356
#> 
#> $Accuracy_Test
#>                   ME      RMSE       MAE      MPE    MAPE
#> Test set -0.00473246 0.8570006 0.6155973 46.79545 106.789
#> 
#> attr(,"class")
#> [1] "foreccomb_res"