Trimmed Eigenvector Forecast Combination

Computes forecast combination weights according to the trimmed eigenvector approach by Hsiao and Wan (2014) and produces forecasts for the test set, if provided.

comb_EIG3(x, ntop_pred = NULL, criterion = "RMSE")

Arguments

x

An object of class foreccomb. Contains training set (actual values + matrix of model forecasts) and optionally a test set.

ntop_pred

Specifies the number of retained predictors. If NULL (default), the inbuilt optimization algorithm selects this number.

criterion

If ntop_pred is not specified, a selection criterion is required for the optimization algorithm: one of "MAE", "MAPE", or "RMSE". If ntop_pred is selected by the user, criterion should be set to NULL (default).

Value

Returns an object of class foreccomb_res with the following components:

Method

Returns the used forecast combination method.

Models

Returns the individual input models that were used for the forecast combinations.

Weights

Returns the combination weights obtained by applying the combination method to the training set.

Top_Predictors

Number of retained predictors.

Ranking

Ranking of the predictors that determines which models are removed in the trimming step.

Fitted

Returns the fitted values of the combination method for the training set.

Accuracy_Train

Returns range of summary measures of the forecast accuracy for the training set.

Forecasts_Test

Returns forecasts produced by the combination method for the test set. Only returned if input included a forecast matrix for the test set.

Accuracy_Test

Returns range of summary measures of the forecast accuracy for the test set. Only returned if input included a forecast matrix and a vector of actual values for the test set.

Input_Data

Returns the data forwarded to the method.

Details

The underlying methodology of the trimmed eigenvector approach by Hsiao and Wan (2014) is the same as their standard eigenvector approach. The only difference is that the trimmed eigenvector approach pre-selects the models that serve as input for the forecast combination, only a subset of the available forecast models is retained, while the models with the worst performance are discarded.

The number of retained forecast models is controlled via ntop_pred. The user can choose whether to select this number, or leave the selection to the inbuilt optimization algorithm (in that case ntop_pred = NULL). If the optimization algorithm should select the best number of retained models, the user must select the optimization criterion: MAE, MAPE, or RMSE. After this trimming step, the weights and the combined forecast are computed in the same way as in the standard eigenvector approach.

The trimmed eigenvector approach takes note of the eigenvector approaches' property to treat \(y\) and \(\mathbf{f}\) symmetrically, which bears the risk that the (non-trimmed) eigenvector approaches' performance could be severely impaired by one or a few models that produce forecasts much worse than the average.

Author

Christoph E. Weiss and Gernot R. Roetzer

References

Hsiao, C., and Wan, S. K. (2014). Is There An Optimal Forecast Combination? Journal of Econometrics, 178(2), 294--309.

See also

Examples

obs <- rnorm(100)
preds <- matrix(rnorm(1000, 1), 100, 10)
train_o<-obs[1:80]
train_p<-preds[1:80,]
test_o<-obs[81:100]
test_p<-preds[81:100,]

## Number of retained models selected by the user:
data<-foreccomb(train_o, train_p, test_o, test_p)
comb_EIG3(data, ntop_pred = 2, criterion = NULL)
#> $Method
#> [1] "Trimmed Eigenvector Approach"
#> 
#> $Models
#>  [1] "Series 1"  "Series 2"  "Series 3"  "Series 4"  "Series 5"  "Series 6" 
#>  [7] "Series 7"  "Series 8"  "Series 9"  "Series 10"
#> 
#> $Fitted
#> Time Series:
#> Start = 1 
#> End = 80 
#> Frequency = 1 
#>  [1]  0.63503230  0.33805443  0.64102692  2.65119043  0.05954581  0.55406611
#>  [7] -0.13960718  1.94750673  0.08882156  0.83471377 -0.09687375  1.16347100
#> [13]  0.81486572  0.08142248  0.79825774  0.29632978  0.64970535  0.96441569
#> [19]  1.94193515 -0.04891691  0.79860225  0.59368303  1.48903318  1.17638849
#> [25] -0.15090166  1.36901373  0.98414862  0.10430382  0.19239848 -0.32736248
#> [31]  0.17452411  0.74800885  0.22554550  0.35824612  1.68491723  2.06302299
#> [37]  1.20398950  0.72143860  0.53305632  0.20278101  0.89115335  0.52735744
#> [43]  0.46557757  1.80439428  1.27438433  1.23184491  1.21633348  0.71046145
#> [49]  1.58076138  0.34719457  0.60358813  0.88571270 -0.27823806  0.99690177
#> [55]  0.99063847  0.33501838  1.71507531  1.36399077  0.96032172  1.51260807
#> [61]  1.09056305  1.63279302  1.08639215  1.22928367  0.51676638  0.72766959
#> [67]  0.62812900  1.33624693  0.11999620  1.17100728  0.78118925 -0.56650059
#> [73]  1.95125431  0.43524071  0.87535202  0.53674631  0.86319405  1.48996098
#> [79]  1.88145997  0.56817535
#> 
#> $Accuracy_Train
#>                  ME     RMSE      MAE       MPE     MAPE      ACF1 Theil's U
#> Test set -0.9145634 1.333001 1.075919 -119.7907 407.7162 0.0795522 0.9743614
#> 
#> $Input_Data
#> $Input_Data$Actual_Train
#> Time Series:
#> Start = 1 
#> End = 80 
#> Frequency = 1 
#>  [1]  0.65952504  0.10661220 -0.66750149  0.36274702 -0.79073513 -1.79370053
#>  [7] -1.47594035  0.19567255 -0.12726827 -0.26861195 -0.22905991  1.93124775
#> [13] -0.41342456 -0.77390259  1.64268542 -1.58987683  0.92924303  0.35594842
#> [19]  0.27144078  0.22346460 -1.04512000  0.77576517  0.02763014  1.15042855
#> [25]  0.25342222 -0.82488682  0.28676319  0.94292281 -0.53107640 -1.29761424
#> [31]  1.22093399  1.47740140 -1.20779824 -0.94193456 -0.58351572  1.13409033
#> [37] -1.75717206  0.05231157 -0.45666876 -0.98754657  0.02126624  0.35400796
#> [43] -1.16106179 -0.47583428  0.11535655 -0.36688686  0.15210927 -0.81243120
#> [49]  0.78293216 -0.44680648 -0.24503039  0.94642071 -0.19796637  1.67688667
#> [55] -0.04331819  0.20424264  0.83721977  1.59664621 -1.34003661  0.10078513
#> [61] -1.81409409  1.23873179  0.65207192  0.77326509  0.42004253 -0.31398419
#> [67] -0.59098124 -0.94971878 -0.31486991 -0.15058437 -1.47426058 -0.55535865
#> [73]  1.83955343 -0.16321110  0.15279233  0.42854347 -2.16824934 -1.59482757
#> [79]  0.05326264 -0.75879443
#> 
#> $Input_Data$Forecasts_Train
#> Time Series:
#> Start = 1 
#> End = 80 
#> Frequency = 1 
#>        Series 1     Series 2    Series 3      Series 4     Series 5
#>  1  0.151472531  1.013408091  0.12088238 -0.2540799447  1.415925606
#>  2  2.122706843  0.494339783  0.19106590  0.8704015498 -1.024651525
#>  3  0.579443812  1.927790118  1.28493468  1.4374770132 -0.017949370
#>  4  0.700750327  0.908519455  2.61245634  0.2410458448  0.843040578
#>  5  0.320435523  1.277478367 -0.37700142  1.7565662622 -0.155023683
#>  6 -0.531963979  0.362952856 -0.15491000  0.3541375504  1.532480225
#>  7  0.797781887  2.808918928 -0.75572848 -0.0350754997  0.778507968
#>  8  2.713334564  1.939060155  2.89781769  0.9712840457  0.467245664
#>  9  1.309577547 -0.385602517  0.69758521 -0.4097690617  0.747243841
#> 10  0.075129886  1.893849557  0.25128882  1.0282804313  1.529752468
#> 11  0.970710520  1.430845693 -0.55368869 -0.3441219972  0.640985802
#> 12  0.885110546  1.031233641  1.00023941  0.7486464757  2.837281304
#> 13  1.118657464 -0.670839387  0.89421588  3.6256685533  0.483101004
#> 14 -0.253492456 -0.951911519 -0.12275891  0.2109346956  1.423808570
#> 15  1.258100081  3.080725168 -0.42797576  0.4475133920 -0.443387802
#> 16  1.218056285  1.210062767  0.24897637  1.2573054199 -0.542972388
#> 17  0.446808013  1.023184942  1.12052624  0.8593242429  0.913927702
#> 18  1.229864210 -0.874587227  1.03982290 -0.3868977646  0.971776054
#> 19  0.552129368  1.678432264  1.08777483  1.7656994621 -0.129392933
#> 20 -0.460672253  0.300630389  1.19854754  0.3367849668  0.249815910
#> 21  1.212417504  1.281414194  1.26232104 -1.1959759257  0.692213769
#> 22  1.995477475  0.095266863 -0.25795301 -1.4400503125  0.362075850
#> 23  1.414672297 -1.544338006  3.01132460  1.0349441277  1.773900348
#> 24  0.430851211 -0.225336598  0.92438315  0.5950828432  1.114979379
#> 25  1.925684094  2.080655681 -1.10454402  1.1411714217  1.463525937
#> 26  2.340730910 -0.463168197  1.47041801  0.8440711514 -0.331105997
#> 27  0.004302507  0.887831310  0.61488530  1.3355255778  1.572895557
#> 28  0.321466803  1.619369237 -0.77035016  0.4622197947  1.119911460
#> 29  0.026281530  1.631459147  0.83854358  0.7889533062  1.131006114
#> 30  1.476279518  1.502040829 -0.20064205  1.6550183101  0.456254190
#> 31  2.037909905  2.014647260  1.24268292  0.6884403011  1.365803879
#> 32 -1.042654413  1.089686492  1.64791279  0.6459110426  2.517514811
#> 33  1.412107058  0.907026678  0.13212091 -0.0588961871 -0.516691371
#> 34  1.163175390 -0.472882341 -0.09613399  0.0737987700  1.607296730
#> 35  0.665722927  2.250331645  1.45791975  2.0728571142  1.401802766
#> 36  3.992526250  1.386094563  2.02898598  0.9671663797  0.383021297
#> 37 -0.393533485  2.239934028  1.71205762  3.1145051894  0.908360338
#> 38 -0.467871941  0.486770731  0.63947275  1.1451636701  1.700486018
#> 39  0.957779358  0.915812770  0.01672906 -0.0907214339  1.753927432
#> 40  0.126245571 -0.099619191  0.38957533  0.3586871739  0.878359082
#> 41  1.310207176  0.445827514  0.72486466  1.5947005355  1.810713363
#> 42  0.603273834  0.118449207  1.18076552  3.4428498598  1.131110184
#> 43 -0.308479659  2.260327059  1.29476417  1.2475594147  1.328237414
#> 44  2.032526436 -0.687740641  1.77807247  2.0673434407  0.835059046
#> 45  0.611108512 -0.007116978  1.46852311  0.5894230508  1.406431103
#> 46  1.099776715  1.466802913  1.58964375  0.0911021050  1.772992972
#> 47  0.592761843  0.546267521  0.09171979  0.5104670835  1.528178306
#> 48  0.604042612  1.615804887  0.21667328  2.7706074627  0.813385084
#> 49  2.739698442  0.535389931  1.49580250  0.5395884386  2.553277176
#> 50  0.536585066  1.773118328  0.02710424  1.2465064009  1.745742277
#> 51  0.582438685  2.228069177  1.46629754  0.0002507064  0.176174085
#> 52 -1.325645178  2.569000174  0.87126797  1.2765163996 -0.008883847
#> 53  0.505135206  1.236422061  0.28373083  0.4484378328  1.215541427
#> 54  0.664636086  0.805445827  0.39624111  2.1234639080  0.358713559
#> 55  3.086589343 -0.129649222  1.20024429  2.1637756624  1.737906965
#> 56 -0.379684310 -0.530536417  0.26523855  1.1803259841  0.800881470
#> 57  1.296460677  1.049470587  1.40226515  0.3474130575  1.712094088
#> 58  1.377309248  1.372637377  0.41014673  1.5503843404  2.532356486
#> 59  0.568257666  2.188422462  0.29781627 -1.7754313359 -0.273104069
#> 60  1.760643633  2.975271353  0.87132264  1.3076926059  0.955366926
#> 61  1.727781691  0.400438597  2.22698840 -0.6108705188  0.141115684
#> 62  2.053069692  1.201069867  1.35451391  2.3264766314  1.235039854
#> 63  0.516789317 -0.147620335  0.84544448  2.1942884387  1.986521627
#> 64 -0.621325569  1.431092717  2.36699882  2.0650927152  1.514561896
#> 65  1.952000711  0.265582842  0.82397053 -0.6124789363  1.399861789
#> 66  0.222929282  1.643590286  0.48064516  1.2371792136  1.169147677
#> 67  0.318271814  1.665469208 -0.08470826  1.0373938290  0.599935063
#> 68  1.790360054  1.085219813  0.91663224  0.6243065282 -0.052477221
#> 69  0.227242872  3.390571763 -0.33158502  3.4713339393  2.099087561
#> 70 -1.537842879  1.912621164  1.53409599  0.9721424288  1.381196645
#> 71  0.138840082 -0.911716874  1.94557562  1.5416670523  1.913229954
#> 72  0.970370010  1.120069457 -0.89282039 -0.3298683035  0.281352829
#> 73  1.555476516  0.193175120  1.88315408  1.7394004833  1.963559790
#> 74  1.094084434  1.826179569  0.54946520  0.3257943890  2.483925346
#> 75  2.073440856  1.292909813  1.25180893  1.7599276984 -0.330591635
#> 76 -0.824516401 -1.141012240  0.77861978  1.5690544557  1.715165807
#> 77  0.780408468  1.338841397  0.14555012  0.8447868820  0.541760522
#> 78  2.255763334  2.373187969  1.08424866  0.7999709883 -1.751203805
#> 79  1.725655635  1.105553385  1.49244422  2.7536625497  0.787755555
#> 80  1.680350577 -0.396527197  1.35951074  1.4877593050  0.975173861
#>         Series 6    Series 7     Series 8    Series 9   Series 10
#>  1  0.3605532883  1.16000563  1.873514072  1.77450066  0.01972010
#>  2  0.4889706994  0.48813723  2.364377026 -1.99837755 -0.48772182
#>  3  1.6980538794 -0.01643577  1.314396547  1.79438333  1.17613907
#>  4  2.4096658214  2.69073992  0.998537136  2.05427431  1.85704103
#>  5  0.4833953940  0.50528281  0.158292487  0.92918755  1.08081510
#>  6  0.6378440700  1.27796691  1.471142795  0.34894456  0.59874574
#>  7  0.7846578810  0.48948412  1.217070444  1.28381845  1.49644467
#>  8  0.1201274321  0.97719072  0.714820634  0.55868038  1.43946666
#>  9  0.8689678677 -0.53275721  0.543927860  2.95156309  1.14408278
#> 10 -0.8173204543  1.43042042  2.459305603  0.11725498  0.02228008
#> 11  1.8429072868  0.36955763 -0.497385873  0.62366080  2.32038645
#> 12  0.2296776375  1.33013879  1.228377157  1.04821772 -0.72477142
#> 13  0.4140122097  0.73384516  0.355667332  1.19234572  1.31288096
#> 14  0.4569673670  0.28990211  0.703011673  0.22370266 -0.26674946
#> 15  0.4564686495  2.05030473  1.795685085  1.83534714  1.39388991
#> 16  0.6294761677  0.34468002  1.733612455  0.20613597 -0.15601862
#> 17  2.1168908993  0.16897319 -0.486459087  0.26644827  0.74566580
#> 18  1.0098584217  0.88742107  0.419621562  1.78628216  0.14587413
#> 19  3.2123073883  2.81407644 -0.378337789  1.01229685  1.10197398
#> 20 -0.8666243408 -1.32264179  1.619948882  0.62791509  1.24500053
#> 21 -1.2810818515  0.32512169  0.931983444 -1.45805176  1.22113762
#> 22 -0.0239834601  1.46324690 -0.431712181  2.31313652  1.18554812
#> 23  1.7660312538 -0.06530407  1.295049092 -0.28826730  2.10749394
#> 24  0.6136927063  1.43369880  1.843869382 -0.18152952  0.70992803
#> 25  0.0552873160  0.82281588  1.507632197  2.07885148  1.63101794
#> 26  0.8773136323  1.26547479  2.790921825  0.31568290  1.59555660
#> 27  1.6773838674  1.36118532  1.706112408  2.54048670  0.73996438
#> 28 -1.0568175367  0.99737017  2.214991705  1.01349936  3.02804261
#> 29  2.4075544711 -0.46734866  1.025955311  0.80445469  1.42126837
#> 30  1.6216405456 -0.45675050  1.324512410  0.73595606  1.64288008
#> 31  2.5511646717 -0.91612056  1.127310021  2.24337142  0.77937121
#> 32  1.6807901989 -0.17083901  1.259764389  0.59709138  0.68720226
#> 33  1.6020473803  0.32093677  2.317789212  3.48695345 -0.10463531
#> 34  1.7121399408  0.82219140  0.901188476 -0.08427562  0.93428789
#> 35  0.5204641194  1.91669325  2.708286369  0.41937032  1.02399795
#> 36  2.2075508069  2.09777651 -1.002800614  0.67019084  1.02212389
#> 37 -0.2216243025  0.68522600  0.073997252  1.55382048  0.37867015
#> 38  0.3267962082  0.80512991  0.324856729  0.62628024  0.54923323
#> 39  2.8156262976  1.06025281  0.816206214  1.94893236  1.40188815
#> 40  0.7717962681  0.01205446  1.011599726 -0.59732569 -0.62438584
#> 41  2.0654497404  1.06094258  0.581196489  1.38322386  0.11639334
#> 42  0.3938646041 -0.13980556  0.050530629  0.55715473  3.42994891
#> 43 -0.0194319330 -0.38106427  1.721929969 -0.20347135 -1.08857016
#> 44  1.0488018032  1.83127019  1.984584178  1.65635207  2.78043421
#> 45 -0.2866644732  1.07615872  0.535754644  1.37354352  1.45864465
#> 46 -0.9105269603  0.86651402  0.010751689  1.56800063  2.55899971
#> 47  1.0871931811  2.36462147 -0.006781854  0.41840996 -0.93521876
#> 48 -0.7261553158  1.21464437 -0.836529589  2.69180102  1.11927490
#> 49  1.0598980517  1.66750874  1.422041456 -0.08995348  1.30446353
#> 50  0.9796372291  0.67402314  0.861853068  1.48342819  2.98081213
#> 51  1.2205164091 -0.27728222  0.850176195 -0.18191786  0.31307831
#> 52 -0.7861140186  0.90046151  1.779203687  0.21163123  0.50118658
#> 53 -0.0002125133 -0.85203698 -0.097631950  0.93807813  0.26356998
#> 54  1.1629329965  1.61020695  1.058288550  2.24276144  2.82220966
#> 55  3.3074041624  0.77662023  0.264497557 -0.01647552  1.71527182
#> 56  1.7701894360  0.40626715  1.451046144  0.79126527 -0.25074095
#> 57  1.6528480255  2.03447045  1.764802413  1.69055947  3.78813836
#> 58  0.2404110430  2.33791423  1.495385702  1.55212294  0.20266457
#> 59  1.1910747222  1.63677361  1.003277224 -0.06742067  0.70006755
#> 60  1.0193474388  2.16739323 -0.916557277  1.11779031  2.51336394
#> 61 -0.0852469417 -0.06978524  2.033981229  1.87635591  0.75918024
#> 62  1.4528089584  1.91693020  0.025624851 -0.12527427  1.14413962
#> 63  0.3082389703  1.33241202  0.421368817  1.16048075  1.98828647
#> 64  1.2323669613  0.06761843 -0.481448261 -0.32205567  0.97702514
#> 65  0.4480564862  0.20309526  2.331488730 -0.24639473  0.67644824
#> 66  1.2708810092  0.97989415 -0.547766374  0.50409628  1.39350893
#> 67  1.8066053428  1.35597223  2.636403207  0.06193129  0.28263538
#> 68  2.0379112035  1.76469494  1.144867024  0.97276441 -0.13180961
#> 69  1.2603076660  0.58108367  1.190522364 -0.17895430  2.55259133
#> 70  0.2485996762  0.80027516 -0.059709094  1.28457014  0.87134687
#> 71  0.9119120172 -0.40770867 -0.844124148  0.53098064 -0.03290005
#> 72  1.0596177340 -0.23331142  0.505592742  1.84321213  1.69881202
#> 73  1.7544568191  2.02078813  1.082733787  1.83989705  0.49067052
#> 74  0.7073674239  0.31861167  1.545246235  1.53410482  0.80463787
#> 75 -0.1484425351  0.49097030  2.331323078  1.53155431  1.35798600
#> 76  1.5934435782  0.28978116  0.062956947  2.35601577  1.52641209
#> 77  0.7558632658  1.59594513  0.654979097  1.18173410 -1.35211999
#> 78 -0.2576330953  1.90421397  0.831424457  2.38343682  1.97195649
#> 79  0.3283783288  2.27866491  0.369966630  1.38549271  1.22983095
#> 80  1.3244936983 -0.23981849  1.314751421  1.22293231  0.78425555
#> 
#> $Input_Data$Actual_Test
#>  [1]  0.19317915  1.03369055  0.22015524  0.74859739  1.64246824  0.21040983
#>  [7] -0.39034542  1.45705626  0.06762969  0.36352669  0.51152515  0.70749592
#> [13]  0.77731586  0.86176382 -0.46124224  1.30342018  1.28738897  1.22215264
#> [19] -0.48624692 -0.57686174
#> 
#> $Input_Data$Forecasts_Test
#>         Series 1     Series 2     Series 3    Series 4    Series 5     Series 6
#>  [1,]  1.3162485  2.724834145  0.016668192  0.45050258  1.55172864  3.546529236
#>  [2,]  0.1695632  0.310168342  1.038760901  2.87433861 -0.06004164  0.780795437
#>  [3,]  2.5436879  0.002946471  1.062661691  0.54502417  0.58505383  2.847830465
#>  [4,]  0.7341872  2.463030290 -0.237983228  1.39765051  0.59009299  0.470709543
#>  [5,]  2.3695464  0.017220731  0.551000134  0.52891580  0.90382670  2.662562580
#>  [6,]  0.8958885  0.072718380  0.091391568  0.26485005  0.85356599 -0.006537504
#>  [7,]  0.9380625  0.084358549 -0.291856586 -0.30671997  1.41725442  1.838305158
#>  [8,]  0.8550175 -0.221701962  1.923831617  0.38217442  1.98702416  1.005604226
#>  [9,]  0.4131958 -0.170347392  1.741200792  0.21495966  1.17600685  3.143180385
#> [10,] -1.8530265  1.285950966  1.152810733  0.99894353  2.02760864  0.249218064
#> [11,]  2.0594499  0.482757466  1.326920857 -0.41625294 -0.47964445  2.128919482
#> [12,]  1.3259487  1.652751303  0.516735948  1.10740376  0.98627107  1.769851636
#> [13,]  0.7922515 -1.226399612  0.726237140  0.03499527  0.90011350  0.534697447
#> [14,]  0.4812397 -0.123596614  1.367651584 -0.44962002  0.40494152 -0.742455464
#> [15,]  0.5705301  1.823502648  0.213812946  1.30988795  0.96013218  2.492284844
#> [16,]  1.2348936  1.741109307 -0.184753630  0.92949151  0.12517027  0.987736643
#> [17,]  2.2198820  1.763053604  2.085267892  1.25390965  2.38075452  0.251214682
#> [18,]  0.8236907  0.278150819  0.771768937  2.17440651  0.84343425  0.576389409
#> [19,]  2.1232110  1.539293364  0.009894599  0.31968079  0.75736235  0.247913105
#> [20,]  2.5875731  0.080710159 -1.125732954  1.25829595  0.83425932  0.295464151
#>         Series 7    Series 8    Series 9   Series 10
#>  [1,]  2.0768465  2.08087935  1.59250264  1.40434545
#>  [2,]  2.4831654 -0.25678452  0.45706602  1.45639006
#>  [3,] -1.0749606 -1.05577294 -0.69226426 -0.71608814
#>  [4,]  0.6644790  0.51717906  1.36770957  1.77108890
#>  [5,]  2.4107721 -1.08817513  1.03505141 -1.74767397
#>  [6,]  1.1150911  1.08584968 -0.37168515  0.09053362
#>  [7,]  1.8149618  1.41527559  0.36755303  0.12885263
#>  [8,]  0.8330834 -1.37910327  1.33061922 -0.61482836
#>  [9,]  0.6204937  1.98089318  1.90976672  2.88352139
#> [10,]  1.0946336  0.08054903  0.03573565 -0.31482633
#> [11,]  1.2400712  1.96807543  0.82928604  1.47672019
#> [12,]  0.3673112  1.30854934  0.83184329  0.70143854
#> [13,]  0.5628535  0.24648624  1.23315601  3.49971481
#> [14,]  1.1681421  1.12045620  0.07809631  0.68412285
#> [15,]  2.0598448  0.85311694  2.34312841 -0.99461349
#> [16,]  0.8608272 -0.99201557  0.95572862  0.64022795
#> [17,]  0.3187403 -0.78336246  2.11672547  1.95532035
#> [18,]  1.6135040  2.39712599 -0.52902076  1.75373353
#> [19,]  0.2843485  1.68961715  0.94671682  0.97109864
#> [20,]  2.2895182  1.03605098  1.33052312  0.32645497
#> 
#> 
#> $Weights
#>  [1] 0.0000000 0.0000000 0.5052079 0.0000000 0.0000000 0.0000000 0.4947921
#>  [8] 0.0000000 0.0000000 0.0000000
#> 
#> $Forecasts_Test
#>  [1] 1.036028062 1.753440784 0.004983157 0.208547927 1.471200507 0.597909941
#>  [7] 0.750580431 1.384138061 1.186683846 1.124025153 1.283948339 0.442801746
#> [13] 0.645396219 1.268935894 1.127214833 0.332591457 1.211204100 1.188252770
#> [19] 0.145692220 0.564106164
#> 
#> $Accuracy_Test
#>                  ME      RMSE       MAE       MPE     MAPE
#> Test set -0.3517301 0.7384866 0.5994236 -73.24683 201.9761
#> 
#> $Top_Predictors
#> [1] 2
#> 
#> $Ranking
#>  [1]  5 10  2  9  3  4  1  8  7  6
#> 
#> attr(,"class")
#> [1] "foreccomb_res"

## Number of retained models selected by algorithm:
data<-foreccomb(train_o, train_p, test_o, test_p)
comb_EIG3(data, ntop_pred = NULL, criterion = "RMSE")
#> Optimization algorithm chooses number of retained models for trimmed eigenvector approach...
#> Algorithm finished. Optimized number of retained models: 3
#> $Method
#> [1] "Trimmed Eigenvector Approach"
#> 
#> $Models
#>  [1] "Series 1"  "Series 2"  "Series 3"  "Series 4"  "Series 5"  "Series 6" 
#>  [7] "Series 7"  "Series 8"  "Series 9"  "Series 10"
#> 
#> $Fitted
#> Time Series:
#> Start = 1 
#> End = 80 
#> Frequency = 1 
#>  [1]  0.89508461 -0.12440946  0.42331652  2.03885128 -0.01670028  0.87935958
#>  [7]  0.16604625  1.45435012  0.31672957  1.06513190  0.14908793  1.72866487
#> [13]  0.70322867  0.53409722  0.36782271  0.01185230  0.74304196  0.96753216
#> [19]  1.23375589  0.06253940  0.76643655  0.50822882  1.59807469  1.15351366
#> [25]  0.38763968  0.79440215  1.18036487  0.44081293  0.51545283 -0.06107531
#> [31]  0.58660096  1.35440735 -0.02645950  0.77725465  1.58720746  1.49409804
#> [37]  1.10813776  1.05214459  0.94201316  0.43299758  1.20102512  0.73713117
#> [43]  0.76444279  1.47607790  1.32068867  1.41797764  1.31256245  0.74120509
#> [49]  1.90923175  0.81791782  0.46607029  0.58279215  0.23203162  0.77590971
#> [55]  1.24530975  0.49212405  1.71147292  1.75154853  0.53734242  1.31867798
#> [61]  0.77861790  1.49585548  1.38906817  1.33527603  0.81822114  0.87505206
#> [67]  0.61267629  0.86271629  0.78613030  1.24516180  1.17401277 -0.28222661
#> [73]  1.95485486  1.12962132  0.47028821  0.93762002  0.74844648  0.38953613
#> [79]  1.50804055  0.71249566
#> 
#> $Accuracy_Train
#>                  ME     RMSE      MAE       MPE     MAPE        ACF1 Theil's U
#> Test set -0.9662489 1.295186 1.087428 -107.9173 444.4685 0.007590397  1.079483
#> 
#> $Input_Data
#> $Input_Data$Actual_Train
#> Time Series:
#> Start = 1 
#> End = 80 
#> Frequency = 1 
#>  [1]  0.65952504  0.10661220 -0.66750149  0.36274702 -0.79073513 -1.79370053
#>  [7] -1.47594035  0.19567255 -0.12726827 -0.26861195 -0.22905991  1.93124775
#> [13] -0.41342456 -0.77390259  1.64268542 -1.58987683  0.92924303  0.35594842
#> [19]  0.27144078  0.22346460 -1.04512000  0.77576517  0.02763014  1.15042855
#> [25]  0.25342222 -0.82488682  0.28676319  0.94292281 -0.53107640 -1.29761424
#> [31]  1.22093399  1.47740140 -1.20779824 -0.94193456 -0.58351572  1.13409033
#> [37] -1.75717206  0.05231157 -0.45666876 -0.98754657  0.02126624  0.35400796
#> [43] -1.16106179 -0.47583428  0.11535655 -0.36688686  0.15210927 -0.81243120
#> [49]  0.78293216 -0.44680648 -0.24503039  0.94642071 -0.19796637  1.67688667
#> [55] -0.04331819  0.20424264  0.83721977  1.59664621 -1.34003661  0.10078513
#> [61] -1.81409409  1.23873179  0.65207192  0.77326509  0.42004253 -0.31398419
#> [67] -0.59098124 -0.94971878 -0.31486991 -0.15058437 -1.47426058 -0.55535865
#> [73]  1.83955343 -0.16321110  0.15279233  0.42854347 -2.16824934 -1.59482757
#> [79]  0.05326264 -0.75879443
#> 
#> $Input_Data$Forecasts_Train
#> Time Series:
#> Start = 1 
#> End = 80 
#> Frequency = 1 
#>        Series 1     Series 2    Series 3      Series 4     Series 5
#>  1  0.151472531  1.013408091  0.12088238 -0.2540799447  1.415925606
#>  2  2.122706843  0.494339783  0.19106590  0.8704015498 -1.024651525
#>  3  0.579443812  1.927790118  1.28493468  1.4374770132 -0.017949370
#>  4  0.700750327  0.908519455  2.61245634  0.2410458448  0.843040578
#>  5  0.320435523  1.277478367 -0.37700142  1.7565662622 -0.155023683
#>  6 -0.531963979  0.362952856 -0.15491000  0.3541375504  1.532480225
#>  7  0.797781887  2.808918928 -0.75572848 -0.0350754997  0.778507968
#>  8  2.713334564  1.939060155  2.89781769  0.9712840457  0.467245664
#>  9  1.309577547 -0.385602517  0.69758521 -0.4097690617  0.747243841
#> 10  0.075129886  1.893849557  0.25128882  1.0282804313  1.529752468
#> 11  0.970710520  1.430845693 -0.55368869 -0.3441219972  0.640985802
#> 12  0.885110546  1.031233641  1.00023941  0.7486464757  2.837281304
#> 13  1.118657464 -0.670839387  0.89421588  3.6256685533  0.483101004
#> 14 -0.253492456 -0.951911519 -0.12275891  0.2109346956  1.423808570
#> 15  1.258100081  3.080725168 -0.42797576  0.4475133920 -0.443387802
#> 16  1.218056285  1.210062767  0.24897637  1.2573054199 -0.542972388
#> 17  0.446808013  1.023184942  1.12052624  0.8593242429  0.913927702
#> 18  1.229864210 -0.874587227  1.03982290 -0.3868977646  0.971776054
#> 19  0.552129368  1.678432264  1.08777483  1.7656994621 -0.129392933
#> 20 -0.460672253  0.300630389  1.19854754  0.3367849668  0.249815910
#> 21  1.212417504  1.281414194  1.26232104 -1.1959759257  0.692213769
#> 22  1.995477475  0.095266863 -0.25795301 -1.4400503125  0.362075850
#> 23  1.414672297 -1.544338006  3.01132460  1.0349441277  1.773900348
#> 24  0.430851211 -0.225336598  0.92438315  0.5950828432  1.114979379
#> 25  1.925684094  2.080655681 -1.10454402  1.1411714217  1.463525937
#> 26  2.340730910 -0.463168197  1.47041801  0.8440711514 -0.331105997
#> 27  0.004302507  0.887831310  0.61488530  1.3355255778  1.572895557
#> 28  0.321466803  1.619369237 -0.77035016  0.4622197947  1.119911460
#> 29  0.026281530  1.631459147  0.83854358  0.7889533062  1.131006114
#> 30  1.476279518  1.502040829 -0.20064205  1.6550183101  0.456254190
#> 31  2.037909905  2.014647260  1.24268292  0.6884403011  1.365803879
#> 32 -1.042654413  1.089686492  1.64791279  0.6459110426  2.517514811
#> 33  1.412107058  0.907026678  0.13212091 -0.0588961871 -0.516691371
#> 34  1.163175390 -0.472882341 -0.09613399  0.0737987700  1.607296730
#> 35  0.665722927  2.250331645  1.45791975  2.0728571142  1.401802766
#> 36  3.992526250  1.386094563  2.02898598  0.9671663797  0.383021297
#> 37 -0.393533485  2.239934028  1.71205762  3.1145051894  0.908360338
#> 38 -0.467871941  0.486770731  0.63947275  1.1451636701  1.700486018
#> 39  0.957779358  0.915812770  0.01672906 -0.0907214339  1.753927432
#> 40  0.126245571 -0.099619191  0.38957533  0.3586871739  0.878359082
#> 41  1.310207176  0.445827514  0.72486466  1.5947005355  1.810713363
#> 42  0.603273834  0.118449207  1.18076552  3.4428498598  1.131110184
#> 43 -0.308479659  2.260327059  1.29476417  1.2475594147  1.328237414
#> 44  2.032526436 -0.687740641  1.77807247  2.0673434407  0.835059046
#> 45  0.611108512 -0.007116978  1.46852311  0.5894230508  1.406431103
#> 46  1.099776715  1.466802913  1.58964375  0.0911021050  1.772992972
#> 47  0.592761843  0.546267521  0.09171979  0.5104670835  1.528178306
#> 48  0.604042612  1.615804887  0.21667328  2.7706074627  0.813385084
#> 49  2.739698442  0.535389931  1.49580250  0.5395884386  2.553277176
#> 50  0.536585066  1.773118328  0.02710424  1.2465064009  1.745742277
#> 51  0.582438685  2.228069177  1.46629754  0.0002507064  0.176174085
#> 52 -1.325645178  2.569000174  0.87126797  1.2765163996 -0.008883847
#> 53  0.505135206  1.236422061  0.28373083  0.4484378328  1.215541427
#> 54  0.664636086  0.805445827  0.39624111  2.1234639080  0.358713559
#> 55  3.086589343 -0.129649222  1.20024429  2.1637756624  1.737906965
#> 56 -0.379684310 -0.530536417  0.26523855  1.1803259841  0.800881470
#> 57  1.296460677  1.049470587  1.40226515  0.3474130575  1.712094088
#> 58  1.377309248  1.372637377  0.41014673  1.5503843404  2.532356486
#> 59  0.568257666  2.188422462  0.29781627 -1.7754313359 -0.273104069
#> 60  1.760643633  2.975271353  0.87132264  1.3076926059  0.955366926
#> 61  1.727781691  0.400438597  2.22698840 -0.6108705188  0.141115684
#> 62  2.053069692  1.201069867  1.35451391  2.3264766314  1.235039854
#> 63  0.516789317 -0.147620335  0.84544448  2.1942884387  1.986521627
#> 64 -0.621325569  1.431092717  2.36699882  2.0650927152  1.514561896
#> 65  1.952000711  0.265582842  0.82397053 -0.6124789363  1.399861789
#> 66  0.222929282  1.643590286  0.48064516  1.2371792136  1.169147677
#> 67  0.318271814  1.665469208 -0.08470826  1.0373938290  0.599935063
#> 68  1.790360054  1.085219813  0.91663224  0.6243065282 -0.052477221
#> 69  0.227242872  3.390571763 -0.33158502  3.4713339393  2.099087561
#> 70 -1.537842879  1.912621164  1.53409599  0.9721424288  1.381196645
#> 71  0.138840082 -0.911716874  1.94557562  1.5416670523  1.913229954
#> 72  0.970370010  1.120069457 -0.89282039 -0.3298683035  0.281352829
#> 73  1.555476516  0.193175120  1.88315408  1.7394004833  1.963559790
#> 74  1.094084434  1.826179569  0.54946520  0.3257943890  2.483925346
#> 75  2.073440856  1.292909813  1.25180893  1.7599276984 -0.330591635
#> 76 -0.824516401 -1.141012240  0.77861978  1.5690544557  1.715165807
#> 77  0.780408468  1.338841397  0.14555012  0.8447868820  0.541760522
#> 78  2.255763334  2.373187969  1.08424866  0.7999709883 -1.751203805
#> 79  1.725655635  1.105553385  1.49244422  2.7536625497  0.787755555
#> 80  1.680350577 -0.396527197  1.35951074  1.4877593050  0.975173861
#>         Series 6    Series 7     Series 8    Series 9   Series 10
#>  1  0.3605532883  1.16000563  1.873514072  1.77450066  0.01972010
#>  2  0.4889706994  0.48813723  2.364377026 -1.99837755 -0.48772182
#>  3  1.6980538794 -0.01643577  1.314396547  1.79438333  1.17613907
#>  4  2.4096658214  2.69073992  0.998537136  2.05427431  1.85704103
#>  5  0.4833953940  0.50528281  0.158292487  0.92918755  1.08081510
#>  6  0.6378440700  1.27796691  1.471142795  0.34894456  0.59874574
#>  7  0.7846578810  0.48948412  1.217070444  1.28381845  1.49644467
#>  8  0.1201274321  0.97719072  0.714820634  0.55868038  1.43946666
#>  9  0.8689678677 -0.53275721  0.543927860  2.95156309  1.14408278
#> 10 -0.8173204543  1.43042042  2.459305603  0.11725498  0.02228008
#> 11  1.8429072868  0.36955763 -0.497385873  0.62366080  2.32038645
#> 12  0.2296776375  1.33013879  1.228377157  1.04821772 -0.72477142
#> 13  0.4140122097  0.73384516  0.355667332  1.19234572  1.31288096
#> 14  0.4569673670  0.28990211  0.703011673  0.22370266 -0.26674946
#> 15  0.4564686495  2.05030473  1.795685085  1.83534714  1.39388991
#> 16  0.6294761677  0.34468002  1.733612455  0.20613597 -0.15601862
#> 17  2.1168908993  0.16897319 -0.486459087  0.26644827  0.74566580
#> 18  1.0098584217  0.88742107  0.419621562  1.78628216  0.14587413
#> 19  3.2123073883  2.81407644 -0.378337789  1.01229685  1.10197398
#> 20 -0.8666243408 -1.32264179  1.619948882  0.62791509  1.24500053
#> 21 -1.2810818515  0.32512169  0.931983444 -1.45805176  1.22113762
#> 22 -0.0239834601  1.46324690 -0.431712181  2.31313652  1.18554812
#> 23  1.7660312538 -0.06530407  1.295049092 -0.28826730  2.10749394
#> 24  0.6136927063  1.43369880  1.843869382 -0.18152952  0.70992803
#> 25  0.0552873160  0.82281588  1.507632197  2.07885148  1.63101794
#> 26  0.8773136323  1.26547479  2.790921825  0.31568290  1.59555660
#> 27  1.6773838674  1.36118532  1.706112408  2.54048670  0.73996438
#> 28 -1.0568175367  0.99737017  2.214991705  1.01349936  3.02804261
#> 29  2.4075544711 -0.46734866  1.025955311  0.80445469  1.42126837
#> 30  1.6216405456 -0.45675050  1.324512410  0.73595606  1.64288008
#> 31  2.5511646717 -0.91612056  1.127310021  2.24337142  0.77937121
#> 32  1.6807901989 -0.17083901  1.259764389  0.59709138  0.68720226
#> 33  1.6020473803  0.32093677  2.317789212  3.48695345 -0.10463531
#> 34  1.7121399408  0.82219140  0.901188476 -0.08427562  0.93428789
#> 35  0.5204641194  1.91669325  2.708286369  0.41937032  1.02399795
#> 36  2.2075508069  2.09777651 -1.002800614  0.67019084  1.02212389
#> 37 -0.2216243025  0.68522600  0.073997252  1.55382048  0.37867015
#> 38  0.3267962082  0.80512991  0.324856729  0.62628024  0.54923323
#> 39  2.8156262976  1.06025281  0.816206214  1.94893236  1.40188815
#> 40  0.7717962681  0.01205446  1.011599726 -0.59732569 -0.62438584
#> 41  2.0654497404  1.06094258  0.581196489  1.38322386  0.11639334
#> 42  0.3938646041 -0.13980556  0.050530629  0.55715473  3.42994891
#> 43 -0.0194319330 -0.38106427  1.721929969 -0.20347135 -1.08857016
#> 44  1.0488018032  1.83127019  1.984584178  1.65635207  2.78043421
#> 45 -0.2866644732  1.07615872  0.535754644  1.37354352  1.45864465
#> 46 -0.9105269603  0.86651402  0.010751689  1.56800063  2.55899971
#> 47  1.0871931811  2.36462147 -0.006781854  0.41840996 -0.93521876
#> 48 -0.7261553158  1.21464437 -0.836529589  2.69180102  1.11927490
#> 49  1.0598980517  1.66750874  1.422041456 -0.08995348  1.30446353
#> 50  0.9796372291  0.67402314  0.861853068  1.48342819  2.98081213
#> 51  1.2205164091 -0.27728222  0.850176195 -0.18191786  0.31307831
#> 52 -0.7861140186  0.90046151  1.779203687  0.21163123  0.50118658
#> 53 -0.0002125133 -0.85203698 -0.097631950  0.93807813  0.26356998
#> 54  1.1629329965  1.61020695  1.058288550  2.24276144  2.82220966
#> 55  3.3074041624  0.77662023  0.264497557 -0.01647552  1.71527182
#> 56  1.7701894360  0.40626715  1.451046144  0.79126527 -0.25074095
#> 57  1.6528480255  2.03447045  1.764802413  1.69055947  3.78813836
#> 58  0.2404110430  2.33791423  1.495385702  1.55212294  0.20266457
#> 59  1.1910747222  1.63677361  1.003277224 -0.06742067  0.70006755
#> 60  1.0193474388  2.16739323 -0.916557277  1.11779031  2.51336394
#> 61 -0.0852469417 -0.06978524  2.033981229  1.87635591  0.75918024
#> 62  1.4528089584  1.91693020  0.025624851 -0.12527427  1.14413962
#> 63  0.3082389703  1.33241202  0.421368817  1.16048075  1.98828647
#> 64  1.2323669613  0.06761843 -0.481448261 -0.32205567  0.97702514
#> 65  0.4480564862  0.20309526  2.331488730 -0.24639473  0.67644824
#> 66  1.2708810092  0.97989415 -0.547766374  0.50409628  1.39350893
#> 67  1.8066053428  1.35597223  2.636403207  0.06193129  0.28263538
#> 68  2.0379112035  1.76469494  1.144867024  0.97276441 -0.13180961
#> 69  1.2603076660  0.58108367  1.190522364 -0.17895430  2.55259133
#> 70  0.2485996762  0.80027516 -0.059709094  1.28457014  0.87134687
#> 71  0.9119120172 -0.40770867 -0.844124148  0.53098064 -0.03290005
#> 72  1.0596177340 -0.23331142  0.505592742  1.84321213  1.69881202
#> 73  1.7544568191  2.02078813  1.082733787  1.83989705  0.49067052
#> 74  0.7073674239  0.31861167  1.545246235  1.53410482  0.80463787
#> 75 -0.1484425351  0.49097030  2.331323078  1.53155431  1.35798600
#> 76  1.5934435782  0.28978116  0.062956947  2.35601577  1.52641209
#> 77  0.7558632658  1.59594513  0.654979097  1.18173410 -1.35211999
#> 78 -0.2576330953  1.90421397  0.831424457  2.38343682  1.97195649
#> 79  0.3283783288  2.27866491  0.369966630  1.38549271  1.22983095
#> 80  1.3244936983 -0.23981849  1.314751421  1.22293231  0.78425555
#> 
#> $Input_Data$Actual_Test
#>  [1]  0.19317915  1.03369055  0.22015524  0.74859739  1.64246824  0.21040983
#>  [7] -0.39034542  1.45705626  0.06762969  0.36352669  0.51152515  0.70749592
#> [13]  0.77731586  0.86176382 -0.46124224  1.30342018  1.28738897  1.22215264
#> [19] -0.48624692 -0.57686174
#> 
#> $Input_Data$Forecasts_Test
#>         Series 1     Series 2     Series 3    Series 4    Series 5     Series 6
#>  [1,]  1.3162485  2.724834145  0.016668192  0.45050258  1.55172864  3.546529236
#>  [2,]  0.1695632  0.310168342  1.038760901  2.87433861 -0.06004164  0.780795437
#>  [3,]  2.5436879  0.002946471  1.062661691  0.54502417  0.58505383  2.847830465
#>  [4,]  0.7341872  2.463030290 -0.237983228  1.39765051  0.59009299  0.470709543
#>  [5,]  2.3695464  0.017220731  0.551000134  0.52891580  0.90382670  2.662562580
#>  [6,]  0.8958885  0.072718380  0.091391568  0.26485005  0.85356599 -0.006537504
#>  [7,]  0.9380625  0.084358549 -0.291856586 -0.30671997  1.41725442  1.838305158
#>  [8,]  0.8550175 -0.221701962  1.923831617  0.38217442  1.98702416  1.005604226
#>  [9,]  0.4131958 -0.170347392  1.741200792  0.21495966  1.17600685  3.143180385
#> [10,] -1.8530265  1.285950966  1.152810733  0.99894353  2.02760864  0.249218064
#> [11,]  2.0594499  0.482757466  1.326920857 -0.41625294 -0.47964445  2.128919482
#> [12,]  1.3259487  1.652751303  0.516735948  1.10740376  0.98627107  1.769851636
#> [13,]  0.7922515 -1.226399612  0.726237140  0.03499527  0.90011350  0.534697447
#> [14,]  0.4812397 -0.123596614  1.367651584 -0.44962002  0.40494152 -0.742455464
#> [15,]  0.5705301  1.823502648  0.213812946  1.30988795  0.96013218  2.492284844
#> [16,]  1.2348936  1.741109307 -0.184753630  0.92949151  0.12517027  0.987736643
#> [17,]  2.2198820  1.763053604  2.085267892  1.25390965  2.38075452  0.251214682
#> [18,]  0.8236907  0.278150819  0.771768937  2.17440651  0.84343425  0.576389409
#> [19,]  2.1232110  1.539293364  0.009894599  0.31968079  0.75736235  0.247913105
#> [20,]  2.5875731  0.080710159 -1.125732954  1.25829595  0.83425932  0.295464151
#>         Series 7    Series 8    Series 9   Series 10
#>  [1,]  2.0768465  2.08087935  1.59250264  1.40434545
#>  [2,]  2.4831654 -0.25678452  0.45706602  1.45639006
#>  [3,] -1.0749606 -1.05577294 -0.69226426 -0.71608814
#>  [4,]  0.6644790  0.51717906  1.36770957  1.77108890
#>  [5,]  2.4107721 -1.08817513  1.03505141 -1.74767397
#>  [6,]  1.1150911  1.08584968 -0.37168515  0.09053362
#>  [7,]  1.8149618  1.41527559  0.36755303  0.12885263
#>  [8,]  0.8330834 -1.37910327  1.33061922 -0.61482836
#>  [9,]  0.6204937  1.98089318  1.90976672  2.88352139
#> [10,]  1.0946336  0.08054903  0.03573565 -0.31482633
#> [11,]  1.2400712  1.96807543  0.82928604  1.47672019
#> [12,]  0.3673112  1.30854934  0.83184329  0.70143854
#> [13,]  0.5628535  0.24648624  1.23315601  3.49971481
#> [14,]  1.1681421  1.12045620  0.07809631  0.68412285
#> [15,]  2.0598448  0.85311694  2.34312841 -0.99461349
#> [16,]  0.8608272 -0.99201557  0.95572862  0.64022795
#> [17,]  0.3187403 -0.78336246  2.11672547  1.95532035
#> [18,]  1.6135040  2.39712599 -0.52902076  1.75373353
#> [19,]  0.2843485  1.68961715  0.94671682  0.97109864
#> [20,]  2.2895182  1.03605098  1.33052312  0.32645497
#> 
#> 
#> $Weights
#>  [1] 0.0000000 0.0000000 0.3383084 0.0000000 0.3384775 0.0000000 0.3232141
#>  [8] 0.0000000 0.0000000 0.0000000
#> 
#> $Forecasts_Test
#>  [1] 1.2021302 1.1336928 0.2100926 0.3339904 1.2715284 0.6802445 0.9675924
#>  [8] 1.5926757 1.1876671 1.4301065 0.6873681 0.6273668 0.7322825 0.9773116
#> [15] 1.0630887 0.2580951 1.6143169 1.0680866 0.3516030 0.6415375
#> 
#> $Accuracy_Test
#>                  ME      RMSE       MAE       MPE     MAPE
#> Test set -0.3668849 0.7615414 0.5789012 -76.64626 202.0812
#> 
#> $Top_Predictors
#> [1] 3
#> 
#> $Ranking
#>  [1]  5 10  2  9  3  4  1  8  7  6
#> 
#> attr(,"class")
#> [1] "foreccomb_res"