Inverse Rank Forecast Combination

Computes forecast combination weights according to the inverse rank approach by Aiolfi and Timmermann (2006) and produces forecasts for the test set, if provided.

comb_InvW(x)

Arguments

x

An object of class foreccomb. Contains training set (actual values + matrix of model forecasts) and optionally a test set.

Value

Returns an object of class foreccomb_res with the following components:

Method

Returns the used forecast combination method.

Models

Returns the individual input models that were used for the forecast combinations.

Weights

Returns the combination weights obtained by applying the combination method to the training set.

Fitted

Returns the fitted values of the combination method for the training set.

Accuracy_Train

Returns range of summary measures of the forecast accuracy for the training set.

Forecasts_Test

Returns forecasts produced by the combination method for the test set. Only returned if input included a forecast matrix for the test set.

Accuracy_Test

Returns range of summary measures of the forecast accuracy for the test set. Only returned if input included a forecast matrix and a vector of actual values for the test set.

Input_Data

Returns the data forwarded to the method.

Details

In the inverse rank approach by Aiolfi and Timmermann (2006), the combination weights are inversely proportional to the forecast model's rank, \(Rank_i\):

$$w_i^{InvW} = \frac{Rank_i^{-1}}{\Sigma_{j=1}^N Rank_j^{-1}}$$

The combined forecast is then obtained by:

$$\hat{y}_t = {\mathbf{f}_{t}}'\mathbf{w}^{InvW}$$

This is a robust variant of the Bates/Granger (1969) approach and also ignores correlations across forecast errors.

Author

Christoph E. Weiss and Gernot R. Roetzer

References

Aiolfi, M., amd Timmermann, A. (2006). Persistence in Forecasting Performance and Conditional Combination Strategies. Journal of Econometrics, 135(1), 31--53.

Bates, J. M., and Granger, C. W. (1969). The Combination of Forecasts. Journal of the Operational Research Society, 20(4), 451--468.

See also

Examples

obs <- rnorm(100)
preds <- matrix(rnorm(1000, 1), 100, 10)
train_o<-obs[1:80]
train_p<-preds[1:80,]
test_o<-obs[81:100]
test_p<-preds[81:100,]

data<-foreccomb(train_o, train_p, test_o, test_p)
comb_InvW(data)
#> $Method
#> [1] "Inverse Ranking Approach"
#> 
#> $Models
#>  [1] "Series 1"  "Series 2"  "Series 3"  "Series 4"  "Series 5"  "Series 6" 
#>  [7] "Series 7"  "Series 8"  "Series 9"  "Series 10"
#> 
#> $Fitted
#> Time Series:
#> Start = 1 
#> End = 80 
#> Frequency = 1 
#>  [1]  0.99343962  0.96983490  0.76812776  0.71130862  1.11155877  1.57765109
#>  [7]  0.51218821  1.31172938  0.33677313  1.30439906  2.03097256  0.94234159
#> [13]  1.66729143  0.50879178  0.78131868  0.92200330  1.31725567  0.25685884
#> [19]  0.49440563  0.54693288  1.80585352  1.21839613  0.30707109  0.55045781
#> [25]  0.77264118  0.80031624  1.28441607  0.80625154  1.40450253  1.29743879
#> [31]  1.07028452  0.46804356  0.59832978  0.79914388  0.75598222  1.07427000
#> [37]  2.15135998  1.06037948  0.80894876  1.49638229  0.95819870  0.55044244
#> [43]  0.53003348  1.06919685  1.13688897  1.06219591  1.15374422  1.61050039
#> [49]  1.18189457  1.51040156  1.81102201  0.60643153  1.02377571  1.34183249
#> [55]  0.93956398  0.98445763  0.99475700  0.50760635  1.33470283  0.94677748
#> [61]  1.18156192  1.20948825  0.58341598  1.11131498 -0.05618095  0.90262820
#> [67]  1.14831073  0.64260837  0.61566996  1.92425403  0.52222418  1.71792950
#> [73]  0.66640712  0.42595601  0.86375872  0.77846171  0.41031200  0.82548112
#> [79]  1.07127337  1.29455817
#> 
#> $Accuracy_Train
#>                 ME     RMSE      MAE      MPE     MAPE       ACF1 Theil's U
#> Test set -1.062617 1.420416 1.162461 89.81241 447.5401 -0.1120205  1.810937
#> 
#> $Input_Data
#> $Input_Data$Actual_Train
#> Time Series:
#> Start = 1 
#> End = 80 
#> Frequency = 1 
#>  [1] -2.13759804  1.72461319 -1.05020471  0.94574804 -0.88427382 -0.85392044
#>  [7] -2.42773966  1.26140669 -0.02957782  0.43722274  0.32156837  0.26731342
#> [13]  2.07082472  0.15774378  0.59540337 -0.05556569  0.04588104 -0.03312922
#> [19] -0.67799005 -0.22838943 -0.51933245  1.73475152 -0.17593131 -2.15875282
#> [25]  0.46209169 -0.21810620  0.90796263 -0.68640677  0.61573184  0.45377320
#> [31]  0.68609684  0.20393010 -0.45595163 -1.02873498 -0.85729534 -1.26388251
#> [37]  0.73706568  1.70795250 -0.64038717  0.07550225  1.08394703 -0.26821324
#> [43] -1.47155409 -0.02432686 -1.45519210 -0.66035028  0.61332221 -0.28779293
#> [49]  0.67612650  0.31557737 -0.79656573  0.99246366 -0.32348141  0.04091126
#> [55] -0.08776347 -0.21569343  0.04130070 -0.10196907 -1.93325595 -0.15533278
#> [61]  0.53982503  0.99990720  0.21157038 -1.07002856 -1.70708575  0.57815738
#> [67] -1.25034818  0.35918701 -1.87848635 -0.10621912 -1.46197230  0.30685160
#> [73]  0.75516966  0.36873976  1.41071967 -0.23438756  0.69990402 -0.62602617
#> [79]  0.58154388  0.19158142
#> 
#> $Input_Data$Forecasts_Train
#> Time Series:
#> Start = 1 
#> End = 80 
#> Frequency = 1 
#>       Series 1    Series 2     Series 3     Series 4     Series 5     Series 6
#>  1  2.49127894  0.49325962  0.277558356  1.332861981  1.203039074  0.300251872
#>  2  1.37325983  1.65015859  0.696366088  0.918129280  1.620088651  1.146697503
#>  3  0.79536517  0.03136918 -0.843550281  1.573390924 -0.272034215  0.780255535
#>  4  0.06737753  0.90501119  3.246244866 -0.742743671 -1.591891622  0.372747881
#>  5  0.80744768 -0.26909415  1.298738896  0.045834290  3.160973139  1.771489640
#>  6  3.10845112  2.53473318  0.939036390  2.045328776  2.226588183  1.359574859
#>  7  0.12028118  0.81008640  1.808229686  0.197891991  0.715071423  0.124122955
#>  8  1.71893388  2.01360388  1.798651990  0.430778114  0.803766393  1.201129735
#>  9  0.71207115  0.71899808  0.206874427  1.989692731  1.640660635 -0.001549802
#> 10  0.41992115  0.34811450  1.909287537  1.099975083 -0.444055008  2.619402946
#> 11  1.11893276  1.38749430  3.295965612  0.670313311  1.375479171  2.786140914
#> 12  2.39000282  0.98948989  0.521256540  0.189364637 -0.895897975  0.885474494
#> 13  0.72493936  1.08291449  2.344712356  0.231071431  0.721541331  1.735558902
#> 14  1.52692815  1.10225354  0.609553249 -0.053257086  0.721357117  0.146916900
#> 15  1.09666163 -0.58954277  0.007812091 -1.042004987 -0.581282109  1.470156987
#> 16 -0.42806465 -1.14852763  1.498123022  0.430689432 -0.757255979  1.279831519
#> 17  0.24721878  1.41084297  0.275809275  0.675085386  3.438700946  1.827780850
#> 18  0.49200506 -1.32245015  1.785495660  0.812386024  2.044326580 -0.567570982
#> 19  1.11787454 -0.42338226  0.835032619  0.656471181  1.664492482  0.078290757
#> 20  0.57210876  1.09850974  0.418473412  0.474945793  0.561723247 -0.068240119
#> 21  1.94135429  2.05009750  0.270292995  1.782103615  0.785915522  2.974237605
#> 22  1.64160030  0.52134838  0.161600058  1.687796229  1.746903487  1.105855107
#> 23  0.96562388 -1.02024420  1.805071065  1.399526435 -0.820645725  0.804630996
#> 24 -0.10796823  0.06357190  0.927704583  0.104703846 -0.587552909  0.332908339
#> 25  1.17833800  0.40224792 -0.009598585  0.221521133  1.035804179 -0.074567774
#> 26  1.06099561 -1.05731272  1.910955170  1.139070814  1.618626136  1.873181108
#> 27  0.58273404  0.98367411  1.547489449 -0.091283295  1.705063448  2.327287564
#> 28  2.47265489  1.10674473  1.699667431  1.227059478  2.107739763  0.026838014
#> 29  1.98184769  1.43478993 -0.090201504  3.314351218  1.106006148  1.647331222
#> 30  0.77175352  1.47899744  1.035559560  2.930521891  0.513816356  1.383831404
#> 31  1.60142377  2.13465363  0.373641511  0.312861194 -1.193913574  0.929574617
#> 32  0.91060203  2.19299846  0.077413762 -0.029150887  1.737257982 -0.479018124
#> 33  1.14512139  1.23715036  2.643706800  0.408806558  1.880974214 -0.513957347
#> 34  1.32652601  0.16243247  2.591992434  1.378332327  1.022276535  0.478787341
#> 35  3.19527803  1.59390309  1.154403449  0.050772818  1.716991641  0.724191989
#> 36  1.33032368  1.89735796  2.651349219  0.612361994  1.099638636  0.493131241
#> 37  1.10978828  2.89456597 -0.096438474  2.546190677  0.785549595  4.588948633
#> 38  0.65833936  1.31692984 -0.020862655  2.223607308  0.914852234  0.572476322
#> 39  1.19231077  1.49444837  2.034978975  2.298155860  0.034236937  0.403920311
#> 40  1.53403829  1.38103838  1.879456483  0.073671622  0.825635467  2.510688293
#> 41 -0.12393687  1.81354629 -0.101720234  0.642652565  0.931351091  1.423859729
#> 42  0.35989862  1.73219214  0.168761921  1.824355772  1.006493354  0.539763804
#> 43  0.60005108  2.25973320  0.700746191  1.053208766 -0.863353588  0.211322015
#> 44  0.66042662 -1.15539328  3.281031368  0.821757786  2.046402788  0.816995460
#> 45  1.64949360  1.59533189  1.183438321 -0.005910307 -0.616348208  0.724661114
#> 46  1.84722171  2.46194181 -0.172614629  1.877084004  0.775482638  0.402751620
#> 47 -0.18998326  0.76783138  2.963673449  0.512103540  0.528437892  2.011384840
#> 48  1.01195940  0.13680891  1.858617577  1.997535603  1.586314671  2.000646701
#> 49  1.31743138  0.78465431  1.080521882  0.094505396  3.365544766  2.011555629
#> 50  2.03797266  2.12663512  1.325656621  1.266324384  0.781104226  1.068034580
#> 51  1.77789716  1.67788692  0.874761247  0.454182800  1.787820319  1.892036784
#> 52  0.60469945  0.41640803  1.348945691  0.678178530  3.172384364  0.280185382
#> 53  2.08479767  0.04730205  0.314454109  0.836806497  3.499766014  0.759450675
#> 54  0.61084001  1.24080453  1.280847717  1.432366594  0.561698157  1.776839134
#> 55  1.96307558  0.76871526  0.136307324  0.144835488  1.497269210  1.319547384
#> 56  1.00730336  1.25148968  0.761222527  0.586292095  1.092287468  1.000423942
#> 57  2.28665394 -0.95472497  0.050174211  2.172175678 -0.300713481  1.474931079
#> 58  1.61424476 -0.53723296  1.438885532  1.460489140  2.015314561  0.377981666
#> 59  1.77528088  2.38984615  1.300870072  2.052162804  1.533892733  2.814469571
#> 60  3.05212680  0.77942209  1.545731320  0.618755917  1.456118910  1.390116256
#> 61 -0.16679363 -0.57576005  1.038712112  0.705049146  0.115084742  1.741079428
#> 62  2.04978826  1.34762637  0.592955963  0.984760997 -0.487882347  1.539252186
#> 63 -0.63799538  1.20046945  1.549775591  0.930887731  1.703481620 -0.388737556
#> 64  1.73101267  0.12920874 -0.233849409  1.472532455  0.024703129  1.110187833
#> 65  1.83162903 -0.87731532 -0.031366399 -0.152779942  0.554219557 -0.407760315
#> 66  0.19607538  0.49941501  1.522084328  2.984352268 -0.142072722  0.865075686
#> 67  1.98912293 -0.74422055  2.136347518  2.420188767 -0.575585845  1.011743077
#> 68  0.12402546  1.30777109  0.724766973  3.466906702 -0.783866629  0.175647471
#> 69  0.40297005  0.73887435  1.873151525  0.641942547 -0.327666546  0.310631973
#> 70 -0.20409597  0.98138717  0.910383865  3.320393001  1.124620020  3.146308154
#> 71 -0.24051658  1.26256382  1.891706043  1.675904320  2.418297813 -0.284744411
#> 72  0.79003381  2.41349081  2.857673706  0.906019607 -2.165099767  2.675554658
#> 73  2.10979100  0.25038814  2.052184185  1.237255032  1.958845979  0.311638511
#> 74  0.45844347  0.62269938  0.139386228  0.812917091  1.230925422 -0.004439169
#> 75 -0.41501872  0.26563031  1.884386950  0.504825510  1.074532691  0.504988588
#> 76  2.06265125  1.22636582  1.379165050  0.723205611  1.698509998  0.350255780
#> 77  0.19867441 -0.34005195 -1.998002496  1.272214917  0.001713166  0.898884379
#> 78  1.16878509 -0.55409127  0.621715728  0.987485531  3.582090132  0.058625456
#> 79  1.50292764  0.07534775 -0.180003450  1.471546620 -0.461015969  1.429372043
#> 80  1.70798869  2.98461804 -0.314491904  0.857364744  2.301956562  1.274358705
#>        Series 7    Series 8     Series 9   Series 10
#>  1  0.137650978  2.19917231  1.533666396  1.46058541
#>  2  0.604035723  0.59541232 -0.417267056  1.15464442
#>  3  0.230223695  1.47508635  0.981695648  1.09751993
#>  4  2.197503281  1.58331661  0.009093323  0.68515546
#>  5  1.597478194  1.06221584  1.393103965 -0.15922503
#>  6 -0.529131564  1.91404682  1.886230215  0.52565943
#>  7  1.344062619  0.10070231  0.092519098  2.00513960
#>  8 -0.134088133  1.68206568  0.989578396  1.52756988
#>  9 -0.012602864 -1.14404423  1.215757993  1.91317067
#> 10 -0.511787985  0.77294881  1.947893378 -0.16832287
#> 11  1.176565341  2.34468633  1.475207708  1.11584520
#> 12  2.523183134 -0.10047376  1.473646271  2.66861253
#> 13  1.652992474  2.12210658  1.396760451  2.90424878
#> 14  3.112200567  0.03846301  0.491950190  0.41826490
#> 15  1.911658327  0.90389797  1.147302207  0.67711719
#> 16  0.164376997  1.37976943  0.898892159  3.09850095
#> 17  1.319050909  0.83088281  0.562717468  1.30155287
#> 18  0.721546405  0.86351711  2.062029074  0.78316862
#> 19  2.460136097  1.18059512 -0.416762104  0.59237368
#> 20  0.868076350  0.65890882  1.438020243  1.24697240
#> 21  1.588165926  0.29457218  2.898366344  0.21097008
#> 22  0.573518478  2.30111805  0.215074934  1.41964284
#> 23  0.158390202 -0.08932915  0.235390554 -0.29721632
#> 24 -1.621320799  1.45386182  1.850229628  1.32061182
#> 25  2.380136578  1.32841290  2.035258009  2.08808731
#> 26  1.662944354 -0.32013668 -0.908836813  0.87665600
#> 27 -0.486595490  1.83666603 -0.085578069 -0.55338265
#> 28  3.434803440  0.58475531  1.119269668  0.22533080
#> 29 -1.534299582  2.31981802 -0.773156841  1.32053681
#> 30  1.591860307  0.66770749  2.010600664  0.81978677
#> 31  0.157329192  1.84188643  1.171156255  0.63464062
#> 32  0.221605259  1.48083010  0.004936911  0.26448405
#> 33  0.752251160  1.28198517  2.724716082 -0.61307335
#> 34  0.716472327  1.35825048  0.557821161  0.44540938
#> 35  1.280701059  0.44298008 -0.660646855  0.02777721
#> 36 -1.320044252  1.49796294  1.943421413  1.52229701
#> 37  0.432868901 -0.32040575  0.173929118  0.78745876
#> 38 -0.988630531  2.96962473  1.074059469  0.01338464
#> 39  1.186298424 -0.02234968  1.464940986  0.97091955
#> 40  0.675509029  0.99084532  0.846309273  0.60165234
#> 41  0.275279184  0.69985541  0.036726716  0.83120679
#> 42  1.508613124 -0.34317435  0.806900908 -0.70413810
#> 43  1.124746231  0.73677923 -0.634006506 -0.23138630
#> 44  0.091814483  1.51074970  2.056882636  2.77920789
#> 45 -0.268841115  2.56315047  1.203006690  1.28446239
#> 46  0.178830768  1.25873362  1.471202296  1.45807621
#> 47  0.669784028  0.10027118  0.973711901  1.13538340
#> 48  1.698992020  2.08915271  1.415148162  1.05685809
#> 49  2.256126691 -0.11643343  1.315396463  0.11082536
#> 50  2.847638944  1.48694388  1.574032034  2.00243353
#> 51  2.466086260  2.77062183  2.224462629  0.43689445
#> 52  1.154417851  1.62195253 -1.094784671 -0.26312303
#> 53  2.524175594  1.74415319  0.951148420  0.16036180
#> 54  0.220591146  0.59919453  1.605226947  2.25885519
#> 55 -1.224801020  1.22142016  2.077580478 -0.33644569
#> 56  0.805475693 -0.17162053  1.753513943  2.68751190
#> 57  0.027396426  1.61938232  0.921912372  0.66652994
#> 58  0.437681840  1.09336707 -0.519153891 -0.08674959
#> 59  0.390820049 -1.56740838  1.474657689 -0.49648045
#> 60 -0.233225171  0.34982021  0.142321483  0.68808624
#> 61  0.077338188  1.51953791  1.807375322  2.15593026
#> 62  0.818736344  1.27916939 -0.219413654  1.69686034
#> 63  1.667785160  0.47124176  2.955444057  0.73637424
#> 64  2.620581809  1.62409512 -0.658149080  2.54386548
#> 65 -0.736390781  0.18711895  0.582213224  0.80516372
#> 66  0.375015115  1.55302397  0.083612346  0.28072987
#> 67  0.997890768  2.11246038  0.240974486  2.07540761
#> 68  2.173238219  0.77120344 -0.775628390  0.50344838
#> 69  2.117512458  0.43518875  0.798594568  0.85358842
#> 70  2.291362242  0.89103594  0.834351109  1.86373843
#> 71  0.009749134  0.21232186  0.262111809  2.01042185
#> 72  0.707446908  2.12918383 -0.392383209  0.34018492
#> 73  0.779640449  0.05776720  2.470670786  0.11740439
#> 74  2.141998558  0.22797834  0.850289191  0.50998951
#> 75  2.874537949  0.43038634  1.206195944  2.75313268
#> 76  2.334835184  0.99411774 -0.094750899  0.06969717
#> 77  1.333351708  0.01656031 -0.595006079  1.40787039
#> 78  1.976090397  1.89285426  1.011804785  1.52969404
#> 79  0.753411470  0.59131768  2.127016701  2.04359058
#> 80 -0.769073790  1.06781744  1.898621982  0.78067256
#> 
#> $Input_Data$Actual_Test
#>  [1]  1.0887519 -1.9746244  0.3234991 -2.0051698 -0.3193853 -1.3588273
#>  [7] -0.2954384  0.4017689 -1.4691333 -0.1264267  0.5150820 -1.0682297
#> [13] -2.4926449 -1.3780717  0.7979893 -0.2623196  0.1383699 -0.5418881
#> [19] -1.6089166 -1.5760554
#> 
#> $Input_Data$Forecasts_Test
#>          Series 1    Series 2     Series 3   Series 4     Series 5   Series 6
#>  [1,]  0.40223291 -2.11228049  0.978778940  0.9437737 -0.535061494  1.6277082
#>  [2,]  2.38090785  2.00665598  1.321395039  1.4144500  0.004322665 -0.0223449
#>  [3,]  0.27912357 -0.69556289  2.116027418  0.9227561  0.657331025  1.5545738
#>  [4,]  1.44876115  0.22591697  2.251282734  0.2816391  2.496140816  1.7593597
#>  [5,]  1.64267227  1.75685368 -0.830939801  1.2803973  1.997623788  0.6949590
#>  [6,] -0.54167373 -0.24655710  1.225605966  0.9232736  0.512813021  1.0635093
#>  [7,]  1.06784926  2.05553311  1.710878121 -0.1148038  1.848746167  0.6754830
#>  [8,] -0.07536435  1.15257967  0.153515057  1.4616943  0.810038728  3.0516007
#>  [9,]  1.72028294  2.44112432  0.998222390  1.7150088  1.580743228  0.9685870
#> [10,]  2.00968714  0.59600099  1.087166627  0.5795159  0.260377595  2.2933377
#> [11,]  0.65964395  2.04411488  1.356916397 -0.4475532  0.105070289  1.5862866
#> [12,]  0.84388648  0.46539898  1.138304307  1.7944607  0.893575255  0.8026687
#> [13,]  1.90106972 -0.22845007  1.721366500  0.1926663  0.649762827  0.3373417
#> [14,]  3.08219627 -0.03977475  0.868896488 -0.6542368  1.437634663  1.4525745
#> [15,]  2.42337019 -0.20480899 -0.030469084  1.4764261  0.334598810  0.2400518
#> [16,]  2.32721413  1.15606696  2.442933934  1.0448703  0.878803429  0.8591156
#> [17,]  1.83264230  0.06970640  2.305641448  0.3667672 -1.075978872  1.9591172
#> [18,]  1.95280357  0.58213542  1.843440633 -0.3598199  1.703154606  0.1872044
#> [19,]  0.25945360 -0.35994712  0.007804326  0.7519888  0.946554515  1.8558013
#> [20,]  1.48861916 -0.08568641  2.041419330 -0.1833115  0.585267363  2.0308258
#>          Series 7   Series 8   Series 9   Series 10
#>  [1,]  3.76468015  1.8785158  0.1392307 -0.53915647
#>  [2,]  1.64346048  1.1361932  1.9291994  0.57392657
#>  [3,]  1.02210763  0.5681858  0.2658133  0.80565748
#>  [4,]  0.69348683  0.7828114  1.6034483  0.67447344
#>  [5,] -0.23799287 -0.4239487  2.0242348  1.73000723
#>  [6,]  0.86140930  0.4190237 -0.2390597  0.35879895
#>  [7,]  1.16913602  1.1578557 -1.0699811  1.72682461
#>  [8,]  1.69016547  1.6195823  0.6360400  1.03729562
#>  [9,]  0.14278210  0.8774546  1.8243473  1.62150224
#> [10,]  1.82440637  1.0253490  1.3601899  2.26494014
#> [11,]  0.61472292  0.3442107 -0.2781094  3.53538171
#> [12,]  0.41510521 -0.0717062  0.0326820 -0.41621808
#> [13,]  2.37624944  2.0667443  1.7357293  1.61971018
#> [14,] -0.59296062 -0.4214791  0.6533802  0.04527681
#> [15,] -0.06573278  1.0880660  2.0263361  1.57314285
#> [16,]  1.48760150  0.7213687  0.8502045 -0.74562445
#> [17,]  1.20860434  2.9236130 -1.1482251 -0.08411144
#> [18,]  1.98706551  1.5472469  0.1748883 -0.20957138
#> [19,]  1.36252171  2.5269054  1.4578522  1.53455779
#> [20,]  0.53854095  2.4386844  1.6387288 -0.31505988
#> 
#> 
#> $Weights
#>  prediction_matrix.Series 1  prediction_matrix.Series 2 
#>                  0.03793524                  0.11380572 
#>  prediction_matrix.Series 3  prediction_matrix.Series 4 
#>                  0.04267714                  0.05690286 
#>  prediction_matrix.Series 5  prediction_matrix.Series 6 
#>                  0.03414172                  0.34141715 
#>  prediction_matrix.Series 7  prediction_matrix.Series 8 
#>                  0.04877388                  0.17070858 
#>  prediction_matrix.Series 9 prediction_matrix.Series 10 
#>                  0.06828343                  0.08535429 
#> 
#> $Forecasts_Test
#>  [1] 0.8755883 0.9029228 0.8612069 1.2131871 0.8070302 0.5646892 0.9636732
#>  [8] 1.7784434 1.2876528 1.5655342 1.2067826 0.5149317 0.9933838 0.6048770
#> [15] 0.6998926 0.8969972 1.3013136 0.6756912 1.4062693 1.3643287
#> 
#> $Accuracy_Test
#>                 ME     RMSE      MAE      MPE     MAPE
#> Test set -1.684803 1.985947 1.715929 138.6704 287.0323
#> 
#> attr(,"class")
#> [1] "foreccomb_res"