Least Absolute Deviation Forecast Combination

Computes forecast combination weights using least absolute deviation (LAD) regression.

comb_LAD(x)

Arguments

x

An object of class 'foreccomb'. Contains training set (actual values + matrix of model forecasts) and optionally a test set.

Value

Returns an object of class foreccomb_res with the following components:

Method

Returns the best-fit forecast combination method.

Models

Returns the individual input models that were used for the forecast combinations.

Weights

Returns the combination weights obtained by applying the combination method to the training set.

Intercept

Returns the intercept of the linear regression.

Fitted

Returns the fitted values of the combination method for the training set.

Accuracy_Train

Returns range of summary measures of the forecast accuracy for the training set.

Forecasts_Test

Returns forecasts produced by the combination method for the test set. Only returned if input included a forecast matrix for the test set.

Accuracy_Test

Returns range of summary measures of the forecast accuracy for the test set. Only returned if input included a forecast matrix and a vector of actual values for the test set.

Input_Data

Returns the data forwarded to the method.

Details

The function integrates the least absolute deviation (LAD) forecast combination implementation of the ForecastCombinations package into ForecastComb.

The defining property of comb_LAD is that it does not minimize the squared error loss like comb_OLS and comb_CLS, but the absolute values of the errors. This makes the method more robust to outliers -- comb_LAD tends to penalize models, which have high errors for some observations, less harshly than the least squares methods would.

Optimal forecast combinations under general loss functions are discussed by Elliott and Timmermann (2004). The LAD method is described in more detail, and used in an empirical context, by Nowotarksi et al. (2014).

The results are stored in an object of class 'foreccomb_res', for which separate plot and summary functions are provided.

References

Elliott, G., and Timmermann, A. (2004). Optimal Forecast Combinations Under General Loss Functions and Forecast Error Distributions. Journal of Econometrics, 122(1), 47--79.

Nowotarski, J., Raviv, E., Tr\"uck, S., and Weron, R. (2014). An Empirical Comparison of Alternative Schemes for Combining Electricity Spot Price Forecasts. Energy Economics, 46, 395--412.

See also

Examples

obs <- rnorm(100)
preds <- matrix(rnorm(1000, 1), 100, 10)
train_o<-obs[1:80]
train_p<-preds[1:80,]
test_o<-obs[81:100]
test_p<-preds[81:100,]

data<-foreccomb(train_o, train_p, test_o, test_p)
comb_LAD(data)
#> $Method
#> [1] "Robust Regression (QR)"
#> 
#> $Models
#>  [1] "Series 1"  "Series 2"  "Series 3"  "Series 4"  "Series 5"  "Series 6" 
#>  [7] "Series 7"  "Series 8"  "Series 9"  "Series 10"
#> 
#> $Fitted
#> Time Series:
#> Start = 1 
#> End = 80 
#> Frequency = 1 
#>  [1] -0.468733961 -0.616415367  0.232877455 -0.153446807 -0.548195419
#>  [6]  0.924500159 -0.410351918  0.448365183 -0.172185682 -0.636313531
#> [11]  0.095881429 -0.231247550 -0.140987255 -0.009064255  0.114757255
#> [16] -0.209333637  0.182983666  0.538603744 -0.015404449  0.316972958
#> [21] -0.226469352 -0.091690618  0.030758570  0.284330381 -0.775381234
#> [26]  0.114297373 -0.017775027  0.524327598 -0.679144134 -0.214455145
#> [31]  0.023243352 -0.377254885  0.003490054 -0.017155977  0.370622922
#> [36]  0.968513633  0.515944703  0.271781229  0.386835458 -0.108864438
#> [41] -0.098062324 -0.197743510 -0.308861942 -0.325495994  0.279370723
#> [46]  1.280549947  0.097490193  0.466694765 -0.238008776 -0.069272963
#> [51] -0.300510120 -0.055829948  0.134806215  0.518853049  0.571698079
#> [56]  0.455615993  0.188522419  0.246202151  0.156401594  0.168207088
#> [61]  0.086387314  0.052326228 -0.097371179 -0.320455870  0.652316942
#> [66] -0.079090789 -0.431766166 -0.233436579 -0.299244026 -0.263524471
#> [71] -0.481520958  0.031722446 -0.567318918  0.178710291 -0.077867662
#> [76] -1.030950304  0.376755983 -0.557627029  0.225968274 -0.337779684
#> 
#> $Accuracy_Train
#>                   ME      RMSE       MAE      MPE     MAPE        ACF1
#> Test set -0.01701429 0.9539734 0.7297963 66.77241 101.0119 -0.08648911
#>          Theil's U
#> Test set  1.018505
#> 
#> $Input_Data
#> $Input_Data$Actual_Train
#> Time Series:
#> Start = 1 
#> End = 80 
#> Frequency = 1 
#>  [1] -0.08611513 -0.61641537  1.88844287 -0.36184976 -1.56225967  1.62756717
#>  [7] -0.38573605 -0.79321639 -0.17218568 -0.63631353 -0.71966146 -1.95746671
#> [13] -0.98944081  0.31868265  1.07884482 -0.98768846 -1.33070315  0.53860374
#> [19]  1.77141996  0.74981860 -0.03637373  0.88834712  0.03075857 -0.16175920
#> [25] -1.51717007  0.66886587 -0.82600274  0.84831515  0.93255380 -0.72453926
#> [31] -2.21431400 -0.64021268  0.42105012 -0.89965744  0.43176265 -0.70749809
#> [37]  1.63934965  1.60831456  1.11961919 -0.65595537 -0.73067993 -0.93895225
#> [43] -0.67057097 -0.32549599 -0.80091681 -1.28443097  1.32952232 -0.69532281
#> [49]  0.87244505 -0.06927296  1.00363665 -0.05582995  1.66005320  0.60971088
#> [55]  1.01637909 -1.52281720  1.13157777  0.18024826  0.89356193  0.16820709
#> [61] -0.06119131  0.34623533 -0.75947674 -1.03179441  1.50997300 -0.13859844
#> [67] -0.13672044  0.88035533 -1.34087325 -0.26352447 -0.48152096 -0.28346876
#> [73]  0.21242227 -0.20129236 -1.56484478  1.24596757  0.15448718  0.34071551
#> [79]  0.26951344  0.61773616
#> 
#> $Input_Data$Forecasts_Train
#> Time Series:
#> Start = 1 
#> End = 80 
#> Frequency = 1 
#>       Series 1    Series 2    Series 3    Series 4    Series 5    Series 6
#>  1  1.56281158  0.73182236  0.46188440  0.83698161  0.62329549  0.95323495
#>  2  1.27456700  0.79131354  1.96066858  2.63204827  2.30118492  2.47300955
#>  3 -1.03420972  1.16617454  1.98870322  1.15800371 -0.94767347  1.26439439
#>  4  1.11799795 -0.00887460  1.23569886  0.96647313  1.90342551  0.66576661
#>  5  0.09024025  0.65701128 -0.44566805  1.08197536  2.05647187  1.73487982
#>  6  2.72192036  0.61603113  0.68454141  2.06197512  0.87350445 -0.54760267
#>  7  2.38209142  0.76956800  0.01123831 -1.72527788  1.04158805  0.17140563
#>  8  0.24062434  0.80090826  0.71782973  0.88304178  1.07013215  0.13488514
#>  9  1.56129354 -0.73386243  1.88839550  1.79560748 -0.13690667  1.38984162
#> 10  1.88525498  2.08290292  1.10458887  0.20510736  1.60707437  0.62618845
#> 11  0.68245348  2.04785755  3.37162544  1.12524893  1.55669536  0.58522286
#> 12  1.38495070  2.38593140  2.34403264  0.65439634 -0.39947994 -0.07042434
#> 13  0.02709285 -0.64822635  0.31615147  1.48966148 -0.19885950 -0.39785550
#> 14 -0.36985797  3.70704774  0.22400241  1.03688103  1.19229618  0.67962286
#> 15  1.89685014  1.40284210  0.18483681  1.94398055  1.39716568 -0.03600468
#> 16  0.37502012 -0.04922946  0.44008407  0.46253796  1.27372655  1.54825438
#> 17  0.22025485  0.79917530  1.34401836  0.79892512  0.89464840 -0.93373427
#> 18  2.27963129  0.28960753  1.95173853  1.03052902 -0.23688597  0.33852509
#> 19 -0.42972379  0.23904067  1.96733650  0.95212988  0.42533514 -0.21979080
#> 20  1.39704764 -0.08154382  2.43630145  2.59039312  1.16169779  0.54728027
#> 21  0.28234507  0.49472461  0.21307812  2.10331728 -0.26324406  2.44715404
#> 22 -1.71009788  1.79843067 -0.63259353 -0.69616120 -0.30552640 -0.88816581
#> 23  3.79584426  0.67568210  0.32235710 -0.47855323  0.96160048  2.72491224
#> 24  1.37940587  1.71895266  1.16277241  1.01997210  0.19783811  2.28551360
#> 25  0.56187259  1.73575435  0.46890533  2.00179059  1.49125096  2.69031627
#> 26  0.93723050  0.85155328  0.94115377 -0.20460638  0.56797162 -1.30540985
#> 27  0.96509078  0.66173564  1.26263296  1.63281310  0.80157926  0.84982522
#> 28  0.78312188 -0.14687221  1.57937379  0.83538928 -1.29743243  1.24387719
#> 29  0.77085766  1.09940033  0.63702744 -0.85149913  2.90465822  1.60626152
#> 30 -0.66090998 -0.42492220 -0.67717913  0.47125570  1.05672787  0.42530088
#> 31  0.22386698  0.26120517  0.72056142  2.01909699  1.31446932  1.40642845
#> 32  0.57933475  1.11351031  0.02323834  1.06931626  3.30034329 -0.15950199
#> 33  1.20002919 -0.89219794  1.00422170  0.04314729 -0.08551014  1.73450380
#> 34  0.84309368  1.44432241  0.70588971  0.56348010  0.49829711 -0.62115351
#> 35  0.67303524  1.77749485  1.99415527 -1.06991393  2.78259938 -1.44575531
#> 36  1.44852873 -0.71450237  1.10934846  2.14563254 -0.73456959 -1.32080701
#> 37  1.09251393  1.74966574  0.55619877  2.16864147  2.41517446 -0.39334211
#> 38 -1.40971504  3.26652038  0.23105340  2.00163270 -0.36922880  2.02659540
#> 39  0.91372625 -0.27219576  1.20965589  1.11491376 -0.38834091  1.41006682
#> 40  0.78252556  1.68230458  0.70162784  2.28670524 -0.05518049  2.42559711
#> 41  1.34108611  0.95113886  0.30529570 -0.16950957  0.47335663  1.53024460
#> 42 -0.77696561  1.96800589  0.15766365 -0.31716424  1.13931115  2.89031594
#> 43  0.70726471  1.56060091  0.89414621 -1.46606245  0.72734302  0.66239903
#> 44  2.19343029  0.86418414  3.94659119 -1.11341038  2.66239854  0.41498937
#> 45  0.41440007  1.34579177 -0.87532529  0.64818259 -0.53237291  1.06945767
#> 46  1.81965054  0.76166567  1.09126123  2.41070976  0.84069735  1.03961805
#> 47  0.25882414  0.73974273  1.00359510  1.94888422  0.29705840  1.03967619
#> 48  0.61663474  2.72848247  0.19711099  1.96406359 -0.19811120 -0.22557922
#> 49  1.45626867  0.50824623 -0.21310395  2.03311345  0.20434698  2.29612279
#> 50  1.58305038 -0.61206667  2.26788054  0.20953580  2.59826720  0.81126455
#> 51  1.57498692  1.42059335  0.50209256  2.11943953  0.58960438  0.74901861
#> 52  0.07817499  0.72788956  2.38288737  0.06637947  0.35426104  1.72768189
#> 53  0.82801764  2.38681782  2.10391874  0.32901133  1.39136836  1.03494924
#> 54 -0.51841431  1.69496293  1.86417910  0.90231360  1.87655323  0.79476788
#> 55  1.18191562  0.92072870  3.07086370  1.05715397  1.37121916 -1.36171573
#> 56  2.30438186  1.01119744  0.06146297  1.65239552  0.60658635  0.06516541
#> 57  0.36379676  1.82709130  1.90199705  2.07600169  1.17725246  1.98785887
#> 58  0.83308389  1.94353578  0.81431178  1.68632378  1.79750273 -0.82324844
#> 59  1.91642832 -0.25337814  2.01438071  1.60635526  1.91380186  2.01169716
#> 60  0.60698421  1.67000705  1.39961615  2.83140469  1.21672778  0.03074837
#> 61  3.00924964  0.51936025  0.22316165  0.11811015  1.08703272 -2.30516982
#> 62  0.43239668  0.49327160  0.38659789  0.41335013  1.42551580 -1.32985245
#> 63 -0.37282554  1.77760081  0.81567337  2.40514416  0.76583660  0.15492002
#> 64  2.82894551  0.78967351  1.52687119  0.03375329  1.94574673  0.59377734
#> 65  1.57536371  1.10147841  1.27973791  1.99456780 -0.22236269  0.08895126
#> 66  0.17328067  1.02585225  1.95503138  1.03270051  1.85474483  0.74158285
#> 67  0.16571861  0.96977911  0.53966957  1.29911149  2.82426427  0.72121439
#> 68  0.20844611  1.80677001  1.87356350  0.06475720  2.23902339 -0.03033375
#> 69  0.63518375  1.75545722 -0.68353452  0.28131538  1.02909032  1.15659843
#> 70  0.83815515 -0.39897153  3.52900054  0.23094361  1.17800027  1.16526851
#> 71  1.08970939  2.16106424 -0.48190159 -0.37064831  1.75245982  2.43290940
#> 72  1.16168289  0.05900275  0.97125072  0.69234508  0.21963200  1.12422258
#> 73  0.90970915 -0.79040486  0.79550606 -0.99853960  1.35469011  1.38045952
#> 74 -0.30556079  0.23366893  0.83164196  1.87373740  0.76808220  0.50145314
#> 75  0.61130020  1.89108079  0.08490891  0.77775655  0.99082184  0.04514855
#> 76  0.71167761  0.16538187 -1.08547940  0.20331381  1.41364451  0.55537217
#> 77  1.04555711  0.28171510  3.74421836 -0.23156451  0.05543789  2.12209919
#> 78  1.10149060 -0.15692793  0.53331661  1.43933331  1.86978509  1.49626696
#> 79  2.28393128  1.75744116  0.41745444  1.34868805  1.12869204 -0.86724504
#> 80 -0.56380070  2.51723972 -0.15857082 -0.05197592  1.10982440  0.17692519
#>      Series 7    Series 8    Series 9   Series 10
#>  1 -0.6279948  2.48525972  1.84081634  0.25697170
#>  2  2.6072511  1.65941719  2.51741054 -0.39742420
#>  3  1.6257826 -0.75210914  2.10868563  0.92320932
#>  4  1.3119582  0.81781709  1.00260958  1.04696695
#>  5  0.6411865  1.32694259  0.77433252 -0.50216224
#>  6  0.3040929  1.34610280 -1.17521078  2.79706177
#>  7  0.1203975  2.62644997  0.94895260 -0.34472591
#>  8  3.6902916  1.27571664 -0.43635831  1.70428990
#>  9  0.3139186  1.72530613  2.55479440 -0.83958161
#> 10  0.7971817  2.79574104  2.13885803  0.12185492
#> 11  1.3416559  0.69892204  1.12214211  1.98581559
#> 12 -0.7991107  1.39770087  3.19803557  1.90233913
#> 13  2.3096546  1.55442884  2.63423649 -0.15724838
#> 14 -0.4153479 -0.08708555  0.54962590  1.75813326
#> 15 -0.3533053  1.62570225  0.58042806  0.46448308
#> 16 -0.3924531 -0.40201949  0.84015787  0.63997599
#> 17  2.0810428 -0.84071876  2.03313603  1.14418780
#> 18  1.5568221  1.10342344  1.03614347  0.97052338
#> 19  1.5519798  1.12763769  1.59199278  2.15491526
#> 20  1.6139520  3.07938876 -0.06971383  2.41823311
#> 21  1.1352759  1.68881886  1.62268652  1.03672538
#> 22  0.6723455  1.03311813  0.85875410  2.10492942
#> 23  2.9276951  1.74537760  0.14339303  1.89275928
#> 24  0.5397171  1.40484822 -0.32920339  2.91283119
#> 25  1.7731924  2.98011127  1.82401379  0.87443517
#> 26  0.5780310  1.08001385  1.25423889  0.75051928
#> 27 -0.2165980  1.61886169  1.03374915 -0.40964627
#> 28  1.4297325  0.10154727  1.13413014  1.65969092
#> 29  0.7131411  0.95537868  0.38559259  2.08631992
#> 30  1.7245660  1.37185551  0.40437570  0.87159737
#> 31  1.8722862  2.25818706 -0.08912718  0.87222307
#> 32 -0.1405944  1.93236303 -0.09314192 -0.20766301
#> 33  2.2497634  1.08198045  1.19662648  0.52467134
#> 34  0.7541029  1.67430371  1.40920066  1.34143510
#> 35  0.7471270  0.27407413 -1.06650036  1.11798090
#> 36  0.9717516  1.23708497  0.33023082  0.65042684
#> 37  1.1991478  0.90422985 -0.84924878 -0.05244234
#> 38  2.4646999  2.28420039 -0.32036864  1.29831846
#> 39  0.7094382  0.97728402  0.45464885 -0.21926849
#> 40  1.5561424  2.36959820  1.33401078  0.80773794
#> 41  0.4809166  0.81820605  0.70797328  2.08445182
#> 42  2.4835638  0.80199456 -0.23302807 -0.34210126
#> 43  1.6073281  0.27078746  1.71397372  1.40456453
#> 44  0.8079554  0.08111335  1.74942941  1.64257811
#> 45  1.8416213 -1.22688654  1.40857159  1.37955901
#> 46  0.7814138  0.17357396 -2.40416260  1.17691919
#> 47  0.9912070  0.88510413  1.40351882 -0.91900127
#> 48  0.8134160  0.54435581  1.29731939 -1.16304790
#> 49  0.5969497  0.96103980  1.87551406  0.91202103
#> 50 -0.7390038  0.15421700  0.12792573  0.39713888
#> 51  1.1809865  2.01590664  2.47204853  1.26158457
#> 52  1.5663734  1.85923311  0.62886342  1.07881702
#> 53  1.9407667  1.54151534  0.21083130 -0.03982302
#> 54  1.1292399  0.61701909 -1.65093584  0.74600078
#> 55  0.4650325  2.12752474 -0.36111468  1.37597148
#> 56  1.9184892  1.20986738  0.52481856  1.30361961
#> 57  1.5191376  0.18381254  0.51247626  2.16371276
#> 58  0.8336587 -0.90963267  1.32123562  0.96854727
#> 59  0.8425015 -0.75400668  0.65930061  1.10169841
#> 60 -0.2491669  0.55527578  1.13424014  2.28064628
#> 61  1.6092847  0.62067370  2.44986953  1.18872118
#> 62  0.6278406  1.64271524  0.48544843 -0.15900565
#> 63  0.2291047  2.23726053  1.19955427  0.18738780
#> 64  0.1785142  0.94890214  1.82460690 -0.21414718
#> 65  0.3170585 -0.94867961  1.39673222  1.52700495
#> 66  0.3440988  2.51900106 -0.47596738  2.81561631
#> 67  1.2333962  2.42970687  0.03608326  1.62037293
#> 68  2.1103819 -0.12430149  1.79353278 -0.71134135
#> 69  1.0896075  1.81347539  0.67717286  0.30056407
#> 70 -0.3787541  1.99080896  0.95735763  2.27431804
#> 71 -0.2043554  2.86741481 -0.58048165 -0.39087307
#> 72  1.5034211  0.92292281  1.23039982  1.93844571
#> 73  0.4560399 -0.26215313  1.78668506  1.82242531
#> 74  1.5052364  1.80867299  0.10890013  1.19882276
#> 75 -0.9104031  1.18447979  0.68687286  0.86616519
#> 76  0.6987767  3.28343562  2.05111182  2.00704767
#> 77  0.8948610  0.30044250  0.36579202  0.85211116
#> 78  2.5617598  0.66210557  2.45681074  0.40195600
#> 79  0.4870198 -0.39178185  1.92583063  0.67618459
#> 80  1.6265639  0.63750626  1.48505236  0.88177639
#> 
#> $Input_Data$Actual_Test
#>  [1]  0.5295705  1.3225437  0.6178421 -0.2250949  0.2511765 -2.5176486
#>  [7]  1.0211342  0.1353909 -1.4685769  0.2996589  0.7877851  0.4098977
#> [13] -0.2091227 -0.2894450  1.4244126 -0.7715258  0.3271078  0.3761074
#> [19] -2.2121270  1.1599430
#> 
#> $Input_Data$Forecasts_Test
#>          Series 1    Series 2    Series 3   Series 4    Series 5     Series 6
#>  [1,]  1.38215234  1.95297019 -0.33504785  0.6629697  1.79403156  1.478724886
#>  [2,] -0.05770553  1.32352231  1.66446373  1.4134997 -0.69288506  1.513093852
#>  [3,]  2.05404413 -1.05403025  0.79907114 -0.2061284  0.21740960  0.489002945
#>  [4,]  3.46331344  1.50583273  0.25983494 -1.3198933  0.08191633  0.775120410
#>  [5,]  0.79904844 -0.08736138  1.29356207  1.9819858  0.71185491 -0.502068661
#>  [6,] -2.36569141  2.42621519  0.87192490  1.8559680  0.51408355  1.212057204
#>  [7,] -0.07902988  0.96171271  2.41937105  1.9563580  0.16865074  1.375230407
#>  [8,] -0.70727486  1.25134365  0.38761342 -0.6085166  0.04669628  0.677026788
#>  [9,] -0.65501863  0.76861512  2.00329550  1.7699367  0.76758511  1.674054706
#> [10,] -0.08747957  1.67121182  0.42247449  1.6973308  1.52441753  1.972112480
#> [11,]  0.35282500  0.11459908  2.41720296  0.7770907 -0.45061787  1.335593390
#> [12,]  0.35390512 -0.14053154  0.52970859  2.9461498  2.18804255 -0.387720626
#> [13,]  0.15747642  1.10106911  0.28988134  1.1979621  0.65876850  2.837075732
#> [14,] -1.45162696  0.86894950  0.82239757  1.1139136  2.28036751 -0.717780020
#> [15,]  2.25104872  0.78985177  2.62787543  0.4880357  1.24135957  1.359416068
#> [16,] -0.72493792  1.18565297 -0.04450503  4.4072346  1.94146102 -1.051314890
#> [17,] -0.69376940  0.92588891  1.22142104  1.6170160  0.91249624  1.075107725
#> [18,]  0.21462890 -0.29758621  2.84647450  1.6126030  0.84435097  0.465484360
#> [19,]  0.84977972  3.46620478  2.71944806  0.6476776  1.71456090  1.894554551
#> [20,]  0.67098197  0.50206722  0.54652822  1.6444737  3.18682622 -0.009640328
#>          Series 7   Series 8   Series 9   Series 10
#>  [1,]  1.76839436 0.86335083  0.1975685 -0.29074418
#>  [2,] -0.08348455 2.68056643  0.8085635  0.79504633
#>  [3,]  1.14346479 1.23227466  3.7198205  0.35535610
#>  [4,]  2.19325689 1.41383654  0.8961256  0.77342446
#>  [5,] -0.35969763 0.90852566  0.1753476  2.17077645
#>  [6,]  0.26647581 2.43769131  0.9013464 -0.17057492
#>  [7,]  1.57835745 1.43134340  1.4493863  0.19333289
#>  [8,]  2.04067223 1.32345623  0.0200958  1.71449650
#>  [9,]  1.82474693 1.37108259 -0.9501928  0.02420584
#> [10,] -0.35937190 1.51008694  0.8042357  1.22690279
#> [11,]  1.76090816 2.33463965  1.3186634  1.75742272
#> [12,]  1.82854644 1.39045359 -0.5248087  3.26098991
#> [13,]  0.50710756 0.60659401  0.6725571  0.55149998
#> [14,] -1.96255930 0.41650197  0.5557045  2.74159944
#> [15,]  0.23620030 0.47968675 -0.3425439  3.03689004
#> [16,]  1.35127202 0.03617077  0.9208550  0.76704225
#> [17,] -0.20058272 0.76803849  0.1659952  1.68076650
#> [18,] -0.08359808 1.39509911  0.4105757  1.70711473
#> [19,]  1.69812792 0.84772351  0.2553349  1.49861161
#> [20,] -0.15649669 0.68119018  0.9732591  2.64299095
#> 
#> 
#> $Predict
#> function (x, newpreds) 
#> {
#>     coef <- c(x$Intercept, x$Weights)
#>     pred <- as.vector(coef %*% t(cbind(1, newpreds)))
#>     return(pred)
#> }
#> <bytecode: 0x56049bea3720>
#> <environment: namespace:ForecastComb>
#> 
#> $Intercept
#> [1] 0.3799036
#> 
#> $Weights
#>  [1]  0.10229060  0.02404552  0.08357150  0.11238880 -0.22739494 -0.13167561
#>  [7]  0.06030042 -0.13405121 -0.27891958 -0.02743919
#> 
#> $Forecasts_Test
#>  [1] -0.04413706  0.05040298 -0.64906421  0.19461701  0.44268293 -0.35621465
#>  [7]  0.09130823  0.09487480  0.49334634 -0.45119730  0.01189755  0.32238259
#> [13] -0.19548114 -0.38219922  0.40449434  0.32155948  0.05847908  0.20711593
#> [19]  0.08730661 -0.47702667
#> 
#> $Accuracy_Test
#>                  ME     RMSE       MAE      MPE     MAPE
#> Test set 0.03719409 1.109901 0.8512677 89.52582 100.3548
#> 
#> attr(,"class")
#> [1] "foreccomb_res"