Newbold/Granger (1974) Forecast Combination

Computes forecast combination weights according to the approach by Newbold and Granger (1974) and produces forecasts for the test set, if provided.

comb_NG(x)

Arguments

x

An object of class foreccomb. Contains training set (actual values + matrix of model forecasts) and optionally a test set.

Value

Returns an object of class foreccomb_res with the following components:

Method

Returns the used forecast combination method.

Models

Returns the individual input models that were used for the forecast combinations.

Weights

Returns the combination weights obtained by applying the combination method to the training set.

Fitted

Returns the fitted values of the combination method for the training set.

Accuracy_Train

Returns range of summary measures of the forecast accuracy for the training set.

Forecasts_Test

Returns forecasts produced by the combination method for the test set. Only returned if input included a forecast matrix for the test set.

Accuracy_Test

Returns range of summary measures of the forecast accuracy for the test set. Only returned if input included a forecast matrix and a vector of actual values for the test set.

Input_Data

Returns the data forwarded to the method.

Details

Building on early research by Bates and Granger (1969), the methodology of Newbold and Granger (1974) also extracts the combination weights from the estimated mean squared prediction error matrix.

Suppose \(y_t\) is the variable of interest, there are \(N\) not perfectly collinear predictors, \(\mathbf{f}_t = (f_{1t}, \ldots, f_{Nt})'\), \(\Sigma\) is the (positive definite) mean squared prediction error matrix of \(\mathbf{f}_t\) and \(\mathbf{e}\) is an \(N \times 1\) vector of \((1, \ldots, 1)'\).

Their approach is a constrained minimization of the mean squared prediction error using the normalization condition \(\mathbf{e}'\mathbf{w} = 1\). This yields the following combination weights:

$$\mathbf{w}^{NG} = \frac{\Sigma^{-1}\mathbf{e}}{\mathbf{e}'\Sigma^{-1}\mathbf{e}}$$

The combined forecast is then obtained by:

$$\hat{y}_t = {\mathbf{f}_{t}}'\mathbf{w}^{NG}$$

While the method dates back to Newbold and Granger (1974), the variant of the method used here does not impose the prior restriction that \(\Sigma\) is diagonal. This approach, called VC in Hsiao and Wan (2014), is a generalization of the original method.

Author

Christoph E. Weiss and Gernot R. Roetzer

References

Bates, J. M., and Granger, C. W. (1969). The Combination of Forecasts. Journal of the Operational Research Society, 20(4), 451--468.

Hsiao, C., and Wan, S. K. (2014). Is There An Optimal Forecast Combination? Journal of Econometrics, 178(2), 294--309.

Newbold, P., and Granger, C. W. J. (1974). Experience with Forecasting Univariate Time Series and the Combination of Forecasts. Journal of the Royal Statistical Society, Series A, 137(2), 131--165.

See also

Examples

obs <- rnorm(100)
preds <- matrix(rnorm(1000, 1), 100, 10)
train_o<-obs[1:80]
train_p<-preds[1:80,]
test_o<-obs[81:100]
test_p<-preds[81:100,]

data<-foreccomb(train_o, train_p, test_o, test_p)
comb_NG(data)
#> $Method
#> [1] "Newbold/Granger (1974)"
#> 
#> $Models
#>  [1] "Series 1"  "Series 2"  "Series 3"  "Series 4"  "Series 5"  "Series 6" 
#>  [7] "Series 7"  "Series 8"  "Series 9"  "Series 10"
#> 
#> $Fitted
#> Time Series:
#> Start = 1 
#> End = 80 
#> Frequency = 1 
#>  [1]  0.75109389  1.24572919 -0.14729307  0.89910134  0.34801216  0.87357478
#>  [7]  0.06446331  1.92349080  1.40500656  0.73239581  1.17315005  1.75237253
#> [13]  1.28055451  0.17000126  1.11357165  1.44191906  0.23126396  1.39669354
#> [19]  0.69223826  0.47035228  0.91028562  1.56812827  0.86105858  0.50323873
#> [25]  0.16483505  0.99788693  0.17090459  0.83864053  0.78593412  0.43335389
#> [31] -0.03854830  0.71122339  1.29325711  0.32017525  1.32769425  0.31672824
#> [37]  0.71969581 -0.35677524  0.86171389  1.66647929  0.72864121  1.06775388
#> [43]  0.23943689  0.04492819  1.16824120  1.13826754  1.09042172  0.67118537
#> [49]  1.04366295  1.17174501  1.17457010  1.50225972  0.52203376  0.04802192
#> [55]  0.51862042  0.54213784  1.67918124  1.32628627  0.76029131  0.66833214
#> [61]  0.96153720  0.16785859  1.27693181  1.20798196  1.65430372  0.87610498
#> [67]  1.19963399  0.65020812  0.85630632  0.90328164  0.94931986  1.24416348
#> [73]  0.45991857 -0.13623868  0.93071813  1.22851435  1.45738201  1.44768447
#> [79]  0.88564066  0.91942511
#> 
#> $Accuracy_Train
#>                  ME     RMSE      MAE      MPE     MAPE       ACF1 Theil's U
#> Test set -0.9970758 1.556945 1.231589 76.44074 245.7406 0.05862683  1.224421
#> 
#> $Input_Data
#> $Input_Data$Actual_Train
#> Time Series:
#> Start = 1 
#> End = 80 
#> Frequency = 1 
#>  [1] -0.03530875 -2.57962689 -1.62720896  1.25676293  0.01399824  0.09724450
#>  [7] -1.54714013  0.38699738 -1.14213997  0.37518377 -1.41528303 -1.76844395
#> [13] -0.41537449 -2.44331385  0.36074945  0.27916566 -1.33775501  0.54140463
#> [19] -0.17609097  0.65375214  0.90762690 -0.81805799  1.09460583 -2.17117176
#> [25] -2.36372989  0.75846007 -0.98986512 -0.25919574 -0.31922038 -0.97210338
#> [31]  0.34742005  0.87193899  1.69150733 -1.79873998 -1.99735527  0.98529485
#> [37] -0.09317414 -0.81280114 -0.34996201  1.94894593  1.91780038  0.21812873
#> [43] -1.02409867  0.02879087  2.11457411  0.36010960 -0.55980527 -0.34293770
#> [49]  0.15292439  0.36188459  1.06989669  0.80108332  1.41484101  0.83607197
#> [55]  0.22075940 -0.40349512 -0.58719864  0.45898479 -0.27641088  0.25763402
#> [61]  1.56312000 -1.53243668 -1.17723674  1.14482499 -1.76253166  0.44479058
#> [67] -1.98852894 -0.23348243  1.26733866  0.42622886 -1.07294703  1.06015122
#> [73]  0.82625794 -3.14924712  1.28380220  2.39009354 -0.23298520 -0.98009594
#> [79] -0.20263722 -1.87777744
#> 
#> $Input_Data$Forecasts_Train
#> Time Series:
#> Start = 1 
#> End = 80 
#> Frequency = 1 
#>       Series 1     Series 2    Series 3     Series 4    Series 5    Series 6
#>  1  2.05363668  0.563875297 -0.01677989  1.484899363 -0.49814865  1.19306441
#>  2  1.06576101  1.140092731  0.67444971 -0.344953847  1.61564589 -0.99912150
#>  3 -0.45602511 -1.146952687  2.85515277  2.475129848  1.97719884  1.56023992
#>  4  1.23157011  1.115781111  1.38667429  1.735021832  2.28344570  0.76854589
#>  5  0.16681894  1.596474068  1.29254493  2.448146115  1.54009640  2.34803462
#>  6  1.11802033  0.434533366  0.33677616 -0.417104795  2.82492346  1.93848369
#>  7 -0.34382028 -1.287608301  2.26362221  1.220429325  0.69255550  2.58219241
#>  8  2.39272378 -2.074020579  1.25227785 -1.602217631  2.02429569  1.58164969
#>  9  1.01025296  2.046954939  1.13133544  0.198145791  1.34963418  2.31220033
#> 10 -0.05864447  1.917637294 -0.19926086  2.404266603 -0.30954127  1.05723937
#> 11 -0.24978368  0.309917556  3.64015409  1.182986952  1.53858698  0.31804768
#> 12  1.53383244  0.903082949  2.70336193 -0.630195867  0.22303166  0.68634681
#> 13  1.21175907  1.640775413 -0.07515557 -0.116653321  0.41984994 -0.17005165
#> 14  0.54129458  0.308765571  1.01912822  2.404470673  0.05156644  1.59257589
#> 15  0.22906433  3.508212626  0.61660192  1.719606693  0.60799598 -0.78472259
#> 16  3.02310555 -0.611364803  2.67930185  1.159006726  0.68777214  2.68111893
#> 17  1.42254762 -0.227908903 -1.23147176  2.246120101 -0.39090973 -0.32709398
#> 18  1.05709859  0.887741601  0.70676719 -0.866771744  0.92945309  1.87097069
#> 19 -0.17913077  2.787537189 -0.39053154  0.827678898  1.76658137  1.00965627
#> 20  0.32373744 -0.584146732  0.17011948 -0.519839111  0.12452454  1.17899945
#> 21  0.80056622  1.919907074  1.02255643  1.039615275  0.31325134  0.90022738
#> 22  1.33299995  1.408400192  1.19982663  2.532611851  3.13610368 -0.88242462
#> 23  1.49738799 -0.369860575 -0.25403394 -0.595913489 -1.27870692 -0.14468988
#> 24  0.98516291  1.383843859  0.36567965  2.133419068 -0.25734673  2.14966700
#> 25  1.18420674  1.757106278 -0.52335626  2.041603966 -0.87159013  1.54682213
#> 26  1.16779087  1.583658765  0.62516470  0.718283577  1.35647496  0.61579591
#> 27  0.74637994  0.789015855  1.38499522  0.355013927 -0.36079428  2.58577352
#> 28  0.30024717  0.623522870  2.41283870  0.548817720  0.44387766  0.70513140
#> 29  1.61112891  1.750218486  0.02366234  0.442830653 -0.19590576  0.95111601
#> 30  0.54948718  0.365340977  0.34688058  1.562321864  0.29788105  0.61914536
#> 31 -0.73713523 -0.005119908  2.23196001  1.885851783  1.05267641  2.91156975
#> 32  1.24976929  1.809079974  0.91690182  2.703493331  0.87090917  1.19412151
#> 33  0.89213642  0.387046588  2.38520596  1.641067127  0.86059820  0.45103761
#> 34  1.07471007  0.204204043  0.05943721  1.318738321 -0.40611608  0.67505399
#> 35  1.71498084  1.118335933  1.66008734  1.107475426  0.31068146 -0.01163537
#> 36  0.82311866  0.574145724 -0.03638264  2.561212647  0.37547448  1.09439586
#> 37 -0.22018574  0.390243078  0.73443139  0.696569438  0.24192644  0.68420085
#> 38  0.17291001 -0.815583150 -0.06360228  2.285348938  1.11539239  1.42083714
#> 39  0.49344701  0.948220254 -0.15668501  2.252691649  0.59004991 -1.92938279
#> 40  1.28795863  1.039952672  2.00251011  0.441274184  2.13481718 -0.63733411
#> 41  0.67997689 -0.353194006  1.11569167 -0.395986006  0.31739292  0.07047733
#> 42  0.40422254 -0.206073736  1.81235016 -0.369148642  1.84627310  2.28157084
#> 43  2.03921561  0.127839047  0.51898696  1.826489742  1.50676617  1.06675795
#> 44 -1.16964618  0.398390336  2.78567556  0.500491465  0.23373307  2.33236734
#> 45  0.84292487  0.537054482  0.01826186 -0.934126856 -1.15359085  1.16087294
#> 46  1.36652278  1.631037676  0.63975744 -1.026691816  0.04055037  1.52783247
#> 47  0.19457983  0.919686671  3.28954059  1.013525786  3.15004150 -0.13675612
#> 48  0.29739525 -0.137124704  0.48001654  0.437395203  2.82631975  0.61964416
#> 49 -0.01658661  2.332133149  1.04613008  1.690293507  0.50454724  1.40731751
#> 50  1.61763468  2.780330329  0.77193344  2.258156220 -0.36189503  1.22849941
#> 51  1.54849041  1.427951742  0.45723650  0.858721324  0.68050166  1.64718946
#> 52  1.74316814  1.039844480 -0.26806404  0.244161805  1.65867416  3.01610561
#> 53  1.68198511  0.952100756  0.53348268  4.175119739  0.42357090 -0.44490483
#> 54  2.01337613 -1.850930139  0.31655114  2.291482080  0.86985572  1.44255832
#> 55  0.23260785  0.854728845  0.21660329  1.410207729  1.04778780  0.53129449
#> 56 -0.18482896  1.004261519  1.00184716  2.109049516  0.47732127 -0.08851122
#> 57  4.05160232  0.342703012 -0.44821558  1.770511853  1.56319544  1.14081828
#> 58  0.79163435 -0.458042259  0.93363352  0.116477681  3.55324627  1.62963575
#> 59  1.44204047  2.072177466  0.91976017  1.390762979  0.43765780  1.74701448
#> 60  0.81411377  0.679852052  1.09408528  1.927360669 -0.44979968  2.50651231
#> 61  1.54078321  1.806492838  1.40894179  1.459958395  1.84382144 -0.02865043
#> 62 -0.64778409  1.070035071  1.35926353  2.678797249  1.99620034 -0.06224062
#> 63  1.62239808  2.817210753  0.79981999  1.893531080 -1.41891684  0.95532333
#> 64  0.67692915  0.757317506  0.61804653  0.263565970  0.87230834  2.42096476
#> 65  1.06676479  2.306317991  0.96295978 -0.007500466  1.55385695  0.86241180
#> 66  2.05443736  0.526617091 -0.47553995  1.039116311  1.92376627  2.25714535
#> 67  2.19823067 -0.396046432  0.90173197  0.445422836  1.90256432  2.97931673
#> 68  0.37976529  0.538231271  1.49681181  1.734440378 -0.53301137 -0.24428007
#> 69  0.56382954  2.196862740  0.32228882 -0.653797291  1.18331975  0.86716581
#> 70  1.26655248  1.457423412 -0.66165101  0.551294484  1.43189217  0.74044833
#> 71  1.41136731  0.181598445  2.03196705  2.590159196  1.32801684 -0.49634852
#> 72  1.14378530  0.611814992  0.44526872  0.664563256  0.82863377  0.51143185
#> 73  1.15796738  0.240319045 -0.21850688  2.653507515  1.28239290  1.87580496
#> 74  0.14289738 -0.106574008 -0.33598680  1.625073872  0.38620657  1.44883153
#> 75  1.73912097 -0.110859001  1.21065123  0.116790059  2.00194926  0.84061517
#> 76  2.24064475  0.859936577  3.38910858  1.586045556 -0.51556061  1.56258377
#> 77  0.83338093  3.794987724  0.16728613  0.648424672  0.28521960  0.88538142
#> 78  0.38848624  2.092461207  2.35898875 -0.587234315  1.60299406  0.54224410
#> 79  1.62790990  1.165801730 -1.08539072  2.217963636  2.66373849  2.86809336
#> 80 -0.44938721  0.379743953  1.04837126 -0.650483777  1.10956898  1.05234666
#>       Series 7      Series 8    Series 9   Series 10
#>  1  0.76523853  1.0955508669  0.01375888  0.73258457
#>  2  0.46537826 -0.0565589203  0.23352721  1.14795765
#>  3  1.99255847  1.3504578923  0.88715388 -1.44273338
#>  4 -1.04386312  1.1264098311  1.36197100  1.25744790
#>  5 -0.49499138  1.0720645168  1.05942656  1.18752351
#>  6  0.33634065  0.4812098751 -0.51299268  0.64203226
#>  7  0.27122737  1.7321511016  1.00720349  1.46168504
#>  8  2.78992734  2.1222413024  0.82070193  1.81685707
#>  9  2.96954294  0.5158309297  0.95697232 -0.41482803
#> 10  2.76992457  2.5284651376  1.77432177  1.34023469
#> 11  2.27212159  1.2662461470  0.87272368  1.85070112
#> 12  0.08595674  2.6287420634  3.02978244  0.50741854
#> 13  0.40642983  0.7025864777  1.03271069  1.62941982
#> 14  1.30278344 -0.6837667921  2.79014169 -0.67073485
#> 15  0.20769924 -0.6201820911  2.03983036  1.78936630
#> 16  1.45674056  1.3021283287  0.01050658  0.70059114
#> 17  1.07654822  1.0701904091 -0.26551604  1.21605433
#> 18  0.07102455  1.3553683556  0.90394843  2.75968055
#> 19  1.60026534  1.9217062288  0.53339155 -0.58079984
#> 20 -0.07348830  2.8859682374  1.41114291  0.81826518
#> 21 -0.24412846  0.5460578472  2.83088848  0.36004575
#> 22  2.03005064  1.3392841337 -0.35389926  2.87918373
#> 23  1.61290811  2.2529034774  1.54448627 -0.71926403
#> 24 -0.29105855  1.4942409738  1.57561861  1.24535793
#> 25 -1.62228804  1.2929841442  0.03951869  1.25947269
#> 26  0.36485381  0.9249986423  0.15460532  0.65756844
#> 27 -1.77867042  0.6611554807  0.67714028 -0.18795190
#> 28 -0.62653788  3.2969508729  1.05026424  1.17705919
#> 29 -0.39035290  0.8056331478  1.25909742 -0.65089263
#> 30  0.17973079 -0.1320751405  1.99957428  1.42945044
#> 31  1.79120978  1.2581366365  0.35526236 -0.16932914
#> 32  0.30238298 -0.9822207408 -0.10350256  1.93791212
#> 33  2.26047706  1.0488237620  1.15750793  1.93999251
#> 34  0.88419036 -0.0948148149  1.16001072 -0.19539351
#> 35 -0.47155677  1.0871187272 -0.05299467  2.88780663
#> 36  0.42384772  1.2498287469  1.39385809  1.13043906
#> 37  1.01388202  2.1217577534  1.80270992  2.10931141
#> 38  0.56366074  0.0375281825  1.89720393 -0.46984377
#> 39  2.86719444  1.7647756722 -1.57590224  2.49058410
#> 40  1.44801921  0.5603151762  0.84930746  1.49571439
#> 41  1.17704599 -1.3947564521  0.65222503  0.81716866
#> 42  2.63363850 -0.0008119264  1.61373500  0.31011432
#> 43 -0.46514829  1.0348811150 -0.56518251 -1.55204294
#> 44  0.35699711 -0.0539978275  1.62274115 -0.02340835
#> 45  0.95921939 -0.2923149984  0.60219729  3.52774867
#> 46 -0.05442454  0.9302271590 -0.59359921  0.75270678
#> 47  1.90132586 -0.0981437888 -0.98896944  0.52490863
#> 48  0.69380177 -0.6941349760  2.68795341  0.59162361
#> 49  1.66762317  1.4493367092  2.21292741  1.58737065
#> 50  1.59860114 -0.3307812494  1.67954658  0.38856091
#> 51  1.50348262 -0.4796962451  1.88794329  0.51906132
#> 52  1.47769328  0.3816240303  1.47071497  2.58065538
#> 53  1.92604605  0.1957599564  0.17224099  0.40608607
#> 54 -0.62143203 -0.1221402068  1.48549640  0.50912440
#> 55  0.47102374 -0.0476459791  1.96082118  1.26765541
#> 56  2.06354919  0.9320134476 -0.12082121  1.68285203
#> 57  1.78482494 -0.4541415853 -0.60233076  1.88525515
#> 58  2.13286141  2.1177241922  1.26516711  1.63896978
#> 59 -0.88253956  0.8184636366  1.81654289 -0.36202606
#> 60  1.99620929  2.4361013825  1.10638239  0.55016758
#> 61 -0.11075308  1.2015580544  0.81006739 -1.07405180
#> 62  1.28169186  1.1981280368  0.67118709  0.05669903
#> 63  0.33079062 -0.2978164166  1.22416543  2.54541640
#> 64  2.69375910  1.3015555164  1.73608640  1.19722086
#> 65  0.61927712  1.4323122044  2.96630590  0.81776827
#> 66  1.23141067  0.4340760266  1.05081267 -0.28935369
#> 67  0.99189026  0.4787211477  2.07072135  0.55491208
#> 68  2.31904355  2.5495630714  1.83897761 -0.81249415
#> 69 -0.78220439 -0.0846816855 -0.20466071  0.84006228
#> 70 -0.40301870  0.4818908011  0.86714059  1.50681465
#> 71  0.73671091  2.8541978138  1.60461203  0.30451573
#> 72  1.65493073  1.1125117287  1.15090957  1.96502831
#> 73  2.34702038  1.3981669122  0.85890775  0.12041588
#> 74  1.04998995  0.3129293090  1.23665578 -0.28392214
#> 75  0.27573904  1.5576712431  0.84715510 -1.00360435
#> 76  0.25472465  1.5985148002  0.27195106  0.25070226
#> 77  1.51436099  2.0381950836  0.07179370  0.98758339
#> 78  2.04226994 -0.7244602505 -0.33206496  0.46671753
#> 79  0.91320437  1.0524820954 -0.30732651  2.75284288
#> 80  2.17192080  1.2518783615  0.32037509  1.65846364
#> 
#> $Input_Data$Actual_Test
#>  [1] -1.10173467  1.48795896 -1.19351184  1.09334917 -0.65562997 -0.18145816
#>  [7] -0.16423389  0.73719266 -0.12668156 -0.89741745  0.02251389  0.52431411
#> [13] -0.64876949 -0.48841087  0.29798790 -0.30615530  0.12501825  0.61106021
#> [19] -0.61174683 -0.59765555
#> 
#> $Input_Data$Forecasts_Test
#>          Series 1   Series 2   Series 3    Series 4    Series 5     Series 6
#>  [1,]  0.64597186  3.2829615  0.1946415 -0.29998713  1.30539471  0.178951266
#>  [2,]  0.41838727  1.1777115  1.0929950  2.26165240  1.16522108  2.283398367
#>  [3,] -0.52829970  1.7260965  0.3581551  0.21309376  1.14500866  0.922843978
#>  [4,]  1.27559101  0.4704141 -0.9158245  2.47831797 -0.53372868  1.915701397
#>  [5,]  0.46932568  2.8667155  0.8489132  0.38056759  0.78212724  0.162578168
#>  [6,]  1.93844837  2.4232756 -0.3744798  2.27207298  1.30314477  1.982043775
#>  [7,]  0.08174119  1.2720598  1.8044605  0.58089539  2.31638679  0.783373581
#>  [8,]  0.33806618  0.9480585 -0.2767791  0.31659251 -0.29379290 -0.468357956
#>  [9,]  0.77125028 -0.2568628  1.2603546  1.42299454  1.26835145 -0.356109252
#> [10,]  0.06549155  2.2873105  1.7917262  1.03067797 -0.04762221  0.984313497
#> [11,]  2.27201994  0.5789162  1.0055890  0.91492502  1.85407185  0.944932744
#> [12,]  1.35137398  0.6907232  0.4513851  1.49515099  0.57949447 -0.026277452
#> [13,]  1.44828372  0.8131609 -0.1259644 -1.63546411  1.92754887  0.845303812
#> [14,]  1.44125092  1.4340330  2.0403718  1.29956716  1.49888078  0.975210387
#> [15,]  1.84577016 -0.3050849  1.9719475  1.02134777  0.29884628 -0.521841708
#> [16,]  2.69749900  2.0373567  2.0006697  1.08497997  2.22785609  0.308150626
#> [17,]  1.19049645  1.2508690  1.9710773 -0.06489629  0.81438479  1.201089159
#> [18,]  1.06013001  0.3967578  1.9362865  0.27760933 -0.04851204 -0.196806991
#> [19,]  2.83177939  2.4096569  2.6963384  1.05979917  0.77528367  0.007149377
#> [20,]  1.65481371  1.4639786  1.5055529  2.04962420  1.40204921  0.524783884
#>         Series 7    Series 8    Series 9  Series 10
#>  [1,]  0.6015497  0.89934638  1.18399385  2.0537908
#>  [2,]  0.2391462  1.23180437  0.36807307  1.9337883
#>  [3,]  1.0570417 -0.41322095  0.45332528  0.6284268
#>  [4,]  0.3919709 -0.71763292  1.82700057  4.3558531
#>  [5,] -0.1539184  0.43704155 -1.11886764 -0.5581414
#>  [6,] -0.2293735  0.73690952  3.14393734  0.6176063
#>  [7,]  1.2869125  0.19997664  2.53261930 -0.8234228
#>  [8,]  0.1125278  1.33708655  0.17728105  0.8165584
#>  [9,]  0.3036276  0.30182053  1.34308047 -0.1019285
#> [10,]  0.8052410  0.70725485  0.48077986  2.1865026
#> [11,]  3.0997132  1.65402393  0.51737186  0.8614262
#> [12,]  1.6886318  1.99055106  0.81981489 -0.4340572
#> [13,]  1.6687467  0.72700400  1.08565875  0.8868952
#> [14,]  2.3461621 -0.03615145  0.62345373  1.6835272
#> [15,]  2.3685941  0.65536802  2.38662867  0.9437545
#> [16,]  0.4615011 -0.57528939  0.99827391  0.1379858
#> [17,]  0.1351285  1.45523586  0.48054499  1.3597907
#> [18,]  0.2140050 -0.35264742 -0.07263044  3.6826985
#> [19,]  1.0681155  2.67559179  2.16976869  2.0628562
#> [20,]  1.2330834  0.03974830 -0.68956655  1.5315062
#> 
#> 
#> $Weights
#>  [1]  0.36235017  0.20440322  0.14524527 -0.19482349  0.08272518 -0.05950967
#>  [7]  0.15732311  0.05594566  0.09252195  0.15381860
#> 
#> $Forecasts_Test
#>  [1] 1.6595797 0.5090137 0.4934832 0.6449220 0.6710732 1.0815536 0.9048686
#>  [8] 0.4524835 0.4321425 1.0352466 1.7671167 0.8407385 1.6405561 1.6082113
#> [15] 1.5249472 1.7930584 1.3381800 1.2737737 2.6032879 1.1709689
#> 
#> $Accuracy_Test
#>                 ME     RMSE      MAE       MPE    MAPE
#> Test set -1.275961 1.648109 1.447169 -213.3704 716.624
#> 
#> attr(,"class")
#> [1] "foreccomb_res"