Simple Average Forecast Combination

Computes forecast combination weights using simple average and produces forecasts for the test set, if provided.

comb_SA(x)

Arguments

x

An object of class foreccomb. Contains training set (actual values + matrix of model forecasts) and optionally a test set.

Value

Returns an object of class foreccomb_res with the following components:

Method

Returns the used forecast combination method.

Models

Returns the individual input models that were used for the forecast combinations.

Weights

Returns the combination weights obtained by applying the combination method to the training set.

Fitted

Returns the fitted values of the combination method for the training set.

Accuracy_Train

Returns range of summary measures of the forecast accuracy for the training set.

Forecasts_Test

Returns forecasts produced by the combination method for the test set. Only returned if input included a forecast matrix for the test set.

Accuracy_Test

Returns range of summary measures of the forecast accuracy for the test set. Only returned if input included a forecast matrix and a vector of actual values for the test set.

Input_Data

Returns the data forwarded to the method.

Details

Suppose \(y_t\) is the variable of interest, there are \(N\) not perfectly collinear predictors, \(\mathbf{f}_t = (f_{1t}, \ldots, f_{Nt})'\). The simple average gives equal weights to all predictors:

$$\mathbf{w}^{SA} = \frac{1}{N}$$

The combined forecast is then obtained by:

$$\hat{y}_t = {\mathbf{f}_{t}}'\mathbf{w}^{SA}$$

It is well-documented that simple average is a robust combination method that is hard to beat (e.g., Stock and Watson, 2004; Timmermann, 2006). This is often associated with the importance of parameter estimation error in sophisticated techniques -- a problem that simple averaging avoids. However, simple averaging may not be a suitable combination method when some of the predictors are biased (Palm and Zellner, 1992).

Author

Christoph E. Weiss and Gernot R. Roetzer

References

Palm, F. C., and Zellner, A. (1992). To Combine or not to Combine? Issues of Combining Forecasts. Journal of Forecasting, 11(8), 687--701.

Stock, J. H., and Watson, M. W. (2004). Combination Forecasts of Output Growth in a Seven-Country Data Set. Journal of Forecasting, 23(6), 405--430.

Timmermann, A. (2006). Forecast Combinations. In: Elliott, G., Granger, C. W. J., and Timmermann, A. (Eds.), Handbook of Economic Forecasting, 1, 135--196.

See also

Examples

obs <- rnorm(100)
preds <- matrix(rnorm(1000, 1), 100, 10)
train_o<-obs[1:80]
train_p<-preds[1:80,]
test_o<-obs[81:100]
test_p<-preds[81:100,]

data<-foreccomb(train_o, train_p, test_o, test_p)
comb_SA(data)
#> $Method
#> [1] "Simple Average"
#> 
#> $Models
#>  [1] "Series 1"  "Series 2"  "Series 3"  "Series 4"  "Series 5"  "Series 6" 
#>  [7] "Series 7"  "Series 8"  "Series 9"  "Series 10"
#> 
#> $Fitted
#> Time Series:
#> Start = 1 
#> End = 80 
#> Frequency = 1 
#>  [1]  0.3561936  1.3043169  0.8925007  1.0419459  0.9115526  1.6456233
#>  [7]  1.6811670  0.8772132  0.7948183  1.3780666  0.7702856  0.8172704
#> [13]  1.0149930  0.9426681  0.6775626  1.0217793  0.4489135  0.4671096
#> [19]  1.0215429  0.5697669  0.9662795  1.2429169  0.3235498  0.4427082
#> [25]  1.1889369  0.8364731  1.0400515  0.9290086  1.2873271  0.7687938
#> [31]  1.1041987  0.4237440  0.6015905  0.7404234  0.9329522  0.8817196
#> [37]  1.0436691  0.9818229  1.2888628  0.8698393  1.0513606  0.8567970
#> [43]  1.1947858  0.9984797  0.6201573  1.2282400  0.9214939 -0.1415124
#> [49]  0.6193054  0.8629051  0.7927605  0.8508809  1.3862864  0.5661011
#> [55]  0.7623480  1.0940169  1.1192927  1.0008678  1.1754182  1.2649456
#> [61]  0.8024332  0.7729645  1.5432437  1.0835350  0.8098354  0.9554576
#> [67]  0.7270153  0.8981271  0.5127164  1.3348535  0.3208708  1.1035797
#> [73]  1.0104665  1.3323799  1.2479196  0.9186076  0.9505744  0.5898825
#> [79]  1.0953523  0.8184407
#> 
#> $Accuracy_Train
#>                 ME     RMSE      MAE      MPE     MAPE      ACF1 Theil's U
#> Test set -1.069462 1.495661 1.218332 56.46138 433.6117 0.0126144  1.119023
#> 
#> $Input_Data
#> $Input_Data$Actual_Train
#> Time Series:
#> Start = 1 
#> End = 80 
#> Frequency = 1 
#>  [1] -0.04249169 -0.41714645 -1.18093192  0.35171480 -0.70454711 -0.29304003
#>  [7] -0.19493523 -0.77587873  0.92983916  0.54265741 -0.54698873  1.00137009
#> [13]  0.75348609  0.13198948 -0.48879928 -0.12277124 -0.38992146  0.78807867
#> [19]  0.40471278 -0.25288389  0.05844535 -1.12092300 -1.93545632 -0.49757896
#> [25]  1.45290265  1.19703846 -1.78325873  1.05682072  0.01290553  1.40384062
#> [31]  2.02546450  0.32457881 -2.26300382 -0.02787734 -0.10489221 -1.41407234
#> [37]  0.39221780 -1.62180678  1.00902037  1.41812938 -0.62021215  0.87872815
#> [43] -0.86205399  0.23121974 -0.71737279  1.60125174  0.70543200 -0.43804411
#> [49] -1.07044859 -0.94234932 -0.01909123 -1.40469739  0.22761002  0.28192347
#> [55]  0.67857451  0.13541009 -1.69687522 -1.29568451  0.58079013 -0.44847303
#> [61] -1.09146289  1.32874106 -1.06767731 -1.15074070 -2.40641091  0.92488878
#> [67] -0.34281184  0.98655676  0.39824549 -1.68090534 -1.86896931 -1.09971642
#> [73]  1.30491940  0.32195814 -0.57901744  0.38123941  2.07473421 -2.10209882
#> [79] -0.68249582 -0.50420638
#> 
#> $Input_Data$Forecasts_Train
#> Time Series:
#> Start = 1 
#> End = 80 
#> Frequency = 1 
#>      Series 1    Series 2    Series 3      Series 4     Series 5    Series 6
#>  1 -0.2184684 -0.62964853  1.13215360  1.0790088338  0.899685922  1.71836286
#>  2  1.7010267  1.20638480  1.60989338  0.9946757791  0.625153736  1.37903005
#>  3  1.3150830  0.55040735  0.48138749  0.3124851312  0.804580376  1.45510940
#>  4 -0.5748842  2.95555629  0.70295462  3.3598949578  0.151185351  0.37834649
#>  5  0.3565508  1.20432816  2.21757074  0.3496305353 -1.121488084  2.99056635
#>  6  1.0580102  1.30094300  3.32438828  2.8110988252  1.092854136  1.95493886
#>  7  1.5254557  0.86848423  0.84239150  2.6770304579  1.534503474  1.66858923
#>  8  0.8487771  2.37717092  0.91141820 -1.4204957080  1.187010124  0.60856446
#>  9  0.9198401  0.94755085 -0.07904363  0.8436624755 -0.089211388  1.44204951
#> 10 -0.9119280  1.97686749  2.05853051  0.3998159527  1.650940423  0.46035122
#> 11 -0.7032093  0.98164335  0.63314145  0.7014331320  2.242670908 -0.66193535
#> 12 -0.1802785  1.43820854  0.82129345  0.0830592711  2.535457031  0.05794416
#> 13  0.9336198  1.16774983  0.59156926  1.4502536219  0.417751725  0.08290414
#> 14  2.2005072  2.48768012 -1.15830104  1.4172655771  0.932210666  0.17812272
#> 15 -1.2038898  1.08267832  1.25113147  0.8428576175  0.856482709  0.95091931
#> 16  1.5087117  1.72339157  1.97552483 -0.1545765940  2.093914454  0.84015481
#> 17  0.4882643  0.40885437 -0.90836737  0.6112847761  1.037397957  0.19845162
#> 18 -0.5121609  0.92377837  1.61721222 -0.0807395121 -1.085459213  0.50036956
#> 19  1.3625370 -1.13837173  0.60871101 -0.1378006016  1.999851145  1.14224224
#> 20  1.6502252  0.78428072 -0.53229384  1.3900360538  0.260906262  0.69693664
#> 21  1.3667083  2.02313765  2.64460111 -0.5777092970  1.123248924 -0.73475086
#> 22  2.7339297  3.36757132  1.76362745  0.3553047939  0.062825224 -0.00578237
#> 23  0.5863131  1.31192376  0.87621734 -0.7839217411 -0.787829872  0.59994829
#> 24 -1.4476993 -0.03595046  1.26729554  2.4851258352  2.105465517 -0.81064695
#> 25  1.4110322  1.38836541  0.60179675  1.8419605480  0.630104200  0.29708705
#> 26 -0.5201504  0.24568371  1.28465427  0.8283457900  0.301679490  1.11057952
#> 27  2.0196285  1.37702194 -0.64365513  1.8402977863  0.358826674  2.57744833
#> 28  0.8531958  1.53338251  1.70023800  0.9769709717  0.836182201  0.84939213
#> 29  1.5141490  1.95474381  1.56790655  1.2557613686 -0.082107159  0.33608064
#> 30  0.1287433  0.14583590  0.62707544  1.9795207587  1.246326009 -1.20673665
#> 31  1.8769401  0.79302819  2.67381286 -0.5079141759  1.446887120  0.16859589
#> 32 -0.6909849  1.64618305  1.13212311 -0.9329047392  0.373553856  2.40239318
#> 33  1.5387477  0.17047752 -1.20303924 -0.0003748651  1.170582456  0.30883076
#> 34  1.1285627  1.41365061  1.42686347 -0.0879494520  0.712264357  0.23017109
#> 35  1.3616280 -0.76122743  3.33292998  1.4174246858 -0.004494804 -0.54058341
#> 36 -0.3405516  2.39697462 -0.46885109  1.5680878360  0.118920179  0.63488393
#> 37  0.6367456 -0.95460224  2.30453650  2.3694050737  0.484919903  2.34578701
#> 38  1.0784545  0.75706334  2.65534196  1.1405989684 -0.287435533  0.53256412
#> 39  0.4265607  2.07895696  3.50766350  2.1437742126  1.970459224  2.33344247
#> 40 -0.5855427  1.09570688  1.33450797  1.9218223190 -0.284394699  0.88724993
#> 41  1.1693011 -0.28366250  0.86421984  2.2552257413  0.210696354  2.02712541
#> 42  0.5081722  0.78839213  0.77444267  0.5690952525  0.666286085  1.17419831
#> 43  1.1736217 -0.41680101  2.08056553  1.8014136131  0.943339601  1.27584811
#> 44  2.3297271  1.42695246  1.72568873  1.1806934722  1.551887706  0.14674101
#> 45  1.7746438  1.40623055  1.17064462  0.6807685291  0.110961733 -0.27611706
#> 46  0.3878130  1.06761579  1.63112211  2.5208626626  0.384409907  1.60318763
#> 47  2.7794978  2.26756969  1.09985942 -0.0931655912 -0.567639438  0.47364304
#> 48 -0.2127188  0.05351471  0.04894204 -1.1459340261  0.393301572 -0.07008226
#> 49  1.8312451  0.56151907 -1.60970147  1.5993647068  1.248331473  2.20190716
#> 50  1.5317903  0.22493892  1.63625439  0.6252170293 -1.650678229  1.53887238
#> 51  1.8334911  1.75187558 -0.09549374 -0.3119534972  2.740471059  1.69594886
#> 52 -0.8096490  0.76238039  0.40007701  0.6024953172  1.374488406 -0.02820405
#> 53  1.2370148  1.56469792  2.49769091  1.3719732775 -0.072376601  2.40026991
#> 54  2.5380052  0.55257958  1.21760805  0.3558010026 -0.677408196  0.34512867
#> 55  0.6289708  0.85475175  0.39240165  1.7200882485 -0.526369014  0.73937950
#> 56  0.8262772  1.19582167  0.69585487  1.4665811255  1.946906493  0.61469011
#> 57  2.1709191  0.91592370  1.44916468  0.0567547687  0.653895277  0.59160286
#> 58  1.8119589  0.81324590  1.85075585  1.8910827439  1.350454778  0.91959392
#> 59  1.4154421 -1.22965188 -0.85193122  2.4162691444  1.665417407  0.76564644
#> 60  1.5040192  1.96180222 -0.22594865  1.0553384233  0.191498044  1.34660168
#> 61 -1.1908905  2.34413851  0.66116618  1.4510021922  2.947744793  1.34616600
#> 62  1.8200660  0.82062455  1.41339856 -0.4747533906  1.912107369  1.10433669
#> 63  3.8313260  0.61186997  1.57137587  1.8580981175  1.537142186  1.69539786
#> 64  2.4635868  0.38975772  2.00894690  0.1055858171  1.605997022  1.55143104
#> 65 -1.0512271  0.40781374 -0.85058495  0.6234779785 -0.775388349  1.91062655
#> 66 -0.3380079 -0.34611257  0.68417161  1.8760916767  1.936111432  0.85050269
#> 67  1.2856218 -0.13463803  0.81354119  1.2990356919  1.324773674  0.01071242
#> 68  2.3212264  0.09760612  0.82745655 -0.3088401930  1.043398512  1.00143733
#> 69  2.2068936  0.97336542 -0.40588593 -1.0510812515  0.132732251  0.48290061
#> 70  1.4970485  2.31986946  1.29381390  0.8218048373  1.588610957  1.50427842
#> 71  0.7391603  0.35599962  1.54001156  0.4172409859  1.438564823 -0.92503289
#> 72  0.2024876 -1.09708057  0.14497590  1.0909599178  2.971648811  1.61157060
#> 73  2.1064396  1.19050779  0.46741005  0.9922900783  1.199923500  0.93160621
#> 74  0.3453661  2.68241126  0.03770578  1.5536709621  1.051388848  0.31932759
#> 75  1.6895657  1.40271157  3.59602268  0.4868465503  1.889118729  0.96734937
#> 76  1.1807965  0.10314555  0.98750214  0.8541644233  0.139908213  0.00957158
#> 77 -1.0109033  1.04464554 -0.04745814  1.6884141634  1.203544474  2.81252816
#> 78  1.0153277  2.37814231  1.93991680 -0.9773213883 -0.294972628  1.28433282
#> 79  1.8963694  2.04584996  1.17304465  2.3463818083 -0.418796598  1.86689293
#> 80  1.9988239 -0.69095099 -0.05710431  1.5823484726  1.865153378  1.97627449
#>       Series 7    Series 8    Series 9   Series 10
#>  1  1.02701921 -0.38002648  0.36897187 -1.43512255
#>  2  2.10623953  1.56967102 -0.10216837  1.95326195
#>  3  0.28926850  1.90243841  0.57549287  1.23875486
#>  4  2.18550020  1.68379775  0.35785855 -0.78075101
#>  5  1.51622048  1.29773510  0.05928899  0.24512289
#>  6  1.15078766 -0.11626044  0.87623057  3.00324189
#>  7  1.96038963  2.69202711  0.66691360  2.37588529
#>  8  0.40161475 -0.10835385  2.11305694  1.85336917
#>  9  0.59675128 -1.28937167  2.58652631  2.06942871
#> 10  1.97416789  1.29026823  3.89091751  0.99073513
#> 11  1.82577027  1.50854784 -0.06673348  1.24152752
#> 12  2.41627764 -0.49968240  0.55239046  0.94803478
#> 13 -0.16721979  1.17803477  2.50357407  1.99169299
#> 14  1.61092254  3.37618277 -0.38981448 -1.22809484
#> 15  0.82790727  2.29750427 -0.52810636  0.39814170
#> 16  1.14263931 -0.95088574  0.81018691  1.22873212
#> 17 -0.38656062  1.86622038  0.93239206  0.24119735
#> 18  2.52382011 -0.55755863  0.73706193  0.60477201
#> 19  1.79132040  2.94935915  0.07819678  1.55938396
#> 20  0.07373524  1.48641660 -0.47246253  0.35988880
#> 21  1.34467735  0.72151966  0.39506497  1.35629702
#> 22  2.15017350  2.28247762  0.71772255 -0.99868109
#> 23 -0.39618535  0.70502324  0.33550476  0.78850409
#> 24 -0.65132774  0.64192526 -0.13986277  1.01275664
#> 25  1.18922813  2.32584920  1.59990722  0.60403806
#> 26  1.27861504  2.60130668  0.91680851  0.31720823
#> 27  0.43155994 -0.04605832  1.90416989  0.58127558
#> 28  1.78256576  0.62251808  0.20667280 -0.07103210
#> 29  1.77927067  1.57831038  2.09563583  0.87351961
#> 30  1.85274761  1.69212670  0.17464917  1.04764954
#> 31  1.97570416  1.44479567  0.67986536  0.49027154
#> 32  0.06261883  0.80493558  0.50516520 -1.06564354
#> 33  0.96798517  0.09212712  1.26708850  1.70347950
#> 34 -0.15923098  0.58228220  1.14324104  1.01437903
#> 35  0.05348603  1.71542476  1.52553810  1.22939633
#> 36  1.44722747  2.77497980  0.06012151  0.62540298
#> 37  1.84792814  0.46419549  1.25691982 -0.31914405
#> 38  0.63352372  1.61115176  1.30997047  0.38699534
#> 39 -0.29777921 -0.91418582  2.26212972 -0.62239347
#> 40 -0.59504495  2.02591651  1.35628269  1.54188867
#> 41  0.29564576  0.97561382  0.28724571  2.71219509
#> 42  1.79990940 -0.38127349  1.56467272  1.10407436
#> 43  1.12829092  1.17053785  1.05131677  1.73972528
#> 44  2.54403684 -0.64975931 -0.22508257 -0.04608879
#> 45  1.15099333 -0.40398016 -0.96395784  1.55138535
#> 46 -1.03921222  1.59315987  3.04826199  1.08517929
#> 47  2.14509772  0.17208518 -0.36592344  1.30391456
#> 48  0.06424101 -0.31136568 -0.09261927 -0.14240328
#> 49  0.65715462  0.43174180 -0.46546380 -0.26304440
#> 50 -0.10570697  0.91785117  1.27173915  2.63877333
#> 51 -1.37644244 -0.65452392  0.24409353  2.10013843
#> 52  0.91260760  0.30640791  2.27019812  2.71800703
#> 53  2.85308243  1.85276045 -0.52189081  0.67964142
#> 54  0.64007683  2.25953005 -0.69839449 -0.87191544
#> 55  0.91442764  0.49764649  1.20998602  1.19219661
#> 56  1.24717205 -0.03185828  0.78493490  2.19378907
#> 57  0.71801709  0.28915394  2.24986966  2.09762540
#> 58  1.08684843 -1.24412268  1.25483267  0.27402778
#> 59  1.59401629  2.71925231  1.05928882  2.20043232
#> 60  1.73973198  1.88369081  1.76780524  1.42491744
#> 61  0.20099145 -0.56985779  0.50510534  0.32876540
#> 62  0.33231341  1.03008721 -0.25373443  0.02519894
#> 63  2.38710626  1.46956993  0.64301410 -0.17246376
#> 64  1.43173119  1.42289429 -0.07656268 -0.06801816
#> 65  3.34469329  1.58614132  1.14577461  1.75702710
#> 66 -0.10163374  1.22685482  2.95349106  0.81310739
#> 67  0.56797057 -0.12741821  1.40940980  0.82114415
#> 68  2.18965182  1.89497522 -0.36594781  0.28030740
#> 69  0.24656581  1.04254185  1.03484010  0.46429182
#> 70  1.70336924  1.34653031  0.98382258  0.28938695
#> 71 -0.75625179 -0.15003431  0.02411302  0.52493671
#> 72  2.13122568  1.37762462  2.78783998 -0.18545568
#> 73  1.98698819  0.63653755 -1.47345935  2.06642142
#> 74  3.61143249  1.30588490  0.47392676  1.94268427
#> 75  1.10494684 -0.69090393  1.07268439  0.96085392
#> 76  2.47961295  0.54639257  0.76896152  2.11602073
#> 77  1.13772016  0.89776144  1.84438090 -0.06488935
#> 78  0.68959865  0.74655328 -1.62523298  0.74248018
#> 79  2.38420052  0.32347892 -1.23215175  0.56825350
#> 80 -0.09819999  1.33891381 -1.46143314  1.73058126
#> 
#> $Input_Data$Actual_Test
#>  [1]  2.93814366 -0.03024096 -1.63192786  1.31258502  0.51347653 -1.89922275
#>  [7]  2.05858228 -0.57913265 -0.12269519 -2.06216184 -0.64820293  1.21898225
#> [13]  1.12262789 -1.32476527 -1.52246068 -0.17074061 -0.35277688  0.25211419
#> [19] -0.21254940 -1.09434913
#> 
#> $Input_Data$Forecasts_Test
#>         Series 1   Series 2    Series 3   Series 4    Series 5   Series 6
#>  [1,]  1.2362291 -0.3324174  1.42063834  2.3014363  0.71135401  1.0378908
#>  [2,]  0.2013982  1.2127054  0.16891184  3.0212128  1.66896737  0.9025812
#>  [3,] -0.2860251  2.1345375  1.77341868  1.1958026  0.11830202 -0.4730781
#>  [4,]  1.0305477  0.4689361  1.24696762  0.6529113  0.73732473  1.8255984
#>  [5,]  1.1737134  2.2656662  1.59263848 -0.3274277  0.39032234  1.3830866
#>  [6,]  1.8397746  1.8017185  2.58046710  1.3256050 -0.59433024  1.7763536
#>  [7,]  1.9510879  0.8598049  0.48606097  0.8762474 -0.04002443  0.6063751
#>  [8,]  0.7465104  0.4392344  1.50863965  0.7777432  0.45745522  0.8179961
#>  [9,]  0.2566115  0.9693214  1.38714745  1.0046752  0.11524296  0.6699261
#> [10,]  1.5550701  0.3634936  3.82014988 -0.5310510  0.65701042  0.8398373
#> [11,]  1.3892796  0.1513573  1.09680754  0.5192801  2.11829694  2.1298140
#> [12,]  2.4242842 -0.2314883  0.07241386  1.6758124  0.95351878  1.0675563
#> [13,]  0.6664764  0.8388485  0.89895334 -0.3589311  1.10885488  1.3269441
#> [14,] -0.7555144  0.5477960  1.52454916  0.2796415 -1.72408664  1.8646939
#> [15,] -0.6555123  0.6658855 -0.01895611  0.2853116  1.56045382  1.3225790
#> [16,]  0.7264913  3.0370770  2.93984016  1.8457638  0.27560251  0.8883780
#> [17,]  2.4750815  1.0110217  1.03635631  1.6306348  2.60712381  0.4997926
#> [18,]  0.2928301  1.9142255  0.35634313  1.3297243  2.40721334  0.8828390
#> [19,]  0.7202949  0.4931269  0.62560850  0.9675699  0.14231022 -0.2465887
#> [20,]  1.4564298  1.9613638  0.85775081  2.4123107 -0.34363447 -0.2244872
#>           Series 7    Series 8   Series 9  Series 10
#>  [1,] -0.209763988  0.53787120  0.5475319 -0.2626617
#>  [2,]  1.045778684  0.04089664  1.9413199  1.4434884
#>  [3,]  0.253629987  1.34732256  1.4499772  0.5040655
#>  [4,]  0.258836487  1.24407075  2.0055950  0.8403591
#>  [5,] -0.508090573  1.65213196  0.3929334  0.2120014
#>  [6,]  0.564955744  1.55874117  1.4731355  1.0099777
#>  [7,]  0.414161243  1.13534775  2.4710015  0.2848401
#>  [8,] -1.257835295 -0.94398224  1.3936147  0.4510322
#>  [9,]  3.828086572  1.07660559  1.0945393  1.8943865
#> [10,] -0.369405003  2.74052774  1.5050813 -0.4522847
#> [11,]  2.070914172 -0.07868193  0.7737953  2.0210669
#> [12,]  0.448184637  0.74391218  2.1967723  0.4053058
#> [13,]  1.030518997  0.74105358  1.9882791  1.7623252
#> [14,]  0.462277390 -0.35770271 -1.1762790  1.4135532
#> [15,]  3.335567647  0.40209352  1.1855468  0.6993398
#> [16,] -0.008964317  0.39490103  1.1805931  1.1236516
#> [17,]  1.209651754  0.47504563  0.3550197  1.4012487
#> [18,]  1.735390199  0.99137761 -0.5492022  1.0234619
#> [19,]  1.048567521  1.17492102  0.5623795  1.6132611
#> [20,] -0.653437351  1.95762620  2.2646691  0.4382070
#> 
#> 
#> $Weights
#>  [1] 0.1 0.1 0.1 0.1 0.1 0.1 0.1 0.1 0.1 0.1
#> 
#> $Forecasts_Test
#>  [1] 0.6988109 1.1647260 0.8017953 1.0311147 0.8226976 1.3336399 0.9044902
#>  [8] 0.4390408 1.2296543 1.0128430 1.2191930 0.9756272 1.0003323 0.2078928
#> [15] 0.8782309 1.2403334 1.2700977 1.0384203 0.7101451 1.0126798
#> 
#> $Accuracy_Test
#>                 ME     RMSE      MAE      MPE     MAPE
#> Test set -1.061324 1.715411 1.465379 399.2511 436.4617
#> 
#> attr(,"class")
#> [1] "foreccomb_res"