Result Rbject for the Forecast Combination Methods

Stores the results and inputs for some combined forecasts.

foreccomb_res(method, modelnames, fitted, accuracy_insample, input_data,
  predict = NULL, intercept = NULL, weights = NULL, pred = NULL,
  accuracy_outsample = NULL, trim_factor = NULL, top_predictors = NULL,
  ranking = NULL)

Arguments

method

Name of the method.

modelnames

Names of the models provided by the user.

fitted

The fitted values.

accuracy_insample

Insample accuracy obtained by the method.

input_data

Contains the input data provided by the user.

predict

(optional) Function used to conduct predictions for new forecasts.

intercept

(optional) Intercept of the method, if it requires one.

weights

(optional) Weights of the method, if it requires one.

pred

= (optional) Predictions on a test set.

accuracy_outsample

= (optional) Accuracy on the test set.

trim_factor

= (optional) Trim factor used in some of the methods.

top_predictors

(optional) Number of retained predictors.

ranking

(optional) Ranking of the predictors that determines which models are removed in the trimming step.

Value

Returns an object of class foreccomb_res.

Author

Christoph E. Weiss, Gernot R. Roetzer