Summary of Forecast Combination

summary method for class ‘foreccomb_res’. Includes information about combination method, combination weights assigned to the individual forecast models, as well as an accuracy evaluation of the combined forecast.

# S3 method for foreccomb_res
summary(object, ...)

# S3 method for foreccomb_res_summary
print(x, ...)

Arguments

object

An object of class 'foreccomb'. Contains training set (actual values + matrix of model forecasts) and optionally a test set.

...

potential further arguments (require by generic)

x

An object of class 'foreccomb'. Contains training set (actual values + matrix of model forecasts) and optionally a test set.

See also

Author

Christoph E. Weiss and Gernot R. Roetzer

Examples

obs <- rnorm(100)
preds <- matrix(rnorm(1000, 1), 100, 10)
train_o<-obs[1:80]
train_p<-preds[1:80,]
test_o<-obs[81:100]
test_p<-preds[81:100,]

data<-foreccomb(train_o, train_p, test_o, test_p)
fit<-comb_BG(data)
summary(fit)
#> 
#> Summary of Forecast Combination 
#> ------------------------------- 
#> 
#> Method:  Bates/Granger (1969) 
#> 
#> Individual Forecasts & Combination Weights: 
#> 
#>           Combination Weight
#> Series 1          0.08531103
#> Series 2          0.10293142
#> Series 3          0.11196901
#> Series 4          0.08819836
#> Series 5          0.11618539
#> Series 6          0.11318259
#> Series 7          0.08537323
#> Series 8          0.11033871
#> Series 9          0.09720048
#> Series 10         0.08930977
#> 
#> Accuracy of Combined Forecast: 
#> 
#>                      ME     RMSE      MAE      MPE     MAPE
#> Training Set -1.2468921 1.634787 1.395013 161.3296 331.6406
#> Test set     -0.8447852 1.337376 1.128924 284.9210 344.8997
#> 
#> Additional information can be extracted from the combination object: 
#> For fitted values (training set):  fit$Fitted 
#> For forecasts (test set):  fit$Forecasts_Test 
#> See  str(fit)  for full list.